### 2021-06-25 #3402 @whedon
[PRE REVIEW]: SIMSOPT: A flexible framework for stellarator optimization
**Submitting author:** @landreman (<a href="http://orcid.org/0000-0002-7233-577X">Matt Landreman</a>)
**Repository:** <a href="https://github.com/hiddenSymmetries/simsopt" target ="_blank">https://github.com/hiddenSymmetries/simsopt</a>
**Version:** v0.3.06
