### 2021-06-25 #3402 @kyleniemeyer
Hi @landreman, thanks for your submission to JOSS! It looks like there is a compilation error in your paper—I don't think we support a footnote in the author list like that (since that is YAML, not Markdown/LaTeX).

Unfortunately, we have extremely limited editor availability to handle submissions right now, and so we need to put your submission on our waitlist untill an appropriate editor becomes available. Thanks for your patience!
