### 2021-06-25 b8cd6c3
Trying to fix whedon compile error by removing email address for corresponding author
