### 2021-06-25 #3402 @landreman
Sorry about that compile error - it compiled fine using the Docker approach before we submitted. I fixed the error so it now compiles via whedon.
