+++ src/simsopt/geo/surfaceobjectives.py
@@ -50,6 +50,139 @@ class Volume(object):
     def dJ_by_dcoilcurrents(self):
         return [0. for c in self.stellarator.coils]
 
+
+class ToroidalFlux(object):
+
+    r"""
+    This objective calculates
+        J = \int_{varphi = constant} B \cdot n ds
+          = \int_{varphi = constant} curlA \cdot n ds
+          from Stokes' theorem
+          = \int_{curve on surface where varphi = constant} A \cdot n dl
+    given a surface and Biot Savart kernel.
+    """
+
+    def __init__(self, surface, biotsavart, stellarator, idx=0, boozer_surface=None):
+        self.stellarator = stellarator
+        self.surface = surface
+        self.boozer_surface = boozer_surface
+        self.biotsavart = biotsavart
+        self.idx = idx
+        self.clear_cached_properties()
+
+    def clear_cached_properties(self):
+        self._J = None
+        self._dJ_by_dcoefficients = None
+        self._dJ_by_dcoilcurrents = None
+        x = self.surface.gamma()[self.idx]
+        self.biotsavart.set_points(x)
+
+    def J(self, compute_derivatives=0):
+        assert compute_derivatives>=0
+        if compute_derivatives>0:
+            assert self.boozer_surface is not None # you need to initialize with a BoozerSurface to have access to Jacobian
+        
+        if self._J is None:
+            self.compute(compute_derivatives=compute_derivatives)
+        return self._J
+
+    def dJ_by_dcoefficients(self, compute_derivatives=1):
+        assert compute_derivatives>=1 and self.boozer_surface is not None
+        if self._dJ_by_dcoefficients is None:
+            self.compute(compute_derivatives=compute_derivatives)
+        return self._dJ_by_dcoefficients
+
+    def dJ_by_dcoilcurrents(self, compute_derivatives=1):
+        assert compute_derivatives>=1 and self.boozer_surface is not None
+        if self._dJ_by_dcoilcurrents is None:
+            self.compute(compute_derivatives=compute_derivatives)
+        return self._dJ_by_dcoilcurrents
+
+
+
+    def compute(self, compute_derivatives=0):
+        xtheta = self.surface.gammadash2()[self.idx]
+        ntheta = self.surface.gamma().shape[1]
+        A = self.biotsavart.A()
+        self._J = np.sum(A * xtheta)/ntheta
+        
+        if compute_derivatives > 0:
+            booz_surf = self.boozer_surface
+            iota = booz_surf.res['iota']
+            G = booz_surf.res['G']
+            P, L, U = booz_surf.res['PLU']
+            dconstraint_dcoils_vjp = booz_surf.res['dconstraint_dcoils_vjp']
+            
+            dJ_by_dA = self.dJ_by_dA()
+            dJ_by_dcoils = self.biotsavart.A_vjp(dJ_by_dA)
+            
+            # tack on dJ_diota = dJ_dG = 0 to the end of dJ_ds
+            dJ_ds = np.zeros(L.shape[0])
[…]
