+++ tests/geo/test_boozersurface.py
@@ -137,65 +137,65 @@ class BoozerSurfaceTests(unittest.TestCase):
 
 
 
-#    def test_dlsqgrad_dcoils_vjp(self):
-#        """
-#        This test verifies that the dlsqgrad_dcoils_vjp calculation is correct.
-#        """
-#
-#        def get_lsqgrad(surface, label, target_label, biotsavart,iota=0.,G=None):
-#            if not isinstance(s, SurfaceXYZTensorFourier):
-#                raise RuntimeError('Exact solution of Boozer Surfaces only supported for SurfaceXYZTensorFourier')
-#        
-#        coils, currents, ma = get_ncsx_data()
-#        stellarator = CoilCollection(coils, currents, 3, True)
-#        bs = BiotSavart(stellarator.coils, stellarator.currents)
-#
-#        mpol = 8  # try increasing this to 8 or 10 for smoother surfaces
-#        ntor = 8  # try increasing this to 8 or 10 for smoother surfaces
-#        stellsym = True
-#        nfp = 3
-#        
-#        phis = np.linspace(0, 1/(2*nfp), ntor+1, endpoint=False)
-#        thetas = np.linspace(0, 1, 2*mpol+1, endpoint=False)
-#        s = SurfaceXYZTensorFourier(mpol=mpol, ntor=ntor, stellsym=stellsym, nfp=nfp, quadpoints_phi=phis, quadpoints_theta=thetas)
-#        s.fit_to_curve(ma, 0.10, flip_theta=True)
-#        iota = -0.4
-#        G = 2. * np.pi * np.sum(np.abs(bs.coil_currents)) * (4 * np.pi * 10**(-7) / (2 * np.pi))
-#        
-#        label = Area(s, stellarator)
-#        target_label = 0.025
-#
-#        booz_surf = BoozerSurface(bs, s, label, target_label)
-#
-#        coeffs = stellarator.get_dofs()
-#        def f(dofs):
-#            stellarator.set_dofs(dofs)
-#            bs.clear_cached_properties()
-#            boozer = boozer_surface_residual(s, iota, G, bs, derivatives=1)
-#            r = boozer[0]
-#            J = boozer[1]
-#            return J.T @ r
-#        f0 = f(coeffs)
-#        
-#        np.random.seed(1)
-#        lm1 = np.random.uniform(size=f0.size)-0.5
-#        lm1_dg_dcoils = boozer_surface_dlsqgrad_dcoils_vjp(lm1, booz_surf, iota, G, bs)
-#        lm1_dg_dcoils = stellarator.reduce_coefficient_derivatives(lm1_dg_dcoils)
-#        
-#        lm2 = np.random.uniform(size=lm1_dg_dcoils.size)-0.5
-#        fd_exact = np.dot(lm1_dg_dcoils , lm2)
-#        
-#        err_old = 1e9
-#        epsilons = np.power(2., -np.asarray(range(7, 20)))
-#        print("################################################################################")
-#        for eps in epsilons:
-#            f1 = f(coeffs + eps * lm2)
-#            Jfd = (f1-f0)/eps
-#            err = np.linalg.norm(np.dot(Jfd, lm1)-fd_exact)/np.linalg.norm(fd_exact)
-#            print(err/err_old)
-#            assert err < err_old * 0.55
-#            err_old = err
-#        print("################################################################################")
+    def test_dlsqgrad_dcoils_vjp(self):
+        """
+        This test verifies that the dlsqgrad_dcoils_vjp calculation is correct.
+        """
+
+        def get_lsqgrad(surface, label, target_label, biotsavart,iota=0.,G=None):
+            if not isinstance(s, SurfaceXYZTensorFourier):
[…]
