+++ src/simsopt/geo/surfaceobjectives.py
@@ -2,9 +2,13 @@ import numpy as np
 import simsoptpp as sopp
 import scipy
 
+
 def forward_backward(P, L, U, rhs):
-    y = scipy.linalg.solve_triangular( U.T, rhs, lower=True) 
-    z = scipy.linalg.solve_triangular( L.T, y, lower=False) 
+    """
+    Solve a linear system of the form PLU*adj = rhs
+    """
+    y = scipy.linalg.solve_triangular(U.T, rhs, lower=True) 
+    z = scipy.linalg.solve_triangular(L.T, y, lower=False) 
     adj = P@z 
     return adj
 
@@ -17,6 +21,7 @@ class Area(object):
     def __init__(self, surface, stellarator):
         self.stellarator = stellarator
         self.surface = surface
+    
     def J(self):
         """
         Compute the area of a surface.
@@ -106,27 +111,25 @@ class ToroidalFlux(object):
         self.biotsavart.set_points(x)
 
     def J(self, compute_derivatives=0):
-        assert compute_derivatives>=0
-        if compute_derivatives>0:
-            assert self.boozer_surface is not None # you need to initialize with a BoozerSurface to have access to Jacobian
+        assert compute_derivatives >= 0
+        if compute_derivatives > 0:
+            assert self.boozer_surface is not None  # you need to initialize with a BoozerSurface to have access to Jacobian
         
         if self._J is None:
             self.compute(compute_derivatives=compute_derivatives)
         return self._J
 
     def dJ_by_dcoefficients(self, compute_derivatives=1):
-        assert compute_derivatives>=1 and self.boozer_surface is not None
+        assert compute_derivatives >= 1 and self.boozer_surface is not None
         if self._dJ_by_dcoefficients is None:
             self.compute(compute_derivatives=compute_derivatives)
         return self._dJ_by_dcoefficients
 
     def dJ_by_dcoilcurrents(self, compute_derivatives=1):
-        assert compute_derivatives>=1 and self.boozer_surface is not None
+        assert compute_derivatives >= 1 and self.boozer_surface is not None
         if self._dJ_by_dcoilcurrents is None:
             self.compute(compute_derivatives=compute_derivatives)
         return self._dJ_by_dcoilcurrents
-
-
 
     def compute(self, compute_derivatives=0):
         xtheta = self.surface.gammadash2()[self.idx]
@@ -151,12 +154,12 @@ class ToroidalFlux(object):
             adj = forward_backward(P, L, U, dJ_ds)
         
             adj_times_dg_dcoil, adj_times_dg_dcurr = dconstraint_dcoils_vjp(adj, booz_surf, iota, G, booz_surf.bs)
-            self._dJ_by_dcoefficients = [dj_dc - adj_dg_dc for dj_dc, adj_dg_dc in zip(dJ_by_dcoils, adj_times_dg_dcoil) ]
[…]
