+++ src/simsopt/geo/surfacexyztensorfourier.py
@@ -124,5 +124,5 @@ class SurfaceXYZTensorFourier(sopp.SurfaceXYZTensorFourier, Surface):
             mask[ntor+1:, 0] = False
         if npsame(phis, np.linspace(0, 1/(2*self.nfp), ntor+1, endpoint=False)) and \
                 npsame(thetas, np.linspace(0, 1, 2*mpol+1, endpoint=False)):
-                    mask[0, mpol+1:] = False
+            mask[0, mpol+1:] = False
         return mask
