+++ tests/geo/test_surface_objectives.py
@@ -14,7 +14,6 @@ from simsopt.geo.boozersurface import BoozerSurface
 from .surface_test_helpers import get_ncsx_data, get_surface, get_exact_surface
 
 
-
 surfacetypes_list = ["SurfaceXYZFourier", "SurfaceRZFourier", "SurfaceXYZTensorFourier"]
 stellsym_list = [True, False]
 
@@ -39,7 +38,8 @@ def taylor_test1(f, df, x, epsilons=None, direction=None):
         err_old = err
     print("################################################################################")
 
-def taylor_test2(f, df, d2f, x, epsilons=None, direction1=None, direction2 = None):
+
+def taylor_test2(f, df, d2f, x, epsilons=None, direction1=None, direction2=None):
     np.random.seed(1)
     if direction1 is None:
         direction1 = np.random.rand(*(x.shape))-0.5
@@ -63,6 +63,7 @@ def taylor_test2(f, df, d2f, x, epsilons=None, direction1=None, direction2 = Non
         err_old = err
     print("################################################################################")
 
+
 class ToroidalFluxTests(unittest.TestCase):
     def test_toroidal_flux_is_constant(self):
         # this test ensures that the toroidal flux does not change, regardless
@@ -75,13 +76,13 @@ class ToroidalFluxTests(unittest.TestCase):
         gamma = s.gamma()
         num_phi = gamma.shape[0]
         
-        tf_list = np.zeros( (num_phi,) )
+        tf_list = np.zeros((num_phi,))
         for idx in range(num_phi):
-            tf = ToroidalFlux(s, bs_tf, stellarator, idx = idx)
+            tf = ToroidalFlux(s, bs_tf, stellarator, idx=idx)
             tf_list[idx] = tf.J()
         mean_tf = np.mean(tf_list)
 
-        max_err = np.max( np.abs(mean_tf - tf_list) ) / mean_tf
+        max_err = np.max(np.abs(mean_tf - tf_list)) / mean_tf
         assert max_err < 1e-2
     
 #    def test_toroidal_flux_dcoilcoeff(self):
@@ -102,17 +103,17 @@ class ToroidalFluxTests(unittest.TestCase):
 #            return dJ_by_dcoils 
 #        taylor_test1(f, df, coeffs )
 
-
     def test_toroidal_flux_first_derivative(self):
         for surfacetype in surfacetypes_list:
             for stellsym in stellsym_list:
-                with self.subTest(surfacetype = surfacetype, stellsym=stellsym):
-                    self.subtest_toroidal_flux1(surfacetype,stellsym)
+                with self.subTest(surfacetype=surfacetype, stellsym=stellsym):
+                    self.subtest_toroidal_flux1(surfacetype, stellsym)
+
     def test_toroidal_flux_second_derivative(self):
         for surfacetype in surfacetypes_list:
             for stellsym in stellsym_list:
-                with self.subTest(surfacetype = surfacetype, stellsym=stellsym):
-                    self.subtest_toroidal_flux2(surfacetype,stellsym)
+                with self.subTest(surfacetype=surfacetype, stellsym=stellsym):
+                    self.subtest_toroidal_flux2(surfacetype, stellsym)
 
     def subtest_toroidal_flux1(self, surfacetype, stellsym):
         coils, currents, ma = get_ncsx_data()
@@ -185,36 +186,35 @@ class NonQuasiAxiSymmetricComponentTests(unittest.TestCase):
  
         G0 = 2. * np.pi * np.sum(np.abs(bs.coil_currents)) * (4 * np.pi * 10**(-7) / (2 * np.pi))
[…]
