+++ examples/optimize_qs_exact.py
@@ -1,263 +1,0 @@
-#!/usr/bin/env python3
-import sys
-import os
-import numpy as np
-from simsopt.geo.biotsavart import BiotSavart
-from simsopt.geo.surfacexyzfourier import SurfaceXYZFourier
-from simsopt.geo.surfacexyztensorfourier import SurfaceXYZTensorFourier
-from simsopt.geo.surfacerzfourier import SurfaceRZFourier
-from simsopt.geo.curve import RotatedCurve
-from simsopt.geo.curverzfourier import CurveRZFourier 
-from simsopt.geo.curvexyzfourier import CurveXYZFourier
-from simsopt.geo.curveobjectives import CurveLength, MinimumDistance, UniformArclength
-from simsopt.geo.boozersurface import BoozerSurface
-from simsopt.geo.surfaceobjectives import NonQuasiAxisymmetricComponentPenalty, Area, Volume, ToroidalFlux, boozer_surface_residual, MajorRadius
-from simsopt.geo.coilcollection import CoilCollection
-sys.path.append(os.path.join("..", "tests", "geo"))
-from surface_test_helpers import get_ncsx_data, get_surface, get_exact_surface
-from scipy.optimize import minimize, least_squares
-from lbfgs import fmin_lbfgs
-
-
-
-
-coils, currents, ma = get_ncsx_data(Nt_coils=6, Nt_ma=10, ppp=20)
-stellarator = CoilCollection(coils, currents, 3, True)
-coils = stellarator.coils
-currents = stellarator.currents
-
-
-mpol = 8  # try increasing this to 8 or 10 for smoother surfaces
-ntor = 8  # try increasing this to 8 or 10 for smoother surfaces
-stellsym = True
-nfp = 3
-exact = True
-
-surf_list = []
-#for idx,target in enumerate(np.linspace(-0.162,-4.0, 20)):
-#for idx,target in enumerate(np.linspace(-0.162,-2., 20)): NON UNIQUENESS HERE
-for idx,target in enumerate(np.linspace(-0.162,-1., 20)):
-    if exact:
-        phis = np.linspace(0, 1/(2*nfp), ntor+1, endpoint=False)
-        thetas = np.linspace(0, 1, 2*mpol+1, endpoint=False)
-    else:
-        phis = np.linspace(0, 1/(2*nfp), ntor+5, endpoint=False)
-        thetas = np.linspace(0, 1, 2*mpol+5, endpoint=False)
-    
-    s = SurfaceXYZTensorFourier(mpol=mpol, ntor=ntor, stellsym=stellsym, nfp=nfp, quadpoints_phi=phis, quadpoints_theta=thetas)
-    bs = BiotSavart(stellarator.coils, stellarator.currents)
-    
-    if len(surf_list) == 0:
-        s.fit_to_curve(ma, 0.10, flip_theta=True)
-        iota0 = -0.4
-        G0 = 2. * np.pi * np.sum(np.abs(bs.coil_currents)) * (4 * np.pi * 10**(-7) / (2 * np.pi))
-    else:
-        s.set_dofs(surf_list[-1].surface.get_dofs())
-        iota0 = surf_list[-1].res['iota'] 
-        G0 = surf_list[-1].res['G']
-
-    label = Volume(s, stellarator)
-    boozer_surface = BoozerSurface(bs, s, label, target)
-    
-    try:
-        res = boozer_surface.solve_residual_equation_exactly_newton(tol=1e-10, maxiter=30,iota=iota0,G=G0)
-        if res['success']:
-            surf_list += [boozer_surface]
[…]
