+++ examples/optimize_qs.py
@@ -1,143 +1,0 @@
-#!/usr/bin/env python3
-import sys
-import os
-import numpy as np
-from simsopt.geo.biotsavart import BiotSavart
-from simsopt.geo.surfacexyzfourier import SurfaceXYZFourier
-from simsopt.geo.surfacexyztensorfourier import SurfaceXYZTensorFourier
-from simsopt.geo.surfacerzfourier import SurfaceRZFourier
-from simsopt.geo.curve import RotatedCurve
-from simsopt.geo.curverzfourier import CurveRZFourier 
-from simsopt.geo.curvexyzfourier import CurveXYZFourier
-from simsopt.geo.curveobjectives import CurveLength 
-from simsopt.geo.boozersurface import BoozerSurface
-from simsopt.geo.surfaceobjectives import NonQuasiAxisymmetricComponentPenalty, Area, Volume, ToroidalFlux, boozer_surface_residual
-from simsopt.geo.coilcollection import CoilCollection
-sys.path.append(os.path.join("..", "tests", "geo"))
-from surface_test_helpers import get_ncsx_data, get_surface, get_exact_surface
-from scipy.optimize import minimize, least_squares
-from lbfgs import fmin_lbfgs
-
-
-
-
-coils, currents, ma = get_ncsx_data()
-stellarator = CoilCollection(coils, currents, 3, True)
-coils = stellarator.coils
-currents = stellarator.currents
-
-
-mpol = 10  # try increasing this to 8 or 10 for smoother surfaces
-ntor = 10  # try increasing this to 8 or 10 for smoother surfaces
-stellsym = True
-nfp = 3
-exact = True
-
-if exact:
-    phis = np.linspace(0, 1/(2*nfp), ntor+1, endpoint=False)
-    thetas = np.linspace(0, 1, 2*mpol+1, endpoint=False)
-else:
-    phis = np.linspace(0, 1/(2*nfp), ntor+5, endpoint=False)
-    thetas = np.linspace(0, 1, 2*mpol+5, endpoint=False)
-
-s = SurfaceXYZTensorFourier(mpol=mpol, ntor=ntor, stellsym=stellsym, nfp=nfp, quadpoints_phi=phis, quadpoints_theta=thetas)
-s.fit_to_curve(ma, 0.10, flip_theta=True)
-iota0 = -0.4
-
-bs = BiotSavart(stellarator.coils, stellarator.currents)
-bs_tf = BiotSavart(stellarator.coils, stellarator.currents)
-G0 = 2. * np.pi * np.sum(np.abs(bs.coil_currents)) * (4 * np.pi * 10**(-7) / (2 * np.pi))
-
-
-label = Volume(s, stellarator)
-label_target = label.J()
-boozer_surface = BoozerSurface(bs, s, label, label_target)
-
-for idx,target in enumerate(np.linspace(-0.3,-1.5,30)):
-    boozer_surface = BoozerSurface(bs, s, label, target)
-    res = boozer_surface.solve_residual_equation_exactly_newton(tol=1e-10, maxiter=300,iota=iota0,G=G0)
-    print(f"iota={res['iota']:.3f}, label={label.J():.3f}, area={s.area():.3f}, |label error|={np.abs(label.J()-target):.3e}, ||residual||={np.linalg.norm(boozer_surface_residual(s, res['iota'], res['G'], bs, derivatives=0)):.3e}")
-    iota0 = res['iota']
-    G0 = res['G']
-
-iota_target = -0.4
-iota_weight = 1.
-boozer_surface.res = boozer_surface.solve_residual_equation_exactly_newton( tol=1e-10, maxiter=10, iota=iota0, G=G0)
[…]

[…]
