+++ examples/2_Intermediate/boozer.py
@@ -41,7 +41,7 @@ s = SurfaceXYZTensorFourier(
 s.fit_to_curve(ma, 0.10, flip_theta=True)
 iota = -0.4
 
-tf = ToroidalFlux(s, bs_tf)
+tf = ToroidalFlux(s, bs_tf, stellarator)
 ar = Area(s)
 ar_target = ar.J()
 
