### 2021-06-25 281b933
more edits for linting
