+++ tests/geo/test_surface_objectives.py
@@ -205,12 +205,6 @@ class ToroidalFluxTests(unittest.TestCase):
             return tf.dJ_by_dsurfacecoefficients() 
         taylor_test1(f, df, coeffs)
 
-
-
-
-
-
-
     def test_toroidal_flux_is_constant(self):
         # this test ensures that the toroidal flux does not change, regardless
         # of the cross section (varphi = constant) across which it is computed
@@ -568,5 +562,3 @@ class MajorRadiusTests(unittest.TestCase):
             mr.clear_cached_properties()
             return mr.dJ_dsurfacecoefficients() 
         taylor_test1(f, df, coeffs)
-
-
