### 2021-06-27 12381c1
SurfaceHenneberg: implemented gamma, gammadash1, and gammadash2
