+++ src/simsopt/geo/surfacehenneberg.py
@@ -5,17 +5,23 @@ from scipy.optimize import minimize_scalar
 from scipy.interpolate import interp1d
 #import matplotlib.pyplot as plt
 
+import simsoptpp as sopp
 from .surface import Surface
 from .surfacerzfourier import SurfaceRZFourier
 
 logger = logging.getLogger(__name__)
 
 
-class SurfaceHenneberg(Surface):
+class SurfaceHenneberg(sopp.Surface, Surface):
     r"""
     This class represents a toroidal surface using the
     parameterization in Henneberg, Helander, and Drevlak,
-    arXiv:2105.00768 (2021). Stellarator symmetry is assumed.
+    arXiv:2105.00768 (2021). The main benefit of this representation
+    is that there is no freedom in the poloidal angle,
+    i.e. :math:`\theta` is uniquely defined, in contrast to other
+    parameterizations like
+    :obj:`~.surfacerzfourier.SurfaceRZFourier`. Stellarator symmetry
+    is assumed.
 
     In this representation by Henneberg et al, the cylindrical
     coordinates :math:`(R,\phi,Z)` are written in terms of a unique
@@ -64,7 +70,13 @@ class SurfaceHenneberg(Surface):
     index corresponding to ``n=0``.
     """
 
-    def __init__(self, nfp: int, alpha_fac: int, mmax: int, nmax: int):
+    def __init__(self,
+                 nfp: int,
+                 alpha_fac: int,
+                 mmax: int,
+                 nmax: int,
+                 quadpoints_phi: int = 63,
+                 quadpoints_theta: int = 62):
         if alpha_fac > 1 or alpha_fac < -1:
             raise ValueError('alpha_fac must be 1, 0, or -1')
 
@@ -74,16 +86,17 @@ class SurfaceHenneberg(Surface):
         self.nmax = nmax
         self.stellsym = True
         self.allocate()
-        # The recalculate attribute could be used in the future for
-        # cacheing, but it is not used for anything yet.
-        self.recalculate = True
+
+        Surface.__init__(self)
+        sopp.Surface.__init__(self,
+                              np.linspace(0, 1.0, quadpoints_phi, endpoint=False),
+                              np.linspace(0, 1.0, quadpoints_theta, endpoint=False))
 
         # Initialize to an axisymmetric torus with major radius 1m and
         # minor radius 0.1m
         self.R0nH[0] = 1.0
         self.bn[0] = 0.1
         self.set_rhomn(1, 0, 0.1)
-        Surface.__init__(self)
 
     def __repr__(self):
         return "SurfaceHenneberg " + str(hex(id(self))) + " (nfp=" + \
@@ -152,7 +165,7 @@ class SurfaceHenneberg(Surface):
         """
         self._validate_mn(m, n)
         self.rhomn[m, n + self.nmax] = val
-        self.recalculate = True
+        self.invalidate_cache()
 
     def get_dofs(self):
         """
@@ -162,23 +175,27 @@ class SurfaceHenneberg(Surface):
                                self.rhomn[0, self.nmax + 1:],
                                np.reshape(self.rhomn[1:, :], (self.mmax * (2 * self.nmax + 1),), order='C')))
 
-    def set_dofs(self, v):
+    def num_dofs(self):
+        """
+        Return the number of degrees of freedom.
+        """
+        ndofs = self.nmax + 1  # R0nH
+        ndofs += self.nmax  # Z0nH
+        ndofs += self.nmax + 1  # b0n
[…]
