### 2021-06-29 a509d02
Added test for energy conservation; run_autopep added/deleted a few empty lines
