+++ tests/field/test_particle.py
@@ -140,3 +140,76 @@ class ParticleTracingTesting(unittest.TestCase):
         for i in range(nparticles):
             assert validate_phi_hits(gc_phi_hits[i], bsh, nphis)
 
+    def test_energy_conservation(self):
+        # Test conservation of Energy = m v^2/2=m vparallel^2/2+mu B
+        # where mu=m v_perp^2/(2B) is the adiabatic invariant. In the
+        # the presence of a strong magnetic field, mu should also be
+        # conserved. Energy is conserved regardless of the magnitude
+        # of the magnetic field.
+        bsh = self.bsh
+        ma = self.ma
+        nparticles = 4
+        m = 1.67e-27
+        tmax = 1e-5
+        Ekin = 9000
+        np.random.seed(1)
+
+        gc_tys, gc_phi_hits = trace_particles_starting_on_axis(
+            ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=1,
+            Ekin=Ekin, umin=-1, umax=+1,
+            phis=[], mode='gc_vac')
+        fo_tys, fo_phi_hits = trace_particles_starting_on_axis(
+            ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=1,
+            Ekin=Ekin, umin=-1, umax=+1,
+            phis=[], mode='full')
+        particles_to_vtk(gc_tys, '/tmp/particles_gc')
+        particles_to_vtk(fo_tys, '/tmp/particles_fo')
+
+        # pick 100 random points on each trace, and ensure that
+        # the energy is being conserved both in the guiding center
+        # and full orbit cases up to some precision
+
+        N = 100
+        max_energy_gc_error = np.array([])
+        max_energy_fo_error = np.array([])
+        max_mu_gc_error = np.array([])
+        max_mu_fo_error = np.array([])
+        for i in range(nparticles):
+            fo_ty = fo_tys[i]
+            gc_ty = gc_tys[i]
+            idxs = np.random.randint(0, gc_ty.shape[0], size=(N, ))
+            gc_xyzs = gc_ty[idxs, 1:4]
+            bsh.set_points(gc_xyzs)
+            Bs = bsh.B()
+            AbsBs = bsh.AbsB()
+            energy_fo = np.array([])
+            energy_gc = np.array([])
+            mu_fo = np.array([])
+            mu_gc = np.array([])
+            vParInitial = gc_ty[0, 4]
+            muInitial = Ekin/AbsBs[0, 0]-m*vParInitial**2/2/AbsBs[0, 0]
+            for j in range(N):
+                v_fo = fo_ty[idxs[j], 4:]
+                v_gc = gc_ty[idxs[j], 4]
+                Bunit = Bs[j, :]/AbsBs[j, 0]
+                vperp = np.linalg.norm(v_fo - np.sum(v_fo*Bunit)*Bunit)
+                energy_fo = np.append(energy_fo, np.sqrt(m/2*(v_fo[[0]]**2+v_fo[[1]]**2+v_fo[[2]]**2)))
+                energy_gc = np.append(energy_gc, np.sqrt(m/2*(v_gc**2+2*muInitial*AbsBs[j, 0]/m)))
+                mu_fo = np.append(mu_fo, m*vperp**2/2/AbsBs[j, 0])
+                mu_gc = np.append(mu_gc, Ekin/AbsBs[j, 0]-m*v_gc**2/2/AbsBs[j, 0])
+            energy_fo_error = np.log10(np.abs(energy_fo-energy_fo[0]+1e-30)/energy_fo[0])
+            energy_gc_error = np.log10(np.abs(energy_gc-energy_gc[0]+1e-30)/energy_gc[0])
[…]
