### 2021-06-29 4b3108d
Bugfix in mhd.__init__
Replaced logger.WARNING with logger.warning function.
