### 2021-06-29 843436f
get energy conservation tests working
