### 2021-06-29 cfd7275
Graph based Optimizable subclassed to allow dof handling outside of Optimizable base class
