### 2021-06-30 bc7ffeb
Basic conda workflow added to github actions
