### 2021-06-30 020f891
bug fix for license file line in metal.yaml
