### 2021-06-30 dbc3a16
numpy dependency with version pinned added to conda recipe
