### 2021-06-30 e4b1c64
Reducing the build matrix till things work to speedup testing
