+++ .github/workflows/conda.yml
@@ -13,8 +13,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        platform: [ubuntu-latest, macos-latest]
-        python-version: ["3.7", "3.8", "3.9"]
+        platform: [ubuntu-latest] # , macos-latest]
+        python-version: ["3.7"] # , "3.8", "3.9"]
 
     runs-on: ${{ matrix.platform }}
 
