### 2021-06-30 4b6081a
python 3.7 removed and dependencies added to meta.yaml
