+++ .github/workflows/conda.yml
@@ -14,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         platform: [ubuntu-latest] # , macos-latest]
-        python-version: ["3.7"] # , "3.8", "3.9"]
+        python-version: ["3.8", "3.9"]
 
     runs-on: ${{ matrix.platform }}
 
