### 2021-06-30 73e0a43
python 3.7 removed and dependencies added to meta.yaml
