### 2021-06-30 f054635
Bug fix in specifying versions of dependencies in meta.yaml
