+++ conda.recipe/meta.yaml
@@ -29,13 +29,13 @@ requirements:
   run:
     - python
     - numpy >=1.19
-    - jax >= 0.2.4
-    - jaxlib >= 0.1.56
-    - scipy >= 1.5.4
-    - Deprecated >= 1.2.10
-    - nptyping >= 1.3.0
-    - pandas >= 1.1.4
-    - monty >= 2021.6.10 
+    - jax >=0.2.4
+    - jaxlib >=0.1.56
+    - scipy >=1.5.4
+    - Deprecated >=1.2.10
+    - nptyping >=1.3.0
+    - pandas >=1.1.4
+    - monty >=2021.6.10 
     - ruamel.yaml
 
 
