### 2021-06-30 4149e44
test script updated for conda recipe
