### 2021-06-30 d607e62
SurfaceHenneberg: quadpoints can be initialized by int or array
