### 2021-07-02 1dd4242
set arch to ivybridge for conda build
