### 2021-07-02 #3402 @dfm
Sure - I'm happy to edit!

@landreman: let me know if you have any reviewer recommendations from [our main list of volunteers](https://bit.ly/joss-reviewers) or if there's anyone else who you think would be particularly well suited (include their github handle, but don't @ them because then they'll get inundated with notifications :D). I'll do a first pass of editing and start recruiting reviewers next week.
