### 2021-07-02 287ed9d
more tracing tests and docstrings
