+++ src/simsoptpp/surface.h
@@ -123,9 +123,12 @@ class Surface {
         virtual void dgammadash1_by_dcoeff_impl(Array& data) { throw logic_error("dgammadash1_by_dcoeff_impl was not implemented"); };
         virtual void dgammadash2_by_dcoeff_impl(Array& data) { throw logic_error("dgammadash2_by_dcoeff_impl was not implemented"); };
 
+
         void normal_impl(Array& data);
         void dnormal_by_dcoeff_impl(Array& data);
         void d2normal_by_dcoeffdcoeff_impl(Array& data);
+
+        void unitnormal_impl(Array& data);
 
         double area();
         void darea_by_dcoeff_impl(Array& data);
@@ -162,6 +165,9 @@ class Surface {
         }
         Array& d2normal_by_dcoeffdcoeff() {
             return check_the_cache("d2normal_by_dcoeffdcoeff", {numquadpoints_phi, numquadpoints_theta,3, num_dofs(), num_dofs() }, [this](Array& A) { return d2normal_by_dcoeffdcoeff_impl(A);});
+        }
+        Array& unitnormal() {
+            return check_the_cache("unitnormal", {numquadpoints_phi, numquadpoints_theta,3}, [this](Array& A) { return unitnormal_impl(A);});
         }
         Array& darea_by_dcoeff() {
             return check_the_cache("darea_by_dcoeff", {num_dofs()}, [this](Array& A) { return darea_by_dcoeff_impl(A);});
