### 2021-07-02 f33bae4
more tracing tests and docstrings
