### 2021-07-02 cf6fb75
more senible dtmax estimate
