### 2021-07-02 cc6a00b
more tests, introduce a constants file
