+++ src/simsopt/field/tracing.py
@@ -3,6 +3,7 @@ import numpy as np
 import simsoptpp as sopp
 import logging
 from simsopt.field.magneticfield import MagneticField
+from simsopt.util.constants import ALPHA_PARTICLE_MASS, ALPHA_PARTICLE_CHARGE, FUSION_ALPHA_PARTICLE_ENERGY
 from nptyping import NDArray, Float
 
 
@@ -18,12 +19,12 @@ def compute_gc_radius(m, vperp, q, absb):
     return m*vperp/(abs(q)*absb)
 
 
-def gc_to_fullorbit_initial_guesses(field, xyz_inits, vtangs, vtotal, m, q):
+def gc_to_fullorbit_initial_guesses(field, xyz_inits, vtangs, vtotal, m, q, eta=0):
     """
     Takes in guiding center positions ``xyz_inits`` as well as a tangential
     velocities ``vtangs`` and total velocities ``vtotal`` to compute orbit
     positions for a full orbit calculation that matches the given guiding
-    center.
+    center. The phase angle can be controll via the `eta` parameter
     """
 
     nparticles = xyz_inits.shape[0]
@@ -34,16 +35,22 @@ def gc_to_fullorbit_initial_guesses(field, xyz_inits, vtangs, vtotal, m, q):
     Bs = field.B()
     AbsBs = field.AbsB()
     eB = Bs/AbsBs
-    ez = np.asarray(nparticles*[[0., 0., -1.]])
-    ez -= eB * np.sum(eB*ez, axis=1)[:, None]
-    ez *= 1./np.linalg.norm(ez, axis=1)[:, None]
-    Bperp = np.cross(eB, ez, axis=1)
-    Bperp *= 1./np.linalg.norm(Bperp, axis=1)[:, None]
-    vperp2s = vtotal**2 - vtangs**2
     for i in range(nparticles):
-        rgs[i] = compute_gc_radius(m, sqrt(vperp2s[i]), q, AbsBs[i, 0])
-        xyz_inits_full[i, :] = xyz_inits[i, :] + rgs[i] * ez[i, :]
-        v_inits[i, :] = -sqrt(vperp2s[i]) * Bperp[i, :] + vtangs[i] * eB[i, :]
+        p1 = eB[i, :]
+        p2 = np.asarray([0, 0, 1])
+        p3 = -np.cross(p1, p2)
+        p3 *= 1./np.linalg.norm(p3)
+        q1 = p1
+        q2 = p2 - np.sum(q1*p2)*q1
+        q2 = q2/np.sum(q2**2)**0.5
+        q3 = p3 - np.sum(q1*p3)*q1 - np.sum(q2*p3)*q2
+        q3 = q3/np.sum(q3**2)**0.5
+        vperpmag = np.sqrt(vtotal**2 - vtangs[i]**2)
+        rgs[i] = compute_gc_radius(m, vperpmag, q, AbsBs[i, 0])
+
+        xyz_inits_full[i, :] = xyz_inits[i, :] + rgs[i] * np.sin(eta) * q2 + rgs[i] * np.cos(eta) * q3
+        vperp = -vperpmag * np.cos(eta) * q2 + vperpmag * np.sin(eta) * q3
+        v_inits[i, :] = vtangs[i] * q1 + vperp
     return xyz_inits_full, v_inits, rgs
 
 
@@ -66,8 +73,8 @@ def parallel_loop_bounds(comm, n):
 
 def trace_particles(field: MagneticField, xyz_inits: NDArray[Float],
                     tangential_velocities: NDArray[Float], tmax=1e-4,
-                    mass=1.6726219e-27, charge=1, Ekin=9000, tol=1e-9,
[…]
