### 2021-07-02 1ecc1bc
wasn't initializing a Vector3d to zero...
