+++ src/simsopt/mhd/vmec_diagnostics.py
@@ -1,0 +1,165 @@
+# coding: utf-8
+# Copyright (c) HiddenSymmetries Development Team.
+# Distributed under the terms of the LGPL License
+
+"""
+This module contains functions that can postprocess VMEC output.
+"""
+
+import logging
+import numpy as np
+from scipy.interpolate import interp1d
+from .._core.optimizable import Optimizable
+from .vmec import Vmec
+
+logger = logging.getLogger(__name__)
+
+
+class Quasisymmetry1(Optimizable):
+    """
+    Note that the supplied value of ``n`` will be multiplied by
+    ``nfp``, so typically ``n`` should be +1 or -1 for quasi-helical
+    symmetry.
+    """
+
+    def __init__(self,
+                 vmec: Vmec,
+                 s,
+                 m=1,
+                 n=0,
+                 weights=None,
+                 ntheta: int = 31,
+                 nphi: int = 32) -> None:
+
+        self.vmec = vmec
+        self.depends_on = ["vmec"]
+        self.ntheta = ntheta
+        self.nphi = nphi
+        self.m = m
+        self.n = n
+
+        # Make sure s is a list:
+        try:
+            self.s = list(s)
+        except:
+            self.s = [s]
+
+        if weights is None:
+            self.weights = np.ones(len(self.s))
+        else:
+            self.weights = weights
+        assert len(self.weights) == len(self.s)
+
+    def get_dofs(self):
+        return np.array([])
+
+    def set_dofs(self, x):
+        self.need_to_run_code = True
+
+    def total(self):
+        """
+        Evaluate the quasisymmetry metric in terms of a scalar total.
+        """
+        residuals = self.residuals()
+        return np.sum(residuals * residuals)
+
+    def residuals(self):
+        """
+        Evaluate the quasisymmetry metric in terms of a 1D numpy vector of
+        residuals. This is the function to use when forming a least-squares
+        objective function.
+        """
+        self.vmec.run()
+        if self.vmec.wout.lasym:
+            raise RuntimeError('Quasisymmetry class cannot yet handle non-stellarator-symmetric configs')
+
+        logger.debug('Evaluating quasisymmetry residuals')
+        ns = len(self.s)
+        ntheta = self.ntheta
+        nphi = self.nphi
+        nfp = self.vmec.wout.nfp
+        d_psi_d_s = -self.vmec.wout.phi[-1] / (2 * np.pi)
+
+        # First, interpolate in s to get the quantities we need on the surfaces we need.
+        method = 'linear'
+
+        interp = interp1d(self.vmec.s_half_grid, self.vmec.wout.iotas[1:], fill_value="extrapolate")
+        iota = interp(self.s)
+
+        interp = interp1d(self.vmec.s_half_grid, self.vmec.wout.bvco[1:], fill_value="extrapolate")
+        G = interp(self.s)
+
+        interp = interp1d(self.vmec.s_half_grid, self.vmec.wout.buco[1:], fill_value="extrapolate")
+        I = interp(self.s)
+
+        interp = interp1d(self.vmec.s_half_grid, self.vmec.wout.gmnc[:, 1:], fill_value="extrapolate")
+        gmnc = interp(self.s)
+
[…]
