+++ tests/mhd/test_vmec_diagnostics.py
@@ -1,0 +1,142 @@
+import unittest
+import numpy as np
+import os
+import logging
+
+from simsopt.mhd.vmec_diagnostics import Quasisymmetry1
+
+try:
+    import vmec
+except ImportError as e:
+    vmec = None 
+
+try:
+    from mpi4py import MPI
+except ImportError as e:
+    MPI = None
+
+if MPI is not None:
+    from simsopt.mhd.vmec import Vmec  # , vmec_found
+from . import TEST_DIR
+
+logger = logging.getLogger(__name__)
+#logging.basicConfig(level=logging.DEBUG)
+
+
+@unittest.skipIf(vmec is None, "vmec python package is not found")
+class Quasisymmetry1Tests(unittest.TestCase):
+    def test_axisymmetry(self):
+        """
+        For an axisymmetric field, the QA error should be 0, while the QH
+        and QP error should be significant.
+        """
+
+        vmec = Vmec(os.path.join(TEST_DIR, 'input.circular_tokamak'))
+
+        for s in [[0.5], [0.3, 0.6]]:
+            qa = Quasisymmetry1(vmec, s, m=1, n=0)
+            residuals = qa.residuals()
+            print('max QA error:', np.max(np.abs(residuals)))
+            np.testing.assert_allclose(residuals, np.zeros_like(residuals), atol=2e-6)
+
+            qh = Quasisymmetry1(vmec, s, m=1, n=1)
+            print('QH error:', qh.total())
+            self.assertTrue(qh.total() > 0.01)
+
+            qp = Quasisymmetry1(vmec, s, m=0, n=1)
+            print('QP error:', qp.total())
+            self.assertTrue(qp.total() > 0.01)
+
+    def test_independent_of_resolution(self):
+        """
+        The total quasisymmetry error should be nearly independent of the
+        resolution parameters used.
+        """
+        vmec = Vmec(os.path.join(TEST_DIR, 'input.li383_low_res'))
+
+        ss = [[0.5], [0.7], [0.3, 0.6], [0.2, 0.4, 0.7]]
+        weightss = [None, [1], None, [2.2, 5.5, 0.9]]
+        ms = [1, 1, 1, 0]
+        ns = [0, 1, -1, 1]
+        for s, weights in zip(ss, weightss):
+            print(f's={s} weights={weights}')
+            for m, n in zip(ms, ns):
+                qs_ref = Quasisymmetry1(vmec, s, m=m, n=n, weights=weights)
+                qs = Quasisymmetry1(vmec, s, m=m, n=n, weights=weights, ntheta=65, nphi=57)
+                total_ref = qs_ref.total()
+                total = qs.total()
+                print(f'm={m} n={n} qs_ref={qs_ref.total()}, qs={qs.total()} diff={total_ref-total}')
+                np.testing.assert_allclose(qs_ref.total(), qs.total(), atol=0, rtol=1e-2)
+
+    def test_good_qa(self):
+        """
+        For a configuration that is known to have good quasi-axisymmetry,
+        the QA error should have a low value. The QH and QP errors should be larger.
+        """
+        vmec = Vmec(os.path.join(TEST_DIR, 'input.simsopt_nfp2_QA_20210328-01-020_000_000251'))
+        qa = Quasisymmetry1(vmec, 0.5, m=1, n=0)
+        total = qa.total()
+        print("QA error for simsopt_nfp2_QA_20210328-01-020_000_000251:", total)
+        self.assertTrue(total < 2e-5)
+
+        qh = Quasisymmetry1(vmec, 0.5, m=1, n=1)
+        print('QH error:', qh.total())
[…]
