### 2021-07-04 1d4bdb2
Fix CI to use new directory for ubuntu.jason in VMEC2000
