+++ .github/workflows/ci.yml
@@ -168,7 +168,7 @@ jobs:
     - name: Configure and install VMEC2000 module
       run: |
         cd VMEC2000
-        cp python/machines/ubuntu.json cmake_config_file.json
+        cp cmake/machines/ubuntu.json cmake_config_file.json
         cat cmake_config_file.json
         pip install .
 
