### 2021-07-04 836b354
Renamed Quasisymmetry1 to QuasisymmetryRatioError and added some to docs
