### 2021-07-04 18b55a2
Renamed Quasisymmetry1 to QuasisymmetryRatioError and added some to docs
