+++ tests/mhd/test_vmec_diagnostics.py
@@ -3,7 +3,7 @@ import numpy as np
 import os
 import logging
 
-from simsopt.mhd.vmec_diagnostics import Quasisymmetry1
+from simsopt.mhd.vmec_diagnostics import QuasisymmetryRatioError
 
 try:
     import vmec
@@ -24,7 +24,7 @@ logger = logging.getLogger(__name__)
 
 
 @unittest.skipIf(vmec is None, "vmec python package is not found")
-class Quasisymmetry1Tests(unittest.TestCase):
+class QuasisymmetryRatioErrorTests(unittest.TestCase):
     def test_axisymmetry(self):
         """
         For an axisymmetric field, the QA error should be 0, while the QH
@@ -34,16 +34,16 @@ class Quasisymmetry1Tests(unittest.TestCase):
         vmec = Vmec(os.path.join(TEST_DIR, 'input.circular_tokamak'))
 
         for s in [[0.5], [0.3, 0.6]]:
-            qa = Quasisymmetry1(vmec, s, m=1, n=0)
+            qa = QuasisymmetryRatioError(vmec, s, m=1, n=0)
             residuals = qa.residuals()
             print('max QA error:', np.max(np.abs(residuals)))
             np.testing.assert_allclose(residuals, np.zeros_like(residuals), atol=2e-6)
 
-            qh = Quasisymmetry1(vmec, s, m=1, n=1)
+            qh = QuasisymmetryRatioError(vmec, s, m=1, n=1)
             print('QH error:', qh.total())
             self.assertTrue(qh.total() > 0.01)
 
-            qp = Quasisymmetry1(vmec, s, m=0, n=1)
+            qp = QuasisymmetryRatioError(vmec, s, m=0, n=1)
             print('QP error:', qp.total())
             self.assertTrue(qp.total() > 0.01)
 
@@ -61,8 +61,8 @@ class Quasisymmetry1Tests(unittest.TestCase):
         for s, weights in zip(ss, weightss):
             print(f's={s} weights={weights}')
             for m, n in zip(ms, ns):
-                qs_ref = Quasisymmetry1(vmec, s, m=m, n=n, weights=weights)
-                qs = Quasisymmetry1(vmec, s, m=m, n=n, weights=weights, ntheta=65, nphi=57)
+                qs_ref = QuasisymmetryRatioError(vmec, s, m=m, n=n, weights=weights)
+                qs = QuasisymmetryRatioError(vmec, s, m=m, n=n, weights=weights, ntheta=65, nphi=57)
                 total_ref = qs_ref.total()
                 total = qs.total()
                 print(f'm={m} n={n} qs_ref={qs_ref.total()}, qs={qs.total()} diff={total_ref-total}')
@@ -74,16 +74,16 @@ class Quasisymmetry1Tests(unittest.TestCase):
         the QA error should have a low value. The QH and QP errors should be larger.
         """
         vmec = Vmec(os.path.join(TEST_DIR, 'input.simsopt_nfp2_QA_20210328-01-020_000_000251'))
-        qa = Quasisymmetry1(vmec, 0.5, m=1, n=0)
+        qa = QuasisymmetryRatioError(vmec, 0.5, m=1, n=0)
         total = qa.total()
         print("QA error for simsopt_nfp2_QA_20210328-01-020_000_000251:", total)
         self.assertTrue(total < 2e-5)
 
-        qh = Quasisymmetry1(vmec, 0.5, m=1, n=1)
+        qh = QuasisymmetryRatioError(vmec, 0.5, m=1, n=1)
[…]
