### 2021-07-04 e14c446
Slight redefinition of QuasisymmetryRatioError so it works even when iota=0
