+++ setup.cfg
@@ -3,7 +3,7 @@ name = simsopt
 description: Simplified framework for optimizing stellarators
 long_description = file: README.md
 long_description_content_type = text/markdown
-version = 0.3.06
+version = 0.4.0
 license = GNU Lesser General Public License Version 3
 license_files = COPYING, COPYING.LESSER
 author = "Matt Landreman, Florian Wechsung, Bharat Medasani"
