### 2021-07-05 805e41f
QuasisymmetryRatioError: refactored to allow users to access intermediate steps of the computation, and added weight to focus on regions where B dot grad B vanishes
