### 2021-07-05 0c3ac78
QuasisymmetryRatioError: added tests
