+++ tests/mhd/test_vmec_diagnostics.py
@@ -58,15 +58,21 @@ class QuasisymmetryRatioErrorTests(unittest.TestCase):
         weightss = [None, [1], None, [2.2, 5.5, 0.9]]
         ms = [1, 1, 1, 0]
         ns = [0, 1, -1, 1]
-        for s, weights in zip(ss, weightss):
-            print(f's={s} weights={weights}')
-            for m, n in zip(ms, ns):
-                qs_ref = QuasisymmetryRatioError(vmec, s, m=m, n=n, weights=weights)
-                qs = QuasisymmetryRatioError(vmec, s, m=m, n=n, weights=weights, ntheta=65, nphi=57)
-                total_ref = qs_ref.total()
-                total = qs.total()
-                print(f'm={m} n={n} qs_ref={qs_ref.total()}, qs={qs.total()} diff={total_ref-total}')
-                np.testing.assert_allclose(qs_ref.total(), qs.total(), atol=0, rtol=1e-2)
+        weight_funcs = [None, 'exp', 'lorentzian']
+        weight_facs = [0, 10.0, 10.0]
+        for weight_func, weight_fac in zip(weight_funcs, weight_facs):
+            for s, weights in zip(ss, weightss):
+                print(f's={s} weights={weights}')
+                for m, n in zip(ms, ns):
+                    qs_ref = QuasisymmetryRatioError(vmec, s, m=m, n=n, weights=weights, ntheta=130, nphi=114,
+                                                     weight_func=weight_func, weight_fac=weight_fac)
+                    qs = QuasisymmetryRatioError(vmec, s, m=m, n=n, weights=weights, ntheta=65, nphi=57,
+                                                 weight_func=weight_func, weight_fac=weight_fac)
+                    total_ref = qs_ref.total()
+                    total = qs.total()
+                    print(f'm={m} n={n} weight_func={weight_func} '
+                          f'qs_ref={qs_ref.total()}, qs={qs.total()} diff={total_ref-total}')
+                    np.testing.assert_allclose(qs_ref.total(), qs.total(), atol=0, rtol=1e-2)
 
     def test_good_qa(self):
         """
@@ -116,26 +122,35 @@ class QuasisymmetryRatioErrorTests(unittest.TestCase):
         The quasisymmetry error should be unchanged under a scaling of the
         configuration in length and/or magnetic field strength.
         """
-        vmec = Vmec(os.path.join(TEST_DIR, 'input.li383_low_res'))
-        qs1 = QuasisymmetryRatioError(vmec, [0, 0.7, 1], m=1, n=-1, weights=[0.8, 1.1, 0.9])
-        residuals1 = qs1.residuals()
+        weight_funcs = [None, 'power', 'exp', 'lorentzian']
+        weight_facs = [0, 100, 100.0, 30.0]
+        for weight_func, weight_fac in zip(weight_funcs, weight_facs):
+            vmec = Vmec(os.path.join(TEST_DIR, 'input.li383_low_res'))
+            qs1 = QuasisymmetryRatioError(vmec, [0, 0.7, 1], m=1, n=-1, weights=[0.8, 1.1, 0.9])
+            results1 = qs1.compute()
+            residuals1 = qs1.residuals()
 
-        R_scale = 2.7
-        B_scale = 1.6
+            R_scale = 2.7
+            B_scale = 1.6
 
-        # Now scale the vmec configuration:
-        vmec.boundary.rc *= R_scale
-        vmec.boundary.zs *= R_scale
-        vmec.indata.phiedge *= B_scale * R_scale * R_scale
-        vmec.indata.pres_scale *= B_scale * B_scale
[…]
