### 2021-07-05 0772e1c
tests and doc strings
