+++ tests/field/test_particle.py
@@ -58,7 +58,7 @@ class ParticleTracingTesting(unittest.TestCase):
         phirange = (0, 2*np.pi, n*6)
         zrange = (-0.3, 0.3, n)
         bsh = InterpolatedField(
-            bs, UniformInterpolationRule(4),
+            bs, UniformInterpolationRule(5),
             rrange, phirange, zrange, True
         )
         self.bsh = bsh
@@ -71,20 +71,24 @@ class ParticleTracingTesting(unittest.TestCase):
         nparticles = 2
         m = PROTON_MASS
         q = ELEMENTARY_CHARGE
+        Ekin = 1000*ONE_EV
         np.random.seed(1)
+        umin = 0.25
+        umax = 0.75
+        tmax = 2e-5
         with self.assertRaises(RuntimeError):
             gc_tys, gc_phi_hits = trace_particles_starting_on_axis(
-                ma.gamma(), bsh, nparticles, tmax=1e-4, seed=1, mass=m, charge=q,
-                Ekin=9000, umin=-0.1, umax=+0.1,
+                ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
+                Ekin=Ekin, umin=umin, umax=umax,
                 phis=[], mode='gc')
 
         gc_tys, gc_phi_hits = trace_particles_starting_on_axis(
-            ma.gamma(), bsh, nparticles, tmax=1e-6, seed=1, mass=m, charge=q,
-            Ekin=9000, umin=-0.1, umax=+0.1,
+            ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
+            Ekin=Ekin, umin=umin, umax=umax,
             phis=[], mode='gc_vac')
         fo_tys, fo_phi_hits = trace_particles_starting_on_axis(
-            ma.gamma(), bsh, nparticles, tmax=1e-6, seed=1, mass=m, charge=q,
-            Ekin=9000, umin=-0.1, umax=+0.1,
+            ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
+            Ekin=Ekin, umin=umin, umax=umax,
             phis=[], mode='full')
         particles_to_vtk(gc_tys, '/tmp/particles_gc')
         particles_to_vtk(fo_tys, '/tmp/particles_fo')
@@ -107,9 +111,9 @@ class ParticleTracingTesting(unittest.TestCase):
                 v = fo_ty[jdx, 4:]
                 Bunit = Bs[j, :]/AbsBs[j, 0]
                 vperp = np.linalg.norm(v - np.sum(v*Bunit)*Bunit)
-                r = compute_gc_radius(m, vperp, 1.6e-19, AbsBs[j, 0])
+                r = compute_gc_radius(m, vperp, q, AbsBs[j, 0])
                 dist = np.linalg.norm(gc_xyzs[j, :] - fo_ty[jdx, 1:4])
-                assert dist < 4*r
+                assert dist < 8*r
                 assert dist > 0.5*r
 
     def test_guidingcenterphihits(self):
@@ -118,6 +122,7 @@ class ParticleTracingTesting(unittest.TestCase):
         nparticles = 2
         m = PROTON_MASS
         q = ELEMENTARY_CHARGE
+        Ekin = 9000 * ONE_EV
         nphis = 10
         phis = np.linspace(0, 2*np.pi, nphis, endpoint=False)
         mpol = 5
@@ -136,8 +141,8 @@ class ParticleTracingTesting(unittest.TestCase):
         np.random.seed(1)
[…]
