### 2021-07-05 ba9a8d2
tests and doc strings
