### 2021-07-05 fa7f39d
test for stopping criteria
