### 2021-07-05 ce8440d
boost optional in extensive ci
