+++ .github/workflows/extensive_ci.yml
@@ -31,6 +31,7 @@ jobs:
         test-type: [unit, integrated]
         spec: [true, false]
         vmec: [true, false]
+        boost: [true, false]
 
     steps:
     # First print out lots of information. We do this in separate 
@@ -55,7 +56,11 @@ jobs:
     - name: apt-get stuff needed for libstell and vmec
       run: |
         sudo apt-get update
-        sudo apt-get install -y build-essential gfortran openmpi-bin libopenmpi-dev libnetcdf-dev libnetcdff-dev liblapack-dev libblacs-mpi-dev libscalapack-mpi-dev libhdf5-dev libhdf5-serial-dev git m4 libfftw3-dev libboost-all-dev
+        sudo apt-get install -y build-essential gfortran openmpi-bin libopenmpi-dev libnetcdf-dev libnetcdff-dev liblapack-dev libblacs-mpi-dev libscalapack-mpi-dev libhdf5-dev libhdf5-serial-dev git m4 libfftw3-dev
+    
+    - name: install boost if required
+      if: matrix.boost == true
+      run: sudo apt-get install -y libboost-all-dev
 
     - uses: actions/checkout@v2
       # If we want submodules downloaded, uncomment the next 2 lines:
