### 2021-07-05 2f8763a
more graceful handling when compiling without boost
