### 2021-07-05 be14fea
more graceful handling when compiling without boost
