### 2021-07-05 3d4eef2
gracefully exit the tracing example
