### 2021-07-05 223925d
reduce memory usage
