### 2021-07-05 2255d61
poincare plot, more tests
