### 2021-07-05 2bed4cd
poincare plot, more tests
