### 2021-07-06 #3402 @landreman
@dfm thank you for serving as editor. For possible reviewers, you could look at the lead developers for https://github.com/boutproject/BOUT-dev/graphs/contributors (ZedThree, bendudson, d7919, ...) or https://github.com/bendudson/freegs/graphs/contributors or https://github.com/PlasmaPy/PlasmaPy/graphs/contributors (namurphy, StanczakDominik, ...). We could also look at the speakers at http://meetings.aps.org/Meeting/DPP20/Session/JM10 and http://meetings.aps.org/Meeting/DPP20/Session/NM10
