### 2021-07-06 e747dd6
Trying pybind11 metaclass
