### 2021-07-06 ce95142
add mpi tracing tests
