### 2021-07-06 c3d3578
Bug fix in _core.util
