+++ src/simsopt/field/tracing.py
@@ -75,7 +75,7 @@ def parallel_loop_bounds(comm, n):
 def trace_particles(field: MagneticField, xyz_inits: NDArray[Float],
                     parallel_speeds: NDArray[Float], tmax=1e-4,
                     mass=ALPHA_PARTICLE_MASS, charge=ALPHA_PARTICLE_CHARGE, Ekin=FUSION_ALPHA_PARTICLE_ENERGY,
-                    tol=1e-9, comm=None, phis=[], stopping_criteria=[], mode='gc_vac'):
+                    tol=1e-9, comm=None, phis=[], stopping_criteria=[], mode='gc_vac', forget_exact_path=False):
     r"""
     Follow particles in a magnetic field.
 
@@ -116,6 +116,8 @@ def trace_particles(field: MagneticField, xyz_inits: NDArray[Float],
             `gc`: general guiding center equations,
             `gc_vac`: simplified guiding center equations for the case :math:`\nabla p=0`,
             `full`: full orbit calculation (slow!)
+        forget_exact_path: return an empty list for the ``res_tys``. To be used when only res_phi_hits
+                           is of interest and one wants to reduce memory usage.
 
     Returns: 2 element tuple containing
         - ``res_tys``:
@@ -161,7 +163,10 @@ def trace_particles(field: MagneticField, xyz_inits: NDArray[Float],
             res_ty, res_phi_hit = sopp.particle_fullorbit_tracing(
                 field, xyz_inits[i, :], v_inits[i, :],
                 m, charge, tmax, tol, phis=phis, stopping_criteria=stopping_criteria)
-        res_tys.append(np.asarray(res_ty))
+        if not forget_exact_path:
+            res_tys.append(np.asarray(res_ty))
+        else:
+            res_tys.append([])
         res_phi_hits.append(np.asarray(res_phi_hit))
         dtavg = res_ty[-1][0]/len(res_ty)
         logger.debug(f"{i+1:3d}/{nparticles}, t_final={res_ty[-1][0]}, average timestep {1000*dtavg:.10f}ms")
@@ -179,7 +184,7 @@ def trace_particles(field: MagneticField, xyz_inits: NDArray[Float],
 def trace_particles_starting_on_axis(axis, field, nparticles, tmax=1e-4,
                                      mass=ALPHA_PARTICLE_MASS, charge=ALPHA_PARTICLE_CHARGE, Ekin=FUSION_ALPHA_PARTICLE_ENERGY,
                                      tol=1e-9, comm=None, seed=1, umin=-1, umax=+1,
-                                     phis=[], stopping_criteria=[], mode='gc_vac'):
+                                     phis=[], stopping_criteria=[], mode='gc_vac', forget_exact_path=False):
     r"""
     Follows particles spawned at random locations on the magnetic axis with random pitch angle.
     See :mod:`simsopt.field.tracing.trace_particles` for the governing equations.
@@ -207,6 +212,8 @@ def trace_particles_starting_on_axis(axis, field, nparticles, tmax=1e-4,
             `gc`: general guiding center equations,
             `gc_vac`: simplified guiding center equations for the case :math:`\nabla p=0`,
             `full`: full orbit calculation (slow!)
+        forget_exact_path: return an empty list for the ``res_tys``. To be used when only res_phi_hits
+                           is of interest and one wants to reduce memory usage.
 
     Returns: see :mod:`simsopt.field.tracing.trace_particles`
     """
[…]
