### 2021-07-07 a44894c
better mpi handling in extensive_ci
