+++ .github/workflows/ci.yml
@@ -195,14 +195,14 @@ jobs:
     - name: Run tests on 1 process using coverage
       if: contains(matrix.test-type, 'unit')
       run: | 
-        coverage run --source=simsopt -m unittest discover -s tests
+        coverage run --source=simsopt -m unittest discover -s tests -v
 
     - name: Run MPI tests using coverage
       if: contains(matrix.test-type, 'unit')
       run: | 
-        mpiexec -n 1 coverage run -m unittest discover -k "mpi" -s tests
-        mpiexec -n 2 coverage run -m unittest discover -k "mpi" -s tests
-        mpiexec -n 3 coverage run -m unittest discover -k "mpi" -s tests
+        mpiexec -n 1 coverage run -m unittest discover -k "mpi" -s tests -v
+        mpiexec -n 2 coverage run -m unittest discover -k "mpi" -s tests -v
+        mpiexec -n 3 coverage run -m unittest discover -k "mpi" -s tests -v
 
     - name: Combine coverage reports
       if: contains(matrix.test-type, 'unit')
