+++ .github/workflows/extensive_ci.yml
@@ -246,14 +246,14 @@ jobs:
     - name: Run tests on 1 MPI process using coverage
       if: "contains(matrix.test-type, 'unit')"
       run: | 
-        coverage run --source=simsopt -m unittest discover -s tests
+        coverage run --source=simsopt -m unittest discover -s tests -v
 
     - name: Run MPI tests using coverage
       if: contains(matrix.test-type, 'unit') && (matrix.vmec == true || matrix.spec == true || matrix.boost == true)
       run: | 
-        mpiexec -n 1 coverage run -m unittest discover -k "mpi" -s tests
-        mpiexec -n 2 coverage run -m unittest discover -k "mpi" -s tests
-        mpiexec -n 3 coverage run -m unittest discover -k "mpi" -s tests
+        mpiexec -n 1 coverage run -m unittest discover -k "mpi" -s tests -v
+        mpiexec -n 2 coverage run -m unittest discover -k "mpi" -s tests -v
+        mpiexec -n 3 coverage run -m unittest discover -k "mpi" -s tests -v
 
     - name: Upload uncombined coverage to github
       if: "contains(matrix.test-type, 'unit')"
