### 2021-07-07 e383805
fewer quadpoints to reduce memory usage
