### 2021-07-07 4f8accb
Comments updated for graph classes
