### 2021-07-07 ed5d602
split tracing example
