+++ examples/1_Simple/tracing_fieldline.py
@@ -4,7 +4,7 @@ from simsopt.field.magneticfieldclasses import InterpolatedField, UniformInterpo
 from simsopt.geo.surfacexyztensorfourier import SurfaceRZFourier
 from simsopt.geo.coilcollection import CoilCollection
 from simsopt.field.tracing import trace_particles_starting_on_axis, SurfaceClassifier, \
-    particles_to_vtk, compute_fieldlines, LevelsetStoppingCriterion
+    particles_to_vtk, compute_fieldlines, LevelsetStoppingCriterion, plot_poincare_data
 from simsopt.geo.curve import curves_to_vtk
 from simsopt.util.zoo import get_ncsx_data
 from simsopt.util.constants import ALPHA_PARTICLE_MASS, ALPHA_PARTICLE_CHARGE, FUSION_ALPHA_PARTICLE_ENERGY
@@ -33,8 +33,8 @@ logger.setLevel(1)
 ci = "CI" in os.environ and os.environ['CI'].lower() in ['1', 'true']
 nfieldlines = 3 if ci else 80
 tmax_fl = 10000 if ci else 40000
-nparticles = 3 if ci else 40
-degree = 2 if ci else 4
+nparticles = 3 if ci else 400
+degree = 2 if ci else 3
 
 
 """
@@ -44,14 +44,6 @@ guiding center trajectories of particles
 
 
 coils, currents, ma = get_ncsx_data(Nt_coils=25, Nt_ma=10)
-
-# scale up the device to larger major radius and stronger B field
-scale = 8.
-for c in coils:
-    c.set_dofs(scale*c.get_dofs())
-ma.set_dofs(scale*ma.get_dofs())
-currents = [32 * c for c in currents]
-
 stellarator = CoilCollection(coils, currents, 3, True)
 coils = stellarator.coils
 currents = stellarator.currents
@@ -68,19 +60,18 @@ phis = np.linspace(0, 1, nfp*2*ntor+1, endpoint=False)
 thetas = np.linspace(0, 1, 2*mpol+1, endpoint=False)
 s = SurfaceRZFourier(
     mpol=mpol, ntor=ntor, stellsym=stellsym, nfp=nfp, quadpoints_phi=phis, quadpoints_theta=thetas)
-s.fit_to_curve(ma, scale*0.70, flip_theta=False)
+s.fit_to_curve(ma, 0.70, flip_theta=False)
+
 s.to_vtk('/tmp/surface')
-sc_fieldline = SurfaceClassifier(s, h=scale*0.1, p=2)
-sc_fieldline.to_vtk('/tmp/levelset', h=scale*0.02)
-s.fit_to_curve(ma, scale*0.20, flip_theta=False)
-sc_particle = SurfaceClassifier(s, h=scale*0.1, p=2)
+sc_fieldline = SurfaceClassifier(s, h=0.1, p=2)
+sc_fieldline.to_vtk('/tmp/levelset', h=0.02)
 
 
 def trace_fieldlines(bfield, label):
     t1 = time.time()
-    phis = [i*2*np.pi/(4*3) for i in range(4)]
-    R0 = [ma.gamma()[0, 0] + i*0.015 for i in range(nfieldlines)]
+    R0 = [ma.gamma()[0, 0] + i*0.002 for i in range(nfieldlines)]
     Z0 = [ma.gamma()[0, 2] for i in range(nfieldlines)]
+    phis = [(i/4)*(2*np.pi/nfp) for i in range(4)]
     fieldlines_tys, fieldlines_phi_hits = compute_fieldlines(
         bfield, R0, Z0, tmax=tmax_fl, tol=1e-7, comm=comm,
         phis=phis, stopping_criteria=[LevelsetStoppingCriterion(sc_fieldline.dist)])
[…]
