+++ examples/1_Simple/tracing_particle.py
@@ -1,0 +1,98 @@
+#!/usr/bin/env python3
+from simsopt.field.biotsavart import BiotSavart
+from simsopt.field.magneticfieldclasses import InterpolatedField, UniformInterpolationRule
+from simsopt.geo.surfacexyztensorfourier import SurfaceRZFourier
+from simsopt.geo.coilcollection import CoilCollection
+from simsopt.field.tracing import trace_particles_starting_on_axis, SurfaceClassifier, \
+    particles_to_vtk, LevelsetStoppingCriterion, plot_poincare_data
+from simsopt.geo.curve import curves_to_vtk
+from simsopt.util.zoo import get_ncsx_data
+from simsopt.util.constants import PROTON_MASS, ELEMENTARY_CHARGE, ONE_EV
+import simsoptpp as sopp
+try:
+    from mpi4py import MPI
+    comm = MPI.COMM_WORLD
+except ImportError:
+    comm = None
+
+import numpy as np
+import time
+import os
+import logging
+import sys
+if not sopp.with_boost():
+    print("Please compile with boost to run this example.")
+    sys.exit(0)
+sys.path.append(os.path.join("..", "tests", "geo"))
+logging.basicConfig()
+logger = logging.getLogger('simsopt.field.tracing')
+logger.setLevel(1)
+
+# check whether we're in CI, in that case we make the run a bit cheaper
+ci = "CI" in os.environ and os.environ['CI'].lower() in ['1', 'true']
+nparticles = 3 if ci else 40
+degree = 2 if ci else 3
+
+
+"""
+This examples demonstrate how to use SIMSOPT to compute guiding center
+trajectories of particles
+"""
+
+
+coils, currents, ma = get_ncsx_data(Nt_coils=25, Nt_ma=10)
+
+stellarator = CoilCollection(coils, currents, 3, True)
+coils = stellarator.coils
+currents = stellarator.currents
+bs = BiotSavart(coils, currents)
+print("Mean(|B|) on axis =", np.mean(np.linalg.norm(bs.set_points(ma.gamma()).B(), axis=1)))
+print("Mean(Axis radius) =", np.mean(np.linalg.norm(ma.gamma(), axis=1)))
+curves_to_vtk(coils + [ma], '/tmp/coils')
+
+mpol = 5
+ntor = 5
+stellsym = True
+nfp = 3
+phis = np.linspace(0, 1, nfp*2*ntor+1, endpoint=False)
+thetas = np.linspace(0, 1, 2*mpol+1, endpoint=False)
+s = SurfaceRZFourier(
+    mpol=mpol, ntor=ntor, stellsym=stellsym, nfp=nfp, quadpoints_phi=phis, quadpoints_theta=thetas)
+
+
+s.fit_to_curve(ma, 0.20, flip_theta=False)
+sc_particle = SurfaceClassifier(s, h=0.1, p=2)
+n = 16
+rs = np.linalg.norm(s.gamma()[:, :, 0:2], axis=2)
+zs = s.gamma()[:, :, 2]
+rrange = (np.min(rs), np.max(rs), n)
+phirange = (0, 2*np.pi, n*6)
+zrange = (np.min(zs), np.max(zs), n)
+bsh = InterpolatedField(
+    bs, UniformInterpolationRule(degree-1),
+    rrange, phirange, zrange, True
+)
+
+
+def trace_particles(bfield, label, mode='gc_vac'):
+    t1 = time.time()
+    phis = [(i/4)*(2*np.pi/nfp) for i in range(4)]
+    gc_tys, gc_phi_hits = trace_particles_starting_on_axis(
[…]
