### 2021-07-07 c46d655
speed up dnormal_by_dcoeff_impl
