### 2021-07-07 e3b34d1
results on different threads are a little different -> gave up on making sure the path is exactly identical and just checking that the phi hits are the same now
