### 2021-07-07 372af7a
results on different threads are a little different -> gave up on making sure the path is exactly identical and just checking that the phi hits are the same now
