+++ tests/geo/test_surface_objectives.py
@@ -148,7 +148,7 @@ class QfmTests(unittest.TestCase):
         for surfacetype in surfacetypes_list:
             for stellsym in stellsym_list:
                 with self.subTest(surfacetype=surfacetype, stellsym=stellsym):
-                    # self.subtest_qfm1(surfacetype, stellsym)
+                    self.subtest_qfm1(surfacetype, stellsym)
                     self.subtest_qfm2(surfacetype, stellsym)
 
     def subtest_qfm1(self, surfacetype, stellsym):
@@ -167,7 +167,7 @@ class QfmTests(unittest.TestCase):
             s.set_dofs(dofs)
             return qfm.dJ_by_dsurfacecoefficients()
         taylor_test1(f, df, coeffs,
-                     epsilons=np.power(2., -np.asarray(range(12, 22))))
+                     epsilons=np.power(2., -np.asarray(range(13, 22))))
 
     def subtest_qfm2(self, surfacetype, stellsym):
         coils, currents, ma = get_ncsx_data()
@@ -191,4 +191,4 @@ class QfmTests(unittest.TestCase):
             return qfm.d2J_by_dsurfacecoefficientsdsurfacecoefficients()
 
         taylor_test2(f, df, d2f, coeffs,
-                     epsilons=np.power(2., -np.asarray(range(10, 20))))
+                     epsilons=np.power(2., -np.asarray(range(12, 20))))
