### 2021-07-07 fe1a506
make unittest verbose
