+++ .github/workflows/conda.yml
@@ -43,6 +43,3 @@ jobs:
 
       - name: Install
         run: conda install -c ${CONDA_PREFIX}/conda-bld/ simsopt
-
-      - name: Test
-        run: ./run_tests
