+++ conda.recipe/meta.yaml
@@ -46,7 +46,7 @@ test:
   source_files:
     - tests
   commands:
-    - python -m unittest discover -s tests -v
+    - ./run_tests
 
 about:
   home: {{ data.get('url') }}
