### 2021-07-08 bc3bacf
move conda testing into workflow
