### 2021-07-08 ce1a2f6
move conda testing into workflow
