### 2021-07-08 ba686ac
prettier plots in the examples
