+++ examples/1_Simple/tracing_particle.py
@@ -30,7 +30,7 @@ logger.setLevel(1)
 
 # check whether we're in CI, in that case we make the run a bit cheaper
 ci = "CI" in os.environ and os.environ['CI'].lower() in ['1', 'true']
-nparticles = 3 if ci else 40
+nparticles = 3 if ci else 100
 degree = 2 if ci else 3
 
 
