### 2021-07-08 d832308
Tweak to constrained objective and lambda in qfm example.
