### 2021-07-08 e8cf362
Tweak to constrained objective and lambda in qfm example.
