+++ conda.recipe/conda_build_config.yaml
@@ -1,6 +1,2 @@
 numpy:
   - 1.19
-boost: 1.76
-pin_run_as_build:
-    boost:
-      max_pin: x.x
