+++ conda.recipe/meta.yaml
@@ -18,6 +18,7 @@ requirements:
     - "{{ compiler('cxx') }}"
     - cmake
     - ninja
+    - boost {{ boost }}
     # - numpy {{ numpy }}
     # - python 
     # - pip
@@ -38,6 +39,7 @@ requirements:
     - pandas >=1.1.4
     - monty >=2021.6.10 
     - ruamel.yaml
+    - boost
 
 about:
   home: {{ data.get('url') }}
