### 2021-07-08 20a70a3
make example a bit faster
