### 2021-07-08 fed7404
Poloidal magnetic field: testing, poincaree, particle
Added new magnetic field class: poloidal magnetic field. This allows for extra testing of the poincare plots and the particle tracer code to check if the canonical angular momentum is conserved.
