+++ src/simsopt/field/magneticfieldclasses.py
@@ -104,6 +104,97 @@ class ToroidalField(MagneticField):
                 np.zeros((3, 3, len(points)))])).transpose((3, 0, 1, 2))
 
 
+class PoloidalField(MagneticField):
+    ''' Magnetic field purely in the poloidal direction, that is, in the theta direction of a poloidal-toroidal coordinate system.
+       Its modulus is given by B = B0*r/(R0*q) so that, together with the toroidal field, it creates a safety factor equals to q
+    Args:
+        B0: modulus of the magnetic field at R0
+        R0: major radius of the magnetic axis
+        q:  safety factor/pitch angle of the magnetic field lines
+    '''
+
+    def __init__(self, R0, B0, q):
+        MagneticField.__init__(self)
+        self.R0 = R0
+        self.B0 = B0
+        self.q = q
+
+    def _B_impl(self, B):
+        points = self.get_points_cart_ref()
+
+        x = points[:, 0]
+        y = points[:, 1]
+        z = points[:, 2]
+
+        phi = np.arctan2(y, x)
+        theta = np.arctan2(z, np.sqrt(x**2+y**2)-self.R0)
+        r = np.sqrt((np.sqrt(x**2+y**2)-self.R0)**2+z**2)
+        thetaUnitVectorOver_times_r = np.vstack((-np.multiply(np.sin(theta), r)*np.cos(phi), -np.multiply(np.sin(theta), r)*np.sin(phi), np.multiply(np.cos(theta), r))).T
+        B[:] = self.B0/self.R0/self.q*thetaUnitVectorOver_times_r
+
+    def _dB_by_dX_impl(self, dB):
+        points = self.get_points_cart_ref()
+
+        x = points[:, 0]
+        y = points[:, 1]
+        z = points[:, 2]
+
+        phi = np.arctan2(y, x)
+        theta = np.arctan2(z, np.sqrt(x**2+y**2)-self.R0)
+        r = np.sqrt((np.sqrt(x**2+y**2)-self.R0)**2+z**2)
+
+        dtheta_by_dX1 = -((x*z)/(np.sqrt(x**2+y**2)*(x**2+y**2+z**2-2*np.sqrt(x**2+y**2)*self.R0+(self.R0)**2)))
+        dtheta_by_dX2 = -((y*z)/(np.sqrt(x**2+y**2)*(x**2+y**2+z**2-2*np.sqrt(x**2+y**2)*self.R0+(self.R0)**2)))
+        dtheta_by_dX3 = 1/((-self.R0+np.sqrt(x**2+y**2))*(1+z**2/(self.R0-np.sqrt(x**2+y**2))**2))
+
+        dphi_by_dX1 = -(y/(x**2 + y**2))
+        dphi_by_dX2 = x/(x**2 + y**2)
+        dphi_by_dX3 = 0.*z
+
+        dthetaunitvector_by_dX1 = np.vstack((
+            -np.cos(theta)*np.cos(phi)*dtheta_by_dX1+np.sin(theta)*np.sin(phi)*dphi_by_dX1,
+            -np.cos(theta)*np.sin(phi)*dtheta_by_dX1-np.sin(theta)*np.cos(phi)*dphi_by_dX1,
+            -np.sin(theta)*dtheta_by_dX1
+        )).T
+        dthetaunitvector_by_dX2 = np.vstack((
+            -np.cos(theta)*np.cos(phi)*dtheta_by_dX2+np.sin(theta)*np.sin(phi)*dphi_by_dX2,
[…]
