### 2021-07-08 de91f55
added proper uniform sampling from a curve and surface
