+++ examples/1_Simple/tracing_fieldline.py
@@ -4,11 +4,10 @@ from simsopt.field.biotsavart import BiotSavart
 from simsopt.field.magneticfieldclasses import InterpolatedField, UniformInterpolationRule
 from simsopt.geo.surfacexyztensorfourier import SurfaceRZFourier
 from simsopt.geo.coilcollection import CoilCollection
-from simsopt.field.tracing import trace_particles_starting_on_axis, SurfaceClassifier, \
+from simsopt.field.tracing import SurfaceClassifier, \
     particles_to_vtk, compute_fieldlines, LevelsetStoppingCriterion, plot_poincare_data
 from simsopt.geo.curve import curves_to_vtk
 from simsopt.util.zoo import get_ncsx_data
-from simsopt.util.constants import ALPHA_PARTICLE_MASS, ALPHA_PARTICLE_CHARGE, FUSION_ALPHA_PARTICLE_ENERGY
 import simsoptpp as sopp
 try:
     from mpi4py import MPI
@@ -16,7 +15,6 @@ try:
 except ImportError:
     comm = None
 
-import matplotlib.pyplot as plt
 import numpy as np
 import time
 import os
