### 2021-07-08 4f3272b
added proper uniform sampling from a curve and surface
