+++ tests/field/test_particle.py
@@ -1,9 +1,9 @@
 from simsopt.geo.coilcollection import CoilCollection
 from simsopt.field.biotsavart import BiotSavart
 from simsopt.util.zoo import get_ncsx_data
-from simsopt.field.tracing import trace_particles_starting_on_axis, SurfaceClassifier, \
+from simsopt.field.tracing import trace_particles_starting_on_curve, SurfaceClassifier, \
     particles_to_vtk, LevelsetStoppingCriterion, compute_gc_radius, gc_to_fullorbit_initial_guesses, \
-    IterationStoppingCriterion
+    IterationStoppingCriterion, trace_particles_starting_on_surface
 from simsopt.geo.surfacerzfourier import SurfaceRZFourier
 from simsopt.field.magneticfieldclasses import InterpolatedField, UniformInterpolationRule
 from simsopt.util.constants import PROTON_MASS, ELEMENTARY_CHARGE, ONE_EV
@@ -91,17 +91,17 @@ class ParticleTracingTesting(unittest.TestCase):
         umax = 0.75
         tmax = 2e-5
         with self.assertRaises(RuntimeError):
-            gc_tys, gc_phi_hits = trace_particles_starting_on_axis(
-                ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
+            gc_tys, gc_phi_hits = trace_particles_starting_on_curve(
+                ma, bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
                 Ekin=Ekin, umin=umin, umax=umax,
                 phis=[], mode='gc')
 
-        gc_tys, gc_phi_hits = trace_particles_starting_on_axis(
-            ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
+        gc_tys, gc_phi_hits = trace_particles_starting_on_curve(
+            ma, bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
             Ekin=Ekin, umin=umin, umax=umax,
             phis=[], mode='gc_vac')
-        fo_tys, fo_phi_hits = trace_particles_starting_on_axis(
-            ma.gamma(), bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
+        fo_tys, fo_phi_hits = trace_particles_starting_on_curve(
+            ma, bsh, nparticles, tmax=tmax, seed=1, mass=m, charge=q,
             Ekin=Ekin, umin=umin, umax=umax,
             phis=[], mode='full')
         if with_evtk:
@@ -155,8 +155,8 @@ class ParticleTracingTesting(unittest.TestCase):
         assert sc.evaluate(ma.gamma()[:1, :]) > 0
         assert sc.evaluate(2*ma.gamma()[:1, :]) < 0
         np.random.seed(1)
-        gc_tys, gc_phi_hits = trace_particles_starting_on_axis(
-            ma.gamma(), bsh, nparticles, tmax=1e-4, seed=1, mass=m, charge=q,
+        gc_tys, gc_phi_hits = trace_particles_starting_on_curve(
+            ma, bsh, nparticles, tmax=1e-4, seed=1, mass=m, charge=q,
             Ekin=Ekin, umin=-0.1, umax=+0.1, phis=phis, mode='gc_vac',
             stopping_criteria=[LevelsetStoppingCriterion(sc)])
 
@@ -208,12 +208,12 @@ class ParticleTracingTesting(unittest.TestCase):
         Ekin = 100*ONE_EV
         np.random.seed(1)
 
[…]
