### 2021-07-08 725b291
get rid of vtk export in sampling test
