### 2021-07-08 9228a00
Curve classes adapted for graph based Optimizable framework
