+++ src/simsopt/geo/graph_curve.py
@@ -1,0 +1,740 @@
+from math import sin, cos
+from abc import abstractmethod
+
+import numpy as np
+from jax import vjp, jacfwd, jvp
+from .jit import jit
+import jax.numpy as jnp
+from monty.dev import requires
+
+try:
+    import matplotlib.pyplot as plt
+except ImportError:
+    plt = None
+
+try:
+    from myavi import mlab
+except ImportError:
+    mlab = None
+
+import simsoptpp as sopp
+from .._core.graph_optimizable import CPPOptimizable
+
+
+@jit
+def incremental_arclength_pure(d1gamma):
+    """
+    This function is used in a Python+Jax implementation of the curve arc length formula.
+    """
+
+    return jnp.linalg.norm(d1gamma, axis=1)
+
+
+incremental_arclength_vjp = jit(lambda d1gamma, v: vjp(lambda d1g: incremental_arclength_pure(d1g), d1gamma)[1](v)[0])
+
+
+@jit
+def kappa_pure(d1gamma, d2gamma):
+    """
+    This function is used in a Python+Jax implementation of formula for curvature.
+    """
+
+    return jnp.linalg.norm(jnp.cross(d1gamma, d2gamma), axis=1)/jnp.linalg.norm(d1gamma, axis=1)**3
+
+
+kappavjp0 = jit(lambda d1gamma, d2gamma, v: vjp(lambda d1g: kappa_pure(d1g, d2gamma), d1gamma)[1](v)[0])
+kappavjp1 = jit(lambda d1gamma, d2gamma, v: vjp(lambda d2g: kappa_pure(d1gamma, d2g), d2gamma)[1](v)[0])
+kappagrad0 = jit(lambda d1gamma, d2gamma: jacfwd(lambda d1g: kappa_pure(d1g, d2gamma))(d1gamma))
+kappagrad1 = jit(lambda d1gamma, d2gamma: jacfwd(lambda d2g: kappa_pure(d1gamma, d2g))(d2gamma))
+
+
+@jit
+def torsion_pure(d1gamma, d2gamma, d3gamma):
+    """
+    This function is used in a Python+Jax implementation of formula for torsion.
+    """
+
+    return jnp.sum(jnp.cross(d1gamma, d2gamma, axis=1) * d3gamma, axis=1) / jnp.sum(jnp.cross(d1gamma, d2gamma, axis=1)**2, axis=1)
+
+
+torsionvjp0 = jit(lambda d1gamma, d2gamma, d3gamma, v: vjp(lambda d1g: torsion_pure(d1g, d2gamma, d3gamma), d1gamma)[1](v)[0])
+torsionvjp1 = jit(lambda d1gamma, d2gamma, d3gamma, v: vjp(lambda d2g: torsion_pure(d1gamma, d2g, d3gamma), d2gamma)[1](v)[0])
+torsionvjp2 = jit(lambda d1gamma, d2gamma, d3gamma, v: vjp(lambda d3g: torsion_pure(d1gamma, d2gamma, d3g), d3gamma)[1](v)[0])
+
+
+class Curve(CPPOptimizable):
+    """
+    Curve  is a base class for various representations of curves in SIMSOPT
+    using the graph based Optimizable framework with external handling of DOFS
+    as well.
+    """
+
+    def __init__(self, **kwargs):
+        # Optimizable.__init__(self)
+        # self.dependencies = []
+        # fixed = np.full(len(self.get_dofs()), False)
+        CPPOptimizable.__init__(self, **kwargs)
+
+    @requires(plt is not None, "Install matplotlib to plot Curve object")
+    def plot(self, ax=None, show=True, plot_derivative=False, closed_loop=True,
+             color=None, linestyle=None):
+        """
[…]
