### 2021-07-08 489e0e3
Curve classes adapted for graph based Optimizable framework
