+++ src/simsopt/geo/graph_curve_objectives.py
@@ -1,0 +1,237 @@
+import numpy as np
+from jax import grad, vjp
+import jax.numpy as jnp
+from .jit import jit
+
+from .._core.graph_optimizable import Optimizable
+
+
+@jit
+def curve_length_pure(l):
+    """
+    This function is used in a Python+Jax implementation of the curve length formula.
+    """
+    return jnp.mean(l)
+
+
+class CurveLength(Optimizable):
+    r"""
+    CurveLength is a class that computes the length of a curve, i.e.
+
+    .. math::
+        J = \int_{\text{curve}}~dl.
+
+    """
+
+    def __init__(self, curve):
+        self.curve = curve
+        self.thisgrad = jit(lambda l: grad(curve_length_pure)(l))
+        super().__init__(opts_in=[curve])
+
+    def J(self):
+        """
+        This returns the value of the quantity.
+        """
+        return curve_length_pure(self.curve.incremental_arclength())
+
+    def dJ(self):
+        """
+        This returns the derivative of the quantity with respect to the curve dofs.
+        """
+        return self.curve.dincremental_arclength_by_dcoeff_vjp(
+            self.thisgrad(self.curve.incremental_arclength()))
+    return_fn_map = {'J': J, 'dJ': dJ}
+
+
+@jit
+def Lp_curvature_pure(kappa, gammadash, p, desired_kappa):
+    """
+    This function is used in a Python+Jax implementation of the curvature penalty term.
+    """
+    arc_length = jnp.linalg.norm(gammadash, axis=1)
+    return (1./p)*jnp.mean(jnp.maximum(kappa-desired_kappa, 0)**p * arc_length)
+
+
+class LpCurveCurvature(Optimizable):
+    r"""
+    This class computes a penalty term based on the :math:`L_p` norm
+    of the curve's curvature, and penalizes where the local curve curvature exceeds a threshold
+
+
+    .. math::
+        J = \frac{1}{p} \int_{\text{curve}} \text{max}(\kappa - \kappa_0, 0)^p ~dl
+
+    where :math:`\kappa_0` is a threshold curvature.  When ``desired_length`` is
+    provided, :math:`\kappa_0 = 2 \pi / \text{desired_length}`, otherwise :math:`\kappa_0=0`.
+    """
+
+    def __init__(self, curve, p, desired_length=None):
+        self.curve = curve
+        # self.depends_on = ["curve"]
+        super().__init__(opts_in=[curve])
+        if desired_length is None:
+            self.desired_kappa = 0
+        else:
+            radius = desired_length/(2*np.pi)
+            self.desired_kappa = 1/radius
+
+        self.J_jax = jit(lambda kappa, gammadash: Lp_curvature_pure(kappa, gammadash, p, self.desired_kappa))
+        self.thisgrad0 = jit(lambda kappa, gammadash: grad(self.J_jax, argnums=0)(kappa, gammadash))
+        self.thisgrad1 = jit(lambda kappa, gammadash: grad(self.J_jax, argnums=1)(kappa, gammadash))
+
+    def J(self):
+        """
+        This returns the value of the quantity.
+        """
+        return self.J_jax(self.curve.kappa(), self.curve.gammadash())
+
+    def dJ(self):
+        """
+        This returns the derivative of the quantity with respect to the curve dofs.
+        """
+        grad0 = self.thisgrad0(self.curve.kappa(), self.curve.gammadash())
[…]
