+++ conda.recipe/meta.yaml
@@ -25,11 +25,11 @@ requirements:
   host:
     - python
     - pip
-    - numpy >=1.19
+    - numpy  {{ numpy }}
 
   run:
     - python
-    - numpy >=1.19
+    - numpy  {{ numpy }}
     - jax >=0.2.4
     - jaxlib >=0.1.56
     - scipy >=1.5.4
