+++ conda.recipe/conda_build_config.yaml
@@ -1,2 +1,3 @@
 numpy:
   - 1.19
+  - 1.20
