### 2021-07-11 c996b51
Bug fix in graph curve
