### 2021-07-11 31b5bf4
Bug fix in graph curve
