### 2021-07-11 6e694b5
Unit test with external handling of dofs
