### 2021-07-12 f6263ae
add ability to exploit stellarator and rotational symmetry
