+++ src/simsoptpp/magneticfield_interpolated.h
@@ -10,6 +10,7 @@ class InterpolatedField : public MagneticField<T> {
         using typename MagneticField<T>::Tensor2;
     private:
 
+        CachedTensor<T, 2> points_cyl_sym;
         std::function<Vec(double, double, double)> f_B;
         std::function<Vec(Vec, Vec, Vec)> fbatch_B;
         std::function<Vec(Vec, Vec, Vec)> fbatch_GradAbsB;
@@ -17,30 +18,117 @@ class InterpolatedField : public MagneticField<T> {
         bool status_B = false;
         bool status_GradAbsB = false;
         const bool extrapolate;
+        const bool stellsym = false;
+        const int nfp = 1;
+        vector<bool> symmetries = vector<bool>(1, false);
 
     protected:
-        void _B_impl(Tensor2& B) override {
+        void _B_cyl_impl(Tensor2& B_cyl) override {
             if(!interp_B)
                 interp_B = std::make_shared<RegularGridInterpolant3D<Tensor2>>(rule, r_range, phi_range, z_range, 3, extrapolate);
             if(!status_B) {
-                Tensor2 old_points = this->field->get_points_cart();
                 interp_B->interpolate_batch(fbatch_B);
-                this->field->set_points_cart(old_points);
                 status_B = true;
             }
-            interp_B->evaluate_batch(this->get_points_cyl_ref(), B);
+            if(nfp > 1 || stellsym){
+                Tensor2& rphiz = this->get_points_cyl_ref();
+                Tensor2& rphiz_sym = points_cyl_sym.get_or_create({npoints, 3});
+                exploit_symmetries_points(rphiz, rphiz_sym);
+                interp_B->evaluate_batch(rphiz_sym, B_cyl);
+                apply_symmetries_to_B_cyl(B_cyl);
+            } else {
+                interp_B->evaluate_batch(this->get_points_cyl_ref(), B_cyl);
+            }
         }
 
-        void _GradAbsB_impl(Tensor2& GradAbsB) override {
+        void _GradAbsB_cyl_impl(Tensor2& GradAbsB_cyl) override {
             if(!interp_GradAbsB)
                 interp_GradAbsB = std::make_shared<RegularGridInterpolant3D<Tensor2>>(rule, r_range, phi_range, z_range, 3, extrapolate);
             if(!status_GradAbsB) {
-                Tensor2 old_points = this->field->get_points_cart();
                 interp_GradAbsB->interpolate_batch(fbatch_GradAbsB);
-                this->field->set_points_cart(old_points);
                 status_GradAbsB = true;
             }
-            interp_GradAbsB->evaluate_batch(this->get_points_cyl_ref(), GradAbsB);
+            if(nfp > 1 || stellsym){
+                Tensor2& rphiz = this->get_points_cyl_ref();
+                Tensor2& rphiz_sym = points_cyl_sym.get_or_create({npoints, 3});
+                exploit_symmetries_points(rphiz, rphiz_sym);
+                interp_GradAbsB->evaluate_batch(rphiz_sym, GradAbsB_cyl);
+                apply_symmetries_to_GradAbsB_cyl(GradAbsB_cyl);
+            } else {
+                interp_GradAbsB->evaluate_batch(this->get_points_cyl_ref(), GradAbsB_cyl);
+            }
+        }
+
+        void _B_impl(Tensor2& B) override {
+            Tensor2& B_cyl = this->B_cyl_ref();
+            Tensor2& rphiz = this->get_points_cyl_ref();
+            int npoints = B.shape(0);
+            for (int i = 0; i < npoints; ++i) {
+                double phi = rphiz(i, 1);
+                B(i, 0) = std::cos(phi)*B_cyl(i, 0) - std::sin(phi)*B_cyl(i, 1);
[…]
