### 2021-07-12 4790a99
add ability to exploit stellarator and rotational symmetry
