### 2021-07-13 64658d3
compute_bell enchancement for external dof handling and bug fixes
