### 2021-07-13 b0ce4b9
compute_bell enchancement for external dof handling and bug fixes
