### 2021-07-13 be41fe2
curve classes with graph framework without the required recompute_bell method
