+++ src/simsopt/geo/graph_curve_rzfourier.py
@@ -32,7 +32,7 @@ class CurveRZFourier(sopp.CurveRZFourier, Curve):
         elif isinstance(quadpoints, np.ndarray):
             quadpoints = list(quadpoints)
         sopp.CurveRZFourier.__init__(self, quadpoints, order, nfp, stellsym)
-        Curve.__init__(self, dof_setter=self.set_dofs, x0=self.get_dofs())
+        Curve.__init__(self, dof_setter=CurveRZFourier.set_dofs, x0=self.get_dofs())
 
     def get_dofs(self):
         """
@@ -45,5 +45,5 @@ class CurveRZFourier(sopp.CurveRZFourier, Curve):
         This function sets the dofs associated to this object.
         """
         sopp.CurveRZFourier.set_dofs(self, dofs)
-        for d in self.dependencies:
-            d.invalidate_cache()
+        # for d in self.dependencies:
+        #    d.invalidate_cache()
