### 2021-07-13 3701bd0
CMakelists updated for conda
