+++ CMakeLists.txt
@@ -18,16 +18,6 @@ message(status "Python header dirs are ${Python_INCLUDE_DIRS}")
 message(status "Python library dirs are ${Python_LIBRARY_DIRS}")
 message(status "Python version is ${Python_VERSION}")
 
-find_package(OpenMP)
-find_package(Boost)
-if(Boost_FOUND)
-    message(STATUS "Boost version" ${Boost_VERSION_STRING})
-    message(STATUS "Boost include" ${Boost_INCLUDE_DIRS})
-    add_compile_definitions(WITH_BOOST)
-else()
-    message(STATUS "Compiling without boost.")
-endif()
-
 
 find_package(Git QUIET)
 if(GIT_FOUND AND EXISTS "${PROJECT_SOURCE_DIR}/.git")
@@ -50,9 +40,22 @@ IF(DEFINED ENV{CI})
 elseif(DEFINED ENV{CONDA_BUILD})
     message(STATUS "conda build environment detected. Set compilation flags accordingly (target ivybridge which supports avx).")
     set(CMAKE_CXX_FLAGS "-O3 -march=ivybridge -mfma -ffp-contract=fast")
+    link_directories("$ENV{BUILD_PREFIX}/lib")
+    include_directories("$ENV{BUILD_PREFIX}/include")
+    message(STATUS "Prefix environment variable is $ENV{BUILD_PREFIX}")
 else()
     message(STATUS "Local build detected. Set compilation flags accordingly (march=native).")
     set(CMAKE_CXX_FLAGS "-O3 -march=native -mfma -ffp-contract=fast")
+endif()
+
+find_package(OpenMP)
+find_package(Boost)
+if(Boost_FOUND)
+    message(STATUS "Boost version" ${Boost_VERSION_STRING})
+    message(STATUS "Boost include" ${Boost_INCLUDE_DIRS})
+    add_compile_definitions(WITH_BOOST)
+else()
+    message(STATUS "Compiling without boost.")
 endif()
 
 add_subdirectory(thirdparty/pybind11)
