+++ conda.recipe/meta.yaml
@@ -11,7 +11,7 @@ build:
   include_recipe: False
   # skip: "true # [win, py<37]"
   number: 0
-  script: python -m pip install . -vvv
+  script: {{ python }} -m pip install . -vvv
 
 requirements:
   build:
@@ -30,7 +30,7 @@ requirements:
 
   run:
     - python
-    - numpy {{ numpy }}
+    - numpy {{ pin_compatible('numpy') }}
     - boost {{ boost }}
     - jax >=0.2.4
     - jaxlib >=0.1.56
