### 2021-07-13 ce389ee
Few more changes to conda recipe
