### 2021-07-13 53d7866
reduce memory usage further
