### 2021-07-13 eb1ff0e
reduce memory usage further
