+++ .github/workflows/extensive_ci.yml
@@ -31,17 +31,18 @@ jobs:
         test-type: [unit, integrated]
         spec: [true, false]
         vmec: [true, false]
+        boost: [true]
         include:
             - python-version: 3.9.5
               test-type: unit
               spec: false
               vmec: false
-              boost: true
+              boost: false
             - python-version: 3.9.5
               test-type: integrated
               spec: false
               vmec: false
-              boost: true
+              boost: false
 
     steps:
     # First print out lots of information. We do this in separate 
