+++ src/simsopt/geo/surface.py
@@ -19,7 +19,7 @@ class Surface(Optimizable):
 
     def plot(self, ax=None, show=True, plot_normal=False, plot_derivative=False, scalars=None, wireframe=True):
         """
-        Plot the surface using mayavi. 
+        Plot the surface using mayavi.
         Note: the `ax` and `show` parameter can be used to plot more than one surface:
 
         .. code-block::
@@ -95,7 +95,7 @@ class Surface(Optimizable):
         if phi < -np.pi:
             phi = phi + 2. * np.pi
 
-        # varphi are the search intervals on which we look for the cross section in 
+        # varphi are the search intervals on which we look for the cross section in
         # at constant cylindrical phi
         # The cross section is sampled at a number of points (theta_resolution) poloidally.
         varphi = np.asarray([0., 0.5, 1.0])
@@ -143,7 +143,7 @@ class Surface(Optimizable):
         # if idx_right == 0, then the subinterval must be idx_left = 0 and idx_right = 1
         idx_right = np.argmax(phi <= cyl_phi, axis=0)
         idx_right = np.where(idx_right == 0, 1, idx_right)
-        idx_left = idx_right-1 
+        idx_left = idx_right-1
 
         varphi_left = varphigrid[idx_left, np.arange(idx_left.size)]
         varphi_right = varphigrid[idx_right, np.arange(idx_right.size)]
@@ -156,7 +156,7 @@ class Surface(Optimizable):
             gamma = np.zeros((varphi_in.size, 3))
             self.gamma_lin(gamma, varphi_in, theta)
             phi = np.arctan2(gamma[:, 1], gamma[:, 0])
-            pinc = (phi < left_bound).astype(int) 
+            pinc = (phi < left_bound).astype(int)
             minc = (phi > right_bound).astype(int)
             phi = phi + 2.*np.pi * (pinc - minc)
             return phi
@@ -176,11 +176,11 @@ class Surface(Optimizable):
                 c = np.where(flag, b, c)
                 err = np.max(np.abs(a-c))
             b = (a + c)/2.
-            return b          
+            return b
         # bisect cyl_phi to compute the cross section
         sol = bisection(cyl_phi_left, varphi_left, cyl_phi_right, varphi_right)
         cross_section = np.zeros((sol.size, 3))
-        self.gamma_lin(cross_section, sol, theta) 
+        self.gamma_lin(cross_section, sol, theta)
         return cross_section
 
     def aspect_ratio(self):
@@ -193,11 +193,11 @@ class Surface(Optimizable):
         .. math::
             AR = R_{\text{major}} / R_{\text{minor}}
 
-        where 
+        where
 
         .. math::
             R_{\text{minor}} &= \sqrt{ \overline{A} / \pi } \\
-            R_{\text{major}} &= \frac{V}{2 \pi^2  R_{\text{minor}}^2} 
+            R_{\text{major}} &= \frac{V}{2 \pi^2  R_{\text{minor}}^2}
 
         and :math:`V` is the volume enclosed by the surface, and :math:`\overline{A}` is the
         average cross sectional area.
@@ -208,18 +208,18 @@ class Surface(Optimizable):
             \overline{A} = \frac{1}{2\pi} \int_{S_{\phi}} ~dS ~d\phi
 
         where :math:`S_\phi` is the cross section of the surface at the cylindrical angle :math:`\phi`.
[…]
