+++ src/simsopt/mhd/vmec.py
@@ -19,7 +19,7 @@ logger = logging.getLogger(__name__)
 try:
     from mpi4py import MPI
 except ImportError as e:
-    MPI = None 
+    MPI = None
     logger.warning(str(e))
 
 # vmec_found = True
@@ -39,6 +39,7 @@ if MPI is not None:
 else:
     MpiPartition = None
 
+print('in vmec')
 
 # Flags used by runvmec():
 restart_flag = 1
@@ -53,7 +54,7 @@ reset_jacdt_flag = 32
 #value flag-name         calls routines to...
 #----- ---------         ---------------------
 #  1   restart_flag      reset internal run-control parameters
-#                        (for example, if jacobian was bad, to try a smaller 
+#                        (for example, if jacobian was bad, to try a smaller
 #                        time-step)
 #  2   readin_flag       read in data from input_file and initialize parameters
 #                        or arrays which do not dependent on radial grid size
@@ -142,13 +143,13 @@ class Vmec(Optimizable):
     ``files_to_delete`` attribute to ``[]`` after that run of VMEC.
 
     Args:
-        filename: Name of a VMEC input file to use for loading the 
+        filename: Name of a VMEC input file to use for loading the
           initial parameters. If ``None``, default parameters will be used.
-        mpi: A :obj:`simsopt.util.mpi.MpiPartition` instance, from which 
+        mpi: A :obj:`simsopt.util.mpi.MpiPartition` instance, from which
           the worker groups will be used for VMEC calculations. If ``None``,
           each MPI process will run VMEC independently.
         keep_all_files: If ``False``, all ``wout`` output files will be deleted
-          except for the first and most recent ones from worker group 0. If 
+          except for the first and most recent ones from worker group 0. If
           ``True``, all ``wout`` files will be kept.
 
     Attributes:
@@ -589,3 +590,139 @@ class Vmec(Optimizable):
 
         well = (dVds_s0 - dVds_s1) / dVds_s0
         return well
+
+    def iota_weighted(self,weight_function):
+        """
+        Computed a weighted average of the rotational transform defined by
+        the prescribed weight_function, f.
+
+        \int ds \, iota * weight_function(s) / \int ds \, weight_function(s)
+
+        Args:
+            weight_function : function handle which takes a single argument, s,
+                the normalized toroidal flux
+        """
+        self.run()
+
+        return np.sum(weight_function(self.s_half_grid) * self.wout.iotas[1:]) \
+            / np.sum(weight_function(self.s_half_grid))
+
+    def iota_target(self,iota_target):
+        """
+        Computes a metric quantifying the deviation of iota from a prescribed
+        target value.
+
+        0.5 * \int ds \, (iota - iota_target)**2
+
+        Args:
+            iota_target : function handle which takes a single argument, s,
+                the normalized toroidal flux, and returns the target rotational
+                transform.
+        """
+        self.run()
+
+        return 0.5 * np.sum((self.wout.iotas[1::]
+            - iota_target(self.s_half_grid))**2) * self.ds
+
+    def d_iota_target(self,iota_target,delta):
+        if self.indata.ncurr != 1:
+            raise RuntimeError('''d_iota_target cannot be computed without
+                running vmec with ncurr = 1''')
+
+        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0,1::] / (4*np.pi*1e-7)
+        ac_aux_f_prev = self.indata.ac_aux_f
[…]
