### 2021-07-14 7a7f8d6
Bug fixes in graph optimizable framework after adding support for external dofs
