### 2021-07-14 119eb5a
Bug fixes in graph optimizable framework after adding support for external dofs
