### 2021-07-14 fb24ab4
Bug fixes in curves using graph optimizable framework
