### 2021-07-14 26203ec
Bug fixes in curves using graph optimizable framework
