+++ tests/geo/test_graph_curve.py
@@ -1,0 +1,399 @@
+import numpy as np
+import unittest
+
+from simsopt.geo.graph_curve_xyzfourier import CurveXYZFourier, JaxCurveXYZFourier
+from simsopt.geo.graph_curve_rzfourier import CurveRZFourier
+from simsopt.geo.graph_curve_helical import CurveHelical
+from simsopt.geo.graph_curve import RotatedCurve, curves_to_vtk
+from simsopt.geo import parameters
+
+try:
+    import pyevtk
+    pyevtk_found = True
+except ImportError:
+    pyevtk_found = False
+
+parameters['jit'] = False
+
+
+def taylor_test(f, df, x, epsilons=None, direction=None):
+    np.random.seed(1)
+    if direction is None:
+        direction = np.random.rand(*(x.shape))-0.5
+    dfx = df(x)@direction
+    if epsilons is None:
+        epsilons = np.power(2., -np.asarray(range(7, 20)))
+    # print("################################################################################")
+    err_old = 1e9
+    counter = 0
+    for eps in epsilons:
+        if counter > 8:
+            break
+
+        fpluseps = f(x + eps * direction)
+        fminuseps = f(x - eps * direction)
+        dfest = (fpluseps-fminuseps)/(2*eps)
+        err = np.linalg.norm(dfest - dfx)
+        # print(err, err_old)
+        assert err < 1e-9 or err < 0.3 * err_old
+        if err < 1e-9:
+            break
+        err_old = err
+        counter += 1
+    if err > 1e-10:
+        assert counter > 3
+    # print("################################################################################")
+
+
+def get_curve(curvetype, rotated, x=np.asarray([0.5])):
+    np.random.seed(2)
+    rand_scale = 0.01
+    order = 4
+
+    if curvetype == "CurveXYZFourier":
+        curve = CurveXYZFourier(x, order)
+    elif curvetype == "JaxCurveXYZFourier":
+        curve = JaxCurveXYZFourier(x, order)
+    elif curvetype == "CurveRZFourier":
+        curve = CurveRZFourier(x, order, 2, True)
+    elif curvetype == "CurveHelical":
+        curve = CurveHelical(x, order, 5, 2, 1.0, 0.3)
+    else:
+        assert False
+    # dofs = np.zeros((curve.num_dofs(), ))
+    dofs = np.zeros((curve.dof_size, ))
+    if curvetype in ["CurveXYZFourier", "JaxCurveXYZFourier"]:
+        dofs[1] = 1.
+        dofs[2*order + 3] = 1.
+        dofs[4*order + 3] = 1.
+    elif curvetype in ["CurveRZFourier"]:
+        dofs[0] = 1.
+        dofs[1] = 0.1
+        dofs[order+1] = 0.1
+    elif curvetype in ["CurveHelical"]:
+        dofs[0] = np.pi/2
+    else:
+        assert False
+    curve.x = dofs
+
+    # dofs = np.asarray(curve.get_dofs())
+    # curve.set_dofs(dofs + rand_scale * np.random.rand(len(dofs)).reshape(dofs.shape))
+    dofs = curve.x
+    curve.x = dofs + rand_scale * np.random.rand(len(dofs)).reshape(dofs.shape)
+    if rotated:
+        curve = RotatedCurve(curve, 0.5, flip=False)
+    return curve
+
+
+class Testing(unittest.TestCase):
+
+    curvetypes = ["CurveXYZFourier", "JaxCurveXYZFourier", "CurveRZFourier", "CurveHelical"]
+
+    def test_curve_helical_xyzfourier(self):
[…]
