### 2021-07-14 07d41f0
plot_poincare_data revert back to s=0.2
