### 2021-07-14 ee11cbb
commented out code removed in graph curve objectives
