### 2021-07-14 dcd585a
curve objective tests updated to graph optimizable framework
