+++ tests/geo/test_graph_curve_objectives.py
@@ -1,0 +1,160 @@
+import unittest
+
+import numpy as np
+
+from simsopt.geo import parameters
+from simsopt.geo.graph_curve import RotatedCurve
+from simsopt.geo.graph_curve_xyzfourier import CurveXYZFourier, JaxCurveXYZFourier
+from simsopt.geo.graph_curve_rzfourier import CurveRZFourier
+from simsopt.geo.graph_curve_objectives import CurveLength, LpCurveCurvature, \
+    LpCurveTorsion, MinimumDistance
+
+parameters['jit'] = False
+
+
+class Testing(unittest.TestCase):
+
+    curvetypes = ["CurveXYZFourier", "JaxCurveXYZFourier", "CurveRZFourier"]
+
+    def create_curve(self, curvetype, rotated):
+        np.random.seed(1)
+        rand_scale = 0.01
+        order = 4
+        nquadpoints = 200
+
+        if curvetype == "CurveXYZFourier":
+            coil = CurveXYZFourier(nquadpoints, order)
+        elif curvetype == "JaxCurveXYZFourier":
+            coil = JaxCurveXYZFourier(nquadpoints, order)
+        elif curvetype == "CurveRZFourier":
+            coil = CurveRZFourier(nquadpoints, order, 2, False)
+        else:
+            # print('Could not find' + curvetype)
+            assert False
+        dofs = np.zeros((coil.dof_size, ))
+        if curvetype in ["CurveXYZFourier", "JaxCurveXYZFourier"]:
+            dofs[1] = 1.
+            dofs[2*order+3] = 1.
+            dofs[4*order+3] = 1.
+        elif curvetype in ["CurveRZFourier"]:
+            dofs[0] = 1.
+            dofs[1] = 0.1
+            dofs[order+1] = 0.1
+        else:
+            assert False
+
+        coil.x = dofs + rand_scale * np.random.rand(len(dofs)).reshape(dofs.shape)
+        if rotated:
+            coil = RotatedCurve(coil, 0.5, flip=False)
+        return coil
+
+    def subtest_curve_length_taylor_test(self, curve):
+        J = CurveLength(curve)
+        J0 = J.J()
+        curve_dofs = curve.x
+        h = 1e-3 * np.random.rand(len(curve_dofs)).reshape(curve_dofs.shape)
+        dJ = J.dJ()
+        deriv = np.sum(dJ * h)
+        err = 1e6
+        for i in range(5, 15):
+            eps = 0.5**i
+            curve.x = curve_dofs + eps * h
+            Jh = J.J()
+            deriv_est = (Jh-J0)/eps
+            err_new = np.linalg.norm(deriv_est-deriv)
+            # print("err_new %s" % (err_new))
+            assert err_new < 0.55 * err
+            err = err_new
+
+    def test_curve_length_taylor_test(self):
+        for curvetype in self.curvetypes:
+            for rotated in [True, False]:
+                with self.subTest(curvetype=curvetype, rotated=rotated):
+                    curve = self.create_curve(curvetype, rotated)
+                    self.subtest_curve_length_taylor_test(curve)
+
+    def subtest_curve_curvature_taylor_test(self, curve):
+        J = LpCurveCurvature(curve, p=2)
+        J0 = J.J()
+        curve_dofs = curve.x
+        h = 1e-2 * np.random.rand(len(curve_dofs)).reshape(curve_dofs.shape)
+        dJ = J.dJ()
+        deriv = np.sum(dJ * h)
+        assert np.abs(deriv) > 1e-10
+        err = 1e6
+        for i in range(5, 15):
+            eps = 0.5**i
+            curve.x = curve_dofs + eps * h
+            Jh = J.J()
[…]
