+++ setup.cfg
@@ -11,7 +11,7 @@ author_email = mattland@umd.edu, wechsung@nyu.edu, mbkumar@gmail.com
 maintainer = "Bharat Medasani"
 maintainer_email = mbkumar@gmail.com
 url = https://github.com/hiddenSymmetries/simsopt
-download_url = https://test.pypi.org/project/simsopt
+download_url = https://pypi.org/project/simsopt
 project_urls = 
     Bug Tracker = https://github.com/hiddenSymmetries/simsopt/issues
     Documentation = https://simsopt.readthedocs.io
