### 2021-07-14 0e8e129
Add force flag for rebuilding documention command
