### 2021-07-14 004acf7
Update to documentation. Added minimize_qfm.
