+++ src/simsopt/geo/qfmsurface.py
@@ -4,24 +4,25 @@ from simsopt.geo.surfaceobjectives import QfmResidual
 
 
 class QfmSurface():
-    """
+    r"""
     QfmSurface is used to compute a quadratic-flux minimizing surface, defined
     as the minimum of the objective function,
 
-        f = \int d^2 x (B \cdot n)^2/ \int d^2 x \, B^2
+    .. math::
+        f(S) = \frac{\int_{S} d^2 x \, \left(\textbf{B} \cdot \hat{\textbf{n}}\right)^2}{\int_{S} d^2 x \, B^2}
 
     subject to a constraint on the surface label, such as the area, volume,
-    or toroidal flux. This objective, f, is computed in QfmResidual
-    (see surfaceobjectives.py). If magnetic surfaces exists, f will be zero.
-    If not, QFM surfaces approximate flux surfaces.
+    or toroidal flux. This objective, :math:`f`, is computed in :class:`QfmResidual`
+    (see :mod:`surfaceobjectives.py`). If magnetic surfaces exists, :math:`f`
+    will be zero. If not, QFM surfaces approximate flux surfaces.
 
     The label constraint can be enforced with a penalty formulation, defined
-    in qfm_penalty_constraints, and whose minimium is computed with LBFGS-B
-    by minimize_qfm_penalty_constraints_LBFGS().
+    in :func:`qfm_penalty_constraints()`, and whose minimium is computed with
+    LBFGS-B by :func:`minimize_qfm_penalty_constraints_LBFGS()'.
 
     Alternatively, the label constraint can be enforced with a constrained
     optimization algorithm. This constrained optimization problem is
-    solved with the SLSQP algorithm by minimize_qfm_exact_constraints_SLSQP().
+    solved with the SLSQP algorithm by :func:`minimize_qfm_exact_constraints_SLSQP()`.
     """
 
     def __init__(self, biotsavart, surface, label, targetlabel):
@@ -31,10 +32,11 @@ class QfmSurface():
         self.targetlabel = targetlabel
 
     def qfm_label_constraint(self, x, derivatives=0):
-        """
+        r"""
         Returns the residual
 
-        0.5 * (label - labeltarget)^2
+        .. math::
+            0.5 (\texttt{label} - \texttt{labeltarget})^2
 
         and optionally the gradient wrt surface params.
         """
@@ -58,13 +60,14 @@ class QfmSurface():
             return val
 
     def qfm_objective(self, x, derivatives=0):
-        """
+        r"""
         Returns the residual
 
-        f(x)
+        .. math::
+            f(S)
 
-        and optionally the gradient wrt surface params where f(x) is the QFM
-        residual defined in surfaceobjectives.py.
+        and optionally the gradient wrt surface params where :math:`f(S)`` is
+        the QFM residual defined in surfaceobjectives.py.
         """
         assert derivatives in [0, 1]
 
@@ -85,13 +88,14 @@ class QfmSurface():
             return r, dr
 
     def qfm_penalty_constraints(self, x, derivatives=0, constraint_weight=1.):
-        """
+        r"""
         Returns the residual
 
-        f(x) + 0.5 * constraint_weight * (label - labeltarget)^2
+        .. math::
+            f(S) + 0.5 \texttt{constraint_weight} (\texttt{label} - \texttt{labeltarget})^2
 
-        and optionally the gradient and Hessian wrt surface params where f(x)
-        is the QFM residual defined in surfaceobjectives.py.
+        and optionally the gradient and Hessian wrt surface params where
+        :math:`f(S)` is the QFM residual defined in :mod:`surfaceobjectives.py`.
         """
 
[…]
