+++ docs/source/simsopt.solve.rst
@@ -11,7 +11,6 @@ simsopt.solve.graph\_mpi module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.solve.graph\_serial module
 ----------------------------------
@@ -20,7 +19,6 @@ simsopt.solve.graph\_serial module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.solve.mpi module
 ------------------------
@@ -29,7 +27,6 @@ simsopt.solve.mpi module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.solve.serial module
 ---------------------------
@@ -38,7 +35,6 @@ simsopt.solve.serial module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 Module contents
 ---------------
@@ -47,4 +43,3 @@ Module contents
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
