+++ docs/source/simsopt.geo.rst
@@ -11,7 +11,6 @@ simsopt.geo.biotsavart module
    :members:
    :undoc-members:
    :show-inheritance:
-   :inherited-members:
 
 simsopt.geo.boozersurface module
 --------------------------------
@@ -20,7 +19,6 @@ simsopt.geo.boozersurface module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.coilcollection module
 ---------------------------------
@@ -29,7 +27,6 @@ simsopt.geo.coilcollection module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.config module
 -------------------------
@@ -38,7 +35,6 @@ simsopt.geo.config module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.curve module
 ------------------------
@@ -47,7 +43,6 @@ simsopt.geo.curve module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.curvehelical module
 -------------------------------
@@ -56,7 +51,6 @@ simsopt.geo.curvehelical module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.curveobjectives module
 ----------------------------------
@@ -65,7 +59,6 @@ simsopt.geo.curveobjectives module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.curverzfourier module
 ---------------------------------
@@ -74,7 +67,6 @@ simsopt.geo.curverzfourier module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.curvexyzfourier module
 ----------------------------------
@@ -83,7 +75,6 @@ simsopt.geo.curvexyzfourier module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.graph\_surface module
 ---------------------------------
@@ -92,7 +83,6 @@ simsopt.geo.graph\_surface module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.jit module
 ----------------------
@@ -101,7 +91,6 @@ simsopt.geo.jit module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.magneticfield module
 --------------------------------
@@ -110,7 +99,6 @@ simsopt.geo.magneticfield module
    :members:
    :undoc-members:
    :show-inheritance:
-   :inherited-members:
 
 simsopt.geo.magneticfieldclasses module
 ---------------------------------------
@@ -119,7 +107,14 @@ simsopt.geo.magneticfieldclasses module
    :members:
    :undoc-members:
    :show-inheritance:
-   :inherited-members:
+
+simsopt.geo.qfmsurface module
+-----------------------------
+
+.. automodule:: simsopt.geo.qfmsurface
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 simsopt.geo.surface module
 --------------------------
@@ -128,7 +123,6 @@ simsopt.geo.surface module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.surfacegarabedian module
 ------------------------------------
@@ -137,16 +131,14 @@ simsopt.geo.surfacegarabedian module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.geo.surfacehenneberg module
-------------------------------------
+-----------------------------------
 
[…]
