+++ docs/source/simsopt.mhd.rst
@@ -11,7 +11,6 @@ simsopt.mhd.boozer module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.mhd.spec module
 -----------------------
@@ -20,7 +19,6 @@ simsopt.mhd.spec module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 simsopt.mhd.vmec module
 -----------------------
@@ -29,7 +27,6 @@ simsopt.mhd.vmec module
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
 
 Module contents
 ---------------
@@ -38,4 +35,3 @@ Module contents
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:
