+++ docs/source/simsopt.rst
@@ -7,7 +7,6 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
-   simsopt._core
    simsopt.geo
    simsopt.mhd
    simsopt.objectives
@@ -21,4 +20,3 @@ Module contents
    :members:
    :undoc-members:
    :show-inheritance:
-   :private-members:

[…]
