### 2021-07-14 1825b14
Update to documentation. Added minimize_qfm.
