### 2021-07-14 4dd0494
Minor cleanup in graph least squares
