+++ src/simsopt/objectives/graph_least_squares.py
@@ -17,7 +17,7 @@ from numbers import Real
 
 import numpy as np
 
-from .._core.graph_optimizable import DOFs, Optimizable
+from .._core.graph_optimizable import Optimizable
 from ..util.types import RealArray, IntArray, BoolArray
 
 
