### 2021-07-14 474088c
make biot savart deterministic even when using openmp
