### 2021-07-14 2f29e30
Added private members to docs.
