+++ docs/source/simsopt.objectives.rst
@@ -11,6 +11,7 @@ simsopt.objectives.functions module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.objectives.graph\_functions module
 ------------------------------------------
@@ -19,6 +20,7 @@ simsopt.objectives.graph\_functions module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.objectives.graph\_least\_squares module
 -----------------------------------------------
@@ -27,6 +29,7 @@ simsopt.objectives.graph\_least\_squares module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.objectives.least\_squares module
 ----------------------------------------
@@ -35,6 +38,7 @@ simsopt.objectives.least\_squares module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 Module contents
 ---------------
@@ -43,3 +47,4 @@ Module contents
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
