### 2021-07-14 9360430
Added private members to docs.
