+++ docs/source/simsopt.util.rst
@@ -11,6 +11,7 @@ simsopt.util.dev module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.util.logging module
 ---------------------------
@@ -19,6 +20,7 @@ simsopt.util.logging module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.util.mpi module
 -----------------------
@@ -27,6 +29,7 @@ simsopt.util.mpi module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.util.mpi\_logger module
 -------------------------------
@@ -35,6 +38,7 @@ simsopt.util.mpi\_logger module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.util.types module
 -------------------------
@@ -43,6 +47,7 @@ simsopt.util.types module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 Module contents
 ---------------
@@ -51,3 +56,4 @@ Module contents
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
