+++ docs/source/simsopt.geo.rst
@@ -11,6 +11,7 @@ simsopt.geo.biotsavart module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.boozersurface module
 --------------------------------
@@ -19,6 +20,7 @@ simsopt.geo.boozersurface module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.coilcollection module
 ---------------------------------
@@ -27,6 +29,7 @@ simsopt.geo.coilcollection module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.config module
 -------------------------
@@ -35,6 +38,7 @@ simsopt.geo.config module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.curve module
 ------------------------
@@ -43,6 +47,7 @@ simsopt.geo.curve module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.curvehelical module
 -------------------------------
@@ -51,6 +56,7 @@ simsopt.geo.curvehelical module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.curveobjectives module
 ----------------------------------
@@ -59,6 +65,7 @@ simsopt.geo.curveobjectives module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.curverzfourier module
 ---------------------------------
@@ -67,6 +74,7 @@ simsopt.geo.curverzfourier module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.curvexyzfourier module
 ----------------------------------
@@ -75,6 +83,7 @@ simsopt.geo.curvexyzfourier module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.graph\_surface module
 ---------------------------------
@@ -83,6 +92,7 @@ simsopt.geo.graph\_surface module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.jit module
 ----------------------
@@ -91,6 +101,7 @@ simsopt.geo.jit module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.magneticfield module
 --------------------------------
@@ -99,6 +110,7 @@ simsopt.geo.magneticfield module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.magneticfieldclasses module
 ---------------------------------------
@@ -107,6 +119,7 @@ simsopt.geo.magneticfieldclasses module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.qfmsurface module
 -----------------------------
@@ -115,6 +128,7 @@ simsopt.geo.qfmsurface module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.surface module
 --------------------------
@@ -123,6 +137,7 @@ simsopt.geo.surface module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.surfacegarabedian module
 ------------------------------------
@@ -131,6 +146,7 @@ simsopt.geo.surfacegarabedian module
    :members:
    :undoc-members:
    :show-inheritance:
+   :private-members:
 
 simsopt.geo.surfacehenneberg module
 -----------------------------------
[…]
