### 2021-07-14 c731617
Obscure numpy related bug is fixed in graph based least squares
