### 2021-07-14 41c7806
Optimization test for graph based curve
