### 2021-07-14 df87b27
added test for minimize_qfm.
