### 2021-07-14 890a953
added test for minimize_qfm.
