### 2021-07-15 bc7c375
curve dof numbering test removed
