+++ tests/geo/test_magneticfields.py
@@ -1,10 +1,10 @@
 from simsopt.geo.magneticfieldclasses import ToroidalField, \
     ScalarPotentialRZMagneticField, CircularCoil, Dommaschk, \
     Reiman, sympy_found, InterpolatedField
-from simsopt.geo.curvexyzfourier import CurveXYZFourier
+from simsopt.geo.curve_xyzfourier import CurveXYZFourier
 from simsopt.geo.magneticfield import MagneticFieldSum
-from simsopt.geo.curverzfourier import CurveRZFourier
-from simsopt.geo.curvehelical import CurveHelical
+from simsopt.geo.curve_rzfourier import CurveRZFourier
+from simsopt.geo.curve_helical import CurveHelical
 from simsopt.geo.biotsavart import BiotSavart
 from simsopt.geo.coilcollection import CoilCollection
 from .surface_test_helpers import get_ncsx_data
