### 2021-07-15 9253538
print simsopt_inc_dirs
