### 2021-07-15 c47d904
pinned numpy removed from meta.yaml
