### 2021-07-15 11a7dde
Trying alternate approach for conda_build_config
