+++ conda.recipe/conda_build_config.yaml
@@ -1,9 +1,7 @@
 numpy:
-  - 1.19
-  - 1.20
+  - >=1.19
 boost:
-  - 1.75
-  - 1.76
+  - >=1.76
 
 pin_run_as_build:
   boost: x.x
