### 2021-07-15 91cc83c
Removed interpolation; more particles; fixed 2*pi factor in angular momentum definition
