### 2021-07-15 19efbdc
introduce general derivative class and demonstrate how to use it for a FOCUS type objective
