### 2021-07-15 de2a75d
introduce general derivative class and demonstrate how to use it for a FOCUS type objective
