+++ src/simsopt/geo/biotsavart.py
@@ -2,9 +2,50 @@ import numpy as np
 import simsoptpp as sopp
 from simsopt.geo.magneticfield import MagneticField
 from simsopt.geo.curve import Curve
+from simsopt._core.graph_optimizable import Optimizable
+from simsopt._core.derivative import Derivative
 
 
-class BiotSavart(sopp.BiotSavart, MagneticField):
+class Coil(sopp.Coil, Optimizable):
+
+    def __init__(self, curve, current):
+        self.__curve = curve
+        self.__current = current
+        sopp.Coil.__init__(self, curve, current)
+        Optimizable.__init__(self, x0=np.asarray([]), opts_in=[curve, current])
+
+    def vjp_graph(self, v_gamma, v_gammadash, v_current):
+        return self.curve.dgamma_by_dcoeff_vjp_graph(v_gamma) \
+            + self.curve.dgammadash_by_dcoeff_vjp_graph(v_gammadash) \
+            + self.current.vjp_graph(v_current)
+
+
+class Current(sopp.Current, Optimizable):
+
+    def __init__(self, current):
+        sopp.Current.__init__(self, current)
+        Optimizable.__init__(self, external_dof_setter=Current.set_dofs,
+                       x0=self.get_dofs())
+
+    def set_dofs(self, dofs):
+        sopp.Current.set_dofs(self, dofs)
+
+    def vjp_graph(self, v_current):
+        return Derivative({self: v_current})
+
+
+class FlippedCurrent(sopp.FlippedCurrent, Optimizable):
+
+    def __init__(self, basecurrent):
+        self.__basecurrent = basecurrent
+        sopp.FlippedCurrent.__init__(self, basecurrent)
+        Optimizable.__init__(self, x0=np.asarray([]), opts_in=[basecurrent])
+
+    def vjp_graph(self, v_current):
+        return self.__basecurrent.vjp_graph(-v_current)
+
+
+class BiotSavart(sopp.BiotSavart, MagneticField, Optimizable):
     r"""
     Computes the MagneticField induced by a list of closed curves :math:`\Gamma_k` with electric currents :math:`I_k`.
     The field is given by
@@ -17,16 +58,23 @@ class BiotSavart(sopp.BiotSavart, MagneticField):
 
     """
 
-    def __init__(self, coils, coil_currents):
-        assert len(coils) == len(coil_currents)
-        assert all(isinstance(item, Curve) for item in coils)
-        assert all(isinstance(item, float) for item in coil_currents)
-        self.currents_optim = [sopp.Current(c) for c in coil_currents]
-        self.coils_optim = [sopp.Coil(curv, curr) for curv, curr in zip(coils, self.currents_optim)]
-        self.coils = coils
-        self.coil_currents = coil_currents
+    # def __init__(self, coils, coil_currents):
+    #     assert len(coils) == len(coil_currents)
+    #     assert all(isinstance(item, Curve) for item in coils)
+    #     assert all(isinstance(item, float) for item in coil_currents)
+    #     self.currents_optim = [sopp.Current(c) for c in coil_currents]
+    #     self.coils_optim = [sopp.Coil(curv, curr) for curv, curr in zip(coils, self.currents_optim)]
+    #     self.coils = coils
+    #     self.coil_currents = coil_currents
+    #     MagneticField.__init__(self)
+    #     sopp.BiotSavart.__init__(self, self.coils_optim)
+
+    def __init__(self, coils):
+        self.__coils = coils
+        sopp.BiotSavart.__init__(self, coils)
[…]
