### 2021-07-15 46ad73a
Changes to conda recipe to get it compile
