+++ conda.recipe/conda_build_config.yaml
@@ -1,7 +1,8 @@
-numpy:
-  - >=1.19
 boost:
-  - >=1.76
+  - 1.76
+numpy:
+  - 1.20
 
 pin_run_as_build:
   boost: x.x
+  numpy: x.x
