### 2021-07-15 51e9409
Changes to conda recipe to get it compile
