### 2021-07-15 acc7820
underscore removed from curve filenames
