### 2021-07-15 747040d
make sure that when the C++ code changes the dofs (e.g. when doing a least squares fit), these new dofs are also copied into the graph optimizable data structure
