### 2021-07-15 522b60d
pip install changed to setup.py install for conda recipe
