### 2021-07-15 f1f31c1
Modified cmakelists for conda build
