+++ CMakeLists.txt
@@ -39,10 +39,10 @@ IF(DEFINED ENV{CI})
     set(CMAKE_CXX_FLAGS "-O3 -march=ivybridge -mfma -ffp-contract=fast")
 elseif(DEFINED ENV{CONDA_BUILD})
     message(STATUS "conda build environment detected. Set compilation flags accordingly (target ivybridge which supports avx).")
-    set(CMAKE_CXX_FLAGS "-O3 -march=ivybridge -mfma -ffp-contract=fast")
-    link_directories("$ENV{BUILD_PREFIX}/lib")
-    include_directories("$ENV{BUILD_PREFIX}/include")
-    message(STATUS "Prefix environment variable is $ENV{BUILD_PREFIX}")
+    # set(CMAKE_CXX_FLAGS "-O3 -march=ivybridge -mfma -ffp-contract=fast")
+    # link_directories("$ENV{PREFIX}/lib")
+    # include_directories("$ENV{PREFIX}/include")
+    message(STATUS "Prefix environment variable is $ENV{PREFIX}")
 else()
     message(STATUS "Local build detected. Set compilation flags accordingly (march=native).")
     set(CMAKE_CXX_FLAGS "-O3 -march=native -mfma -ffp-contract=fast")
@@ -82,7 +82,7 @@ target_include_directories(${PROJECT_NAME} PRIVATE "thirdparty/xtensor/include" 
 target_link_libraries(${PROJECT_NAME} PRIVATE fmt::fmt)
 if(Boost_FOUND)
     target_include_directories(${PROJECT_NAME} PRIVATE ${Boost_INCLUDE_DIRS})
-    target_link_libraries(${PROJECT_NAME} PRIVATE Boost::headers)
+    target_link_libraries(${PROJECT_NAME} PRIVATE pybind11::module Boost::headers)
 endif()
 
 add_executable(profiling EXCLUDE_FROM_ALL src/profiling/profiling.cpp src/simsoptpp/biot_savart_c.cpp src/simsoptpp/biot_savart_vjp_c.cpp src/simsoptpp/regular_grid_interpolant_3d_c.cpp)
