+++ tests/geo/test_biotsavart.py
@@ -3,10 +3,10 @@ import unittest
 import numpy as np
 
 from simsopt.geo.curvexyzfourier import CurveXYZFourier
-from simsopt.geo.biotsavart import BiotSavart
+from simsopt.geo.biotsavart import BiotSavart, Coil, Current
 
 
-def get_coil(num_quadrature_points=200, perturb=False):
+def get_curve(num_quadrature_points=200, perturb=False):
     coil = CurveXYZFourier(num_quadrature_points, 3)
     coeffs = coil.dofs
     coeffs[1][0] = 1.
@@ -22,64 +22,64 @@ def get_coil(num_quadrature_points=200, perturb=False):
 class Testing(unittest.TestCase):
 
     def test_biotsavart_both_interfaces_give_same_result(self):
-        coils = [get_coil()]
-        currents = [1e4]
+        curve = get_curve()
+        coil = Coil(curve, Current(1e4))
         points = np.asarray(10 * [[-1.41513202e-03, 8.99999382e-01, -3.14473221e-04]])
-        B1 = BiotSavart(coils, currents).set_points(points).B()
+        B1 = BiotSavart([coil]).set_points(points).B()
         from simsoptpp import biot_savart_B
-        B2 = biot_savart_B(points, [c.gamma() for c in coils], [c.gammadash() for c in coils], currents)
+        B2 = biot_savart_B(points, [curve.gamma()], [curve.gammadash()], [1e4])
         assert np.linalg.norm(B1) > 1e-5
         assert np.allclose(B1, B2)
 
     def test_biotsavart_exponential_convergence(self):
-        coil = get_coil()
+        coil = BiotSavart([Coil(get_curve(), Current(1e4))])
         from time import time
         # points = np.asarray(17 * [[-1.41513202e-03,  8.99999382e-01, -3.14473221e-04 ]])
         points = np.asarray(10 * [[-1.41513202e-03, 8.99999382e-01, -3.14473221e-04]])
         tic = time()
-        btrue = BiotSavart([get_coil(1000)], [1e4]).set_points(points).B()
+        btrue = BiotSavart([Coil(get_curve(1000), Current(1e4))]).set_points(points).B()
         # print(btrue)
-        bcoarse = BiotSavart([get_coil(10)], [1e4]).set_points(points).B()
-        bfine = BiotSavart([get_coil(20)], [1e4]).set_points(points).B()
+        bcoarse = BiotSavart([Coil(get_curve(10), Current(1e4))]).set_points(points).B()
+        bfine = BiotSavart([Coil(get_curve(20), Current(1e4))]).set_points(points).B()
         assert np.linalg.norm(btrue-bfine) < 1e-4 * np.linalg.norm(bcoarse-bfine)
         # print(time()-tic)
 
         tic = time()
-        dbtrue = BiotSavart([get_coil(1000)], [1e4]).set_points(points).dB_by_dX()
+        dbtrue = BiotSavart([Coil(get_curve(1000), Current(1e4))]).set_points(points).dB_by_dX()
         # print(dbtrue)
-        dbcoarse = BiotSavart([get_coil(10)], [1e4]).set_points(points).dB_by_dX()
-        dbfine = BiotSavart([get_coil(20)], [1e4]).set_points(points).dB_by_dX()
+        dbcoarse = BiotSavart([Coil(get_curve(10), Current(1e4))]).set_points(points).dB_by_dX()
+        dbfine = BiotSavart([Coil(get_curve(20), Current(1e4))]).set_points(points).dB_by_dX()
[…]

[…]
