### 2021-07-16 722917c
get more tests working
