+++ src/simsopt/geo/biotsavart.py
@@ -14,10 +14,10 @@ class Coil(sopp.Coil, Optimizable):
         sopp.Coil.__init__(self, curve, current)
         Optimizable.__init__(self, x0=np.asarray([]), opts_in=[curve, current])
 
-    def vjp_graph(self, v_gamma, v_gammadash, v_current):
-        return self.curve.dgamma_by_dcoeff_vjp_graph(v_gamma) \
-            + self.curve.dgammadash_by_dcoeff_vjp_graph(v_gammadash) \
-            + self.current.vjp_graph(v_current)
+    def vjp(self, v_gamma, v_gammadash, v_current):
+        return self.curve.dgamma_by_dcoeff_vjp(v_gamma) \
+            + self.curve.dgammadash_by_dcoeff_vjp(v_gammadash) \
+            + self.current.vjp(v_current)
 
 
 class Current(sopp.Current, Optimizable):
@@ -30,8 +30,11 @@ class Current(sopp.Current, Optimizable):
     def set_dofs(self, dofs):
         sopp.Current.set_dofs(self, dofs)
 
-    def vjp_graph(self, v_current):
+    def vjp(self, v_current):
         return Derivative({self: v_current})
+
+    def __neg__(self):
+        return FlippedCurrent(self)
 
 
 class FlippedCurrent(sopp.FlippedCurrent, Optimizable):
@@ -41,11 +44,14 @@ class FlippedCurrent(sopp.FlippedCurrent, Optimizable):
         sopp.FlippedCurrent.__init__(self, basecurrent)
         Optimizable.__init__(self, x0=np.asarray([]), opts_in=[basecurrent])
 
-    def vjp_graph(self, v_current):
-        return self.__basecurrent.vjp_graph(-v_current)
+    def vjp(self, v_current):
+        return self.__basecurrent.vjp(-v_current)
+
+    def __neg__(self):
+        return FlippedCurrent(self)
 
 
-class BiotSavart(sopp.BiotSavart, MagneticField, Optimizable):
+class BiotSavart(sopp.BiotSavart, MagneticField):
     r"""
     Computes the MagneticField induced by a list of closed curves :math:`\Gamma_k` with electric currents :math:`I_k`.
     The field is given by
@@ -58,22 +64,10 @@ class BiotSavart(sopp.BiotSavart, MagneticField, Optimizable):
 
     """
 
-    # def __init__(self, coils, coil_currents):
-    #     assert len(coils) == len(coil_currents)
-    #     assert all(isinstance(item, Curve) for item in coils)
-    #     assert all(isinstance(item, float) for item in coil_currents)
-    #     self.currents_optim = [sopp.Current(c) for c in coil_currents]
-    #     self.coils_optim = [sopp.Coil(curv, curr) for curv, curr in zip(coils, self.currents_optim)]
-    #     self.coils = coils
-    #     self.coil_currents = coil_currents
-    #     MagneticField.__init__(self)
-    #     sopp.BiotSavart.__init__(self, self.coils_optim)
-
     def __init__(self, coils):
         self.__coils = coils
         sopp.BiotSavart.__init__(self, coils)
-        MagneticField.__init__(self)
-        Optimizable.__init__(self, x0=np.asarray([]), opts_in=coils)
+        MagneticField.__init__(self, opts_in=coils)
 
 
     def compute_A(self, compute_derivatives=0):
@@ -90,7 +84,7 @@ class BiotSavart(sopp.BiotSavart, MagneticField, Optimizable):
 
         points = self.get_points_cart_ref()
         npoints = len(points)
[…]
