### 2021-07-16 0d2b67b
Fixed missing import os in optimize_qs_and_islands_simultaneously example
