+++ CMakeLists.txt
@@ -2,12 +2,8 @@ cmake_minimum_required(VERSION 3.5)
 project(simsoptpp)
 #set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake ${CMAKE_MODULE_PATH})
 
-#find_package(PythonInterp 3 REQUIRED)
-#find_package(PythonLibs 3 REQUIRED)
-#find_package(NumPy REQUIRED)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
-
 
 set(Python_FIND_STRATEGY LOCATION)
 find_package(Python 3 COMPONENTS Interpreter Development.Module NumPy)
@@ -38,11 +34,8 @@ IF(DEFINED ENV{CI})
     message(STATUS "CI environment detected. Set compilation flags accordingly (target ivybridge which supports avx).")
     set(CMAKE_CXX_FLAGS "-O3 -march=ivybridge -mfma -ffp-contract=fast")
 elseif(DEFINED ENV{CONDA_BUILD})
-    message(STATUS "conda build environment detected. Set compilation flags accordingly (target ivybridge which supports avx).")
+    message(STATUS "conda build environment detected. Let conda set compilation flags accordingly.")
     # set(CMAKE_CXX_FLAGS "-O3 -march=ivybridge -mfma -ffp-contract=fast")
-    # link_directories("$ENV{PREFIX}/lib")
-    # include_directories("$ENV{PREFIX}/include")
-    message(STATUS "Prefix environment variable is $ENV{PREFIX}")
 else()
     message(STATUS "Local build detected. Set compilation flags accordingly (march=native).")
     set(CMAKE_CXX_FLAGS "-O3 -march=native -mfma -ffp-contract=fast")
@@ -97,9 +90,6 @@ target_link_libraries(profiling PRIVATE fmt::fmt)
 if(OpenMP_CXX_FOUND)
     target_link_libraries(${PROJECT_NAME} PRIVATE OpenMP::OpenMP_CXX)
 endif()
-
-get_target_property(simsopt_inc_dirs ${PROJECT_NAME} INCLUDE_DIRECTORIES)
-message(STATUS "simsopt_inc_dirs: ${simsopt_inc_dirs}")
 
 #get_cmake_property(_variableNames VARIABLES)
 #list (SORT _variableNames)
