### 2021-07-16 8fa3693
Now running combined vmec+spec example in CI integrated tests
