### 2021-07-19 f8aab14
Added documentation for the magnetic field class to concepts.rst
