+++ docs/source/concepts.rst
@@ -235,3 +235,60 @@ If a quantity for the curve is requested, the cache is checked to see whether it
 This cache can be cleared manually by calling ``Curve.invalidate_cache()``.
 This function is called everytime ``Curve.set_dofs()`` is called (and the shape of the curve changes).
 
+
+Magnetic Field Classes
+-----------------
+
+Simsopt contains several magnetic field classes available to be called directly. Any field can be summed with any other field and/or multiplied by a constant parameter.
+
+ToroidalField
+~~~~~~
+
+The :obj:`simsopt.field.magneticfieldclasses.ToroidalField` class initializes a toroidal magnetic field vector acording to the formula :math:`\mathbf B = B_0 \frac{R_0}{R} \mathbf e_\phi`, where :math:`R_0` and :math:`B_0` are input scalar quantities, with :math:`R_0` representing the major radius of the magnetic axis and :math:`B_0` the magnetic field at :math:`R_0`. :math:`R` is the radial coordinate of the cylindrical coordinate system :math:`(R,Z,\phi)`, so that :math:`B` has the expected :math:`1/R` dependence. Given a set of points :math:`(x,y,z)`, :math:`R` is calculated as :math:`R=\sqrt{x^2+y^2}`. The vector :math:`\mathbf e_\phi` is a unit vector pointing in the direction of increasing :math:`\phi`, with :math:`\phi` the standard azimuthal angle of the cylindrical coordinate system :math:`(R,Z,\phi)`. Given a set of points :math:`(x,y,z)`, :math:`\mathbf e_\phi` is calculated as :math:`\mathbf e_\phi=-\sin \phi \mathbf e_x+\cos \phi \mathbf e_y` with :math:`\phi=\arctan(y/x)`. 
+
+PoloidalField
+~~~~~~
+
+The :obj:`simsopt.field.magneticfieldclasses.PoloidalField` class initializes a poloidal magnetic field vector acording to the formula :math:`\mathbf B = B_0 \frac{r}{q R_0} \mathbf e_\theta`, where :math:`R_0, q` and :math:`B_0` are input scalar quantities. :math:`R_0` represents the major radius of the magnetic axis, :math:`B_0` the magnetic field at :math:`r=R_0 q` and :math:`q` the safety factor associated with the sum of a poloidal magnetic field and a toroidal magnetic field with major radius :math:`R_0` and magnetic field on-axis :math:`B_0`. :math:`r` is the radial coordinate of the simple toroidal coordinate system :math:`(r,\phi,\theta)`. Given a set of points :math:`(x,y,z)`, :math:`r` is calculated as :math:`r=\sqrt{(\sqrt{x^2+y^2}-R_0)^2+z^2}`. The vector :math:`\mathbf e_\theta` is a unit vector pointing in the direction of increasing :math:`\theta`, with :math:`\theta` the poloidal angle in the simple toroidal coordinate system :math:`(r,\phi,\theta)`. Given a set of points :math:`(x,y,z)`, :math:`\mathbf e_\theta` is calculated as :math:`\mathbf e_\theta=-\sin \theta \cos \phi \mathbf e_x+\sin \theta \sin \phi \mathbf e_y+\cos \theta \mathbf e_z` with :math:`\phi=\arctan(y/x)` and :math:`\theta=\arctan(z/(\sqrt{x^2+y^2}-R_0))`.
+
+ScalarPotentialRZMagneticField
+~~~~~~
+
[…]
