### 2021-07-19 dbbc0b5
Version bump to 0.4.2 after merging field line tracing PR
