### 2021-07-20 f7e5923
DOFs moved from pandas to numpy arrays
