+++ src/simsopt/_core/graph_optimizable.py
@@ -25,7 +25,7 @@ from ..util.types import RealArray, StrArray, BoolArray, Key
 from .util import ImmutableId, OptimizableMeta
 
 
-class DOFs(pd.DataFrame):
+class PandasDOFs(pd.DataFrame):
     """
     Defines the (D)egrees (O)f (F)reedom(s) associated with optimization
 
@@ -305,6 +305,355 @@ class DOFs(pd.DataFrame):
             string identifiers of the DOFs
         """
         return self.index.values
+
+
+class DOFs:
+    """
+    Defines the (D)egrees (O)f (F)reedom(s) associated with optimization
+
+    This class holds data related to the degrees of freedom
+    associated with an Optimizable object. The class subclasses
+    pandas.DataFrame. To access the data stored in the DOFs class as a
+    pandas dataframe, use the labels under internal column shown in the
+    table below.
+
+    DOFs Dataframe column index table
+
+    =====   =============  ===============
+    Index   External name  Internal column
+    =====   =============  ===============
+    0       x              _x
+    1       free           free
+    2       lower_bounds   _lb
+    3       upper_bounds   _ub
+    =====   =============  ===============
+
+    The class implements the external name column properties in the above
+    table as properties. Additional methods to update bounds, fix/unfix DOFs,
+    etc. are also defined.
+    """
+
+    def __init__(self,
+                 x: RealArray = None,  # To enable empty DOFs object
+                 names: StrArray = None,
+                 free: BoolArray = None,
+                 lower_bounds: RealArray = None,
+                 upper_bounds: RealArray = None) -> None:
+        """
+        Args:
+            x: Numeric values of the DOFs
+            names: Names of the dofs
+            free: Array of boolean values denoting if the DOFs is are free.
+                  False values implies the corresponding DOFs are fixed
+            lower_bounds: Lower bounds for the DOFs. Meaningful only if
+                DOF is not fixed. Default is np.NINF
+            upper_bounds: Upper bounds for the DOFs. Meaningful only if
+                DOF is not fixed. Default is np.inf
+        """
+        if x is None:
+            x = np.array([])
+        else:
+            x = np.array(x, dtype=np.double)
+
+        if names is None:
+            names = [f"x{i}" for i in range(len(x))]
+        assert(len(np.unique(names)) == len(names)) # DOF names should be unique
+
+        if free is not None:
+            free = np.array(free, dtype=np.bool_)
+        else:
+            free = np.full(len(x), True)
+
+        if lower_bounds is not None:
+            lb = np.array(lower_bounds, np.double)
+        else:
+            lb = np.full(len(x), np.NINF)
+
+        if upper_bounds is not None:
+            ub = np.array(upper_bounds, np.double)
+        else:
+            ub = np.full(len(x), np.inf)
+
+        assert(len(x) == len(free) == len(lb) == len(ub) == len(names))
+        self._x = x
+        self._free = free
+        self._lb = lb
+        self._ub = ub
+        self._names = list(names)
+
+    def __len__(self):
+        return len(self._free)
+
+    def fix(self, key: Key) -> None:
+        """
+        Fixes the specified DOF
+
+        Args:
+            key: Key to identify the DOF
+        """
+        if isinstance(key, str):
+            i = self._names.index(key)
+            self._free[i] = False
+        else:
+            self._free[key] = False
+
+    def unfix(self, key: Key) -> None:
+        """
+        Unfixes the specified DOF
+
+        Args:
+            key: Key to identify the DOF
+        """
+        if isinstance(key, str):
+            i = self._names.index(key)
+            self._free[i] = True
+        else:
+            self._free[key] = True
[…]
