### 2021-07-20 4ac390d
Bug fixes in numpy styled DOFs
