### 2021-07-20 4161d78
Disabling tests for graph_surface which is going to be discarded
