+++ tests/geo/test_graph_surface.py
@@ -5,7 +5,7 @@ from simsopt.geo.graph_surface import Surface, SurfaceRZFourier, SurfaceGarabedi
 
 TEST_DIR = Path(__file__).parent / ".." / "test_files"  # ) #.resolve()
 
-
+@unittest.skip
 class SurfaceTests(unittest.TestCase):
     def test_init(self):
         """
@@ -36,6 +36,7 @@ class SurfaceTests(unittest.TestCase):
         #self.assertTrue(s.stellsym)
 
 
+@unittest.skip
 class SurfaceRZFourierTests(unittest.TestCase):
     def test_init(self):
         s = SurfaceRZFourier(mpol=3, ntor=2)
@@ -183,6 +184,7 @@ class SurfaceRZFourierTests(unittest.TestCase):
         self.assertAlmostEqual(s.volume(), true_volume, places=3)
 
 
+@unittest.skip
 class SurfaceGarabedianTests(unittest.TestCase):
     def test_init(self):
         """
