+++ src/simsopt/geo/surfacerzfourier.py
@@ -1,0 +1,303 @@
+import numpy as np
+import simsoptpp as sopp
+from .surface import Surface
+
+
+class SurfaceRZFourier(sopp.SurfaceRZFourier, Surface):
+    r"""`SurfaceRZFourier` is a surface that is represented in cylindrical
+    coordinates using the following Fourier series:
+
+    .. math::
+           r(\theta, \phi) = \sum_{m=0}^{m_{\text{pol}}} \sum_{n=-n_{\text{tor}}}^{n_\text{tor}} [
+               r_{c,m,n} \cos(m \theta - n_{\text{fp}} n \phi)
+               + r_{s,m,n} \sin(m \theta - n_{\text{fp}} n \phi) ]
+
+    and the same for :math:`z(\theta, \phi)`.
+
+    Here, :math:`(r,\phi, z)` are standard cylindrical coordinates, and theta
+    is any poloidal angle.
+
+    Note that for :math:`m=0` we skip the :math:`n<0` term for the cos terms, and the :math:`n \leq 0`
+    for the sin terms.
+
+    In addition, in the ``stellsym=True`` case, we skip the sin terms for :math:`r`, and
+    the cos terms for :math:`z`.
+    """
+
+    def __init__(self, nfp=1, stellsym=True, mpol=1, ntor=0, quadpoints_phi=63, quadpoints_theta=62):
+        if isinstance(quadpoints_phi, np.ndarray):
+            quadpoints_phi = list(quadpoints_phi)
+            quadpoints_theta = list(quadpoints_theta)
+        sopp.SurfaceRZFourier.__init__(self, mpol, ntor, nfp, stellsym, quadpoints_phi, quadpoints_theta)
+        self.rc[0, ntor] = 1.0
+        self.rc[1, ntor] = 0.1
+        self.zs[1, ntor] = 0.1
+        Surface.__init__(self)
+        self.make_names()
+
+    def get_dofs(self):
+        """
+        Return the dofs associated to this surface.
+        """
+        return np.asarray(sopp.SurfaceRZFourier.get_dofs(self))
+
+    def make_names(self):
+        """
+        Form a list of names of the `rc`, `zs`, `rs`, or `zc` array elements.
+        """
+        self.names = self.make_names_helper('rc', True) + self.make_names_helper('zs', False)
+        if not self.stellsym:
+            self.names += self.make_names_helper('rs', False) + self.make_names_helper('zc', True)
+
+    def make_names_helper(self, prefix, include0):
+        if include0:
+            names = [prefix + "(0,0)"]
+        else:
+            names = []
+
+        names += [prefix + '(0,' + str(n) + ')' for n in range(1, self.ntor + 1)]
+        for m in range(1, self.mpol + 1):
+            names += [prefix + '(' + str(m) + ',' + str(n) + ')' for n in range(-self.ntor, self.ntor + 1)]
+        return names
+
+    @classmethod
+    def from_focus(cls, filename, nphi=32, ntheta=32):
+        """
+        Read in a surface from a FOCUS-format file.
+        """
+        f = open(filename, 'r')
+        lines = f.readlines()
+        f.close()
+
+        # Read the line containing Nfou and nfp:
+        splitline = lines[1].split()
+        errmsg = "This does not appear to be a FOCUS-format file."
+        assert len(splitline) == 3, errmsg
+        Nfou = int(splitline[0])
+        nfp = int(splitline[1])
+
+        # Now read the Fourier amplitudes:
+        n = np.full(Nfou, 0)
+        m = np.full(Nfou, 0)
+        rc = np.zeros(Nfou)
+        rs = np.zeros(Nfou)
+        zc = np.zeros(Nfou)
[…]

[…]
