+++ tests/geo/test_surfacehenneberg.py
@@ -1,0 +1,271 @@
+import unittest
+import os
+import logging
+from pathlib import Path
+import numpy as np
+
+from simsopt.geo.surfacehenneberg import SurfaceHenneberg
+
+#logging.basicConfig(level=logging.DEBUG)
+logger = logging.getLogger(__name__)
+
+try:
+    from mpi4py import MPI
+except:
+    MPI = None
+
+try:
+    import vmec
+    vmec_found = True
+except ImportError:
+    vmec_found = False
+
+if (MPI is not None) and vmec_found:
+    from simsopt.mhd.vmec import Vmec
+else:
+    Vmec = None
+
+TEST_DIR = (Path(__file__).parent / ".." / "test_files").resolve()
+
+
+class SurfaceHennebergTests(unittest.TestCase):
+    def test_init_quadpoints(self):
+        """
+        Test that we can initialize the points either by an integer, list,
+        or numpy array.
+        """
+        for phi_points in [32,
+                           np.linspace(0.0, 1.0, 32, endpoint=False),
+                           list(np.linspace(0.0, 1.0, 32, endpoint=False))]:
+
+            for theta_points in [32,
+                                 np.linspace(0.0, 1.0, 32, endpoint=False),
+                                 list(np.linspace(0.0, 1.0, 32, endpoint=False))]:
+
+                surf = SurfaceHenneberg(nfp=5, alpha_fac=-1, mmax=3, nmax=4,
+                                        quadpoints_phi=phi_points,
+                                        quadpoints_theta=theta_points)
+
+    def test_names(self):
+        """
+        Check that the names of the dofs are set correctly.
+        """
+        surf = SurfaceHenneberg(nfp=1, alpha_fac=1, mmax=2, nmax=1)
+        names_correct = ['R0nH(0)', 'R0nH(1)',
+                         'Z0nH(1)',
+                         'bn(0)', 'bn(1)',
+                         'rhomn(0,1)',
+                         'rhomn(1,-1)', 'rhomn(1,0)', 'rhomn(1,1)',
+                         'rhomn(2,-1)', 'rhomn(2,0)', 'rhomn(2,1)']
+        self.assertEqual(surf.names, names_correct)
+
+    def test_set_get_dofs(self):
+        """
+        If the dofs are set to random numbers, and then you get the dofs,
+        you should get what you set.
+        """
+        for mmax in range(1, 4):
+            for nmax in range(4):
+                surf = SurfaceHenneberg(nfp=1, alpha_fac=1, mmax=mmax, nmax=nmax)
+                self.assertEqual(len(surf.names), len(surf.get_dofs()))
+
+                # R0nH has nmax+1
+                # Z0nH has nmax
+                # bn has nmax + 1
+                # rhomn for m=0 has nmax
+                # rhomn for m>0 has 2 * nmax + 1
+                nvals = nmax + 1 \
+                    + nmax \
+                    + nmax + 1 \
+                    + nmax \
+                    + mmax * (2 * nmax + 1)
+                vals = np.random.rand(nvals) - 0.5
+                surf.set_dofs(vals)
+                np.testing.assert_allclose(surf.get_dofs(), vals)
+
+    def test_set_get_rhomn(self):
+        """
+        Test get_rhomn() and set_rhomn(). If you set a particular mode and
+        then get that mode, you should get what you set.
+        """
+        for mmax in range(1, 4):
+            for nmax in range(4):
+                surf = SurfaceHenneberg(nfp=2, alpha_fac=-1, mmax=mmax, nmax=nmax)
+                for m in range(mmax + 1):
[…]
