+++ tests/geo/test_boozersurface.py
@@ -1,0 +1,290 @@
+import unittest
+import numpy as np
+from simsopt.geo.coilcollection import CoilCollection
+from simsopt.geo.boozersurface import BoozerSurface
+from simsopt.geo.biotsavart import BiotSavart
+from simsopt.geo.surfaceobjectives import ToroidalFlux
+from simsopt.geo.surfaceobjectives import Area
+from .surface_test_helpers import get_ncsx_data, get_surface, get_exact_surface
+
+
+surfacetypes_list = ["SurfaceXYZFourier", "SurfaceXYZTensorFourier"]
+stellsym_list = [True, False]
+
+
+class BoozerSurfaceTests(unittest.TestCase):
+    def test_residual(self):
+        """
+        This test loads a SurfaceXYZFourier that interpolates the xyz coordinates
+        of a surface in the NCSX configuration that was computed on a previous
+        branch of pyplasmaopt.  Here, we verify that the  Boozer residual at these 
+        interpolation points is small.
+        """
+
+        s = get_exact_surface()
+        coils, currents, ma = get_ncsx_data()
+        stellarator = CoilCollection(coils, currents, 3, True)
+        bs = BiotSavart(stellarator.coils, stellarator.currents)
+        bs_tf = BiotSavart(stellarator.coils, stellarator.currents)
+
+        weight = 1.
+        tf = ToroidalFlux(s, bs_tf)
+
+        # these data are obtained from `boozer` branch of pyplamsaopt
+        tf_target = 0.41431152
+        iota = -0.44856192
+
+        boozer_surface = BoozerSurface(bs, s, tf, tf_target)
+        x = np.concatenate((s.get_dofs(), [iota]))
+        r0 = boozer_surface.boozer_penalty_constraints(
+            x, derivatives=0, constraint_weight=weight, optimize_G=False, scalarize=False)
+        # the residual should be close to zero for all entries apart from the y
+        # and z coordinate at phi=0 and theta=0 (and the corresponding rotations)
+        ignores_idxs = np.zeros_like(r0)
+        ignores_idxs[[1, 2, 693, 694, 695, 1386, 1387, 1388, -2, -1]] = 1
+        assert np.max(np.abs(r0[ignores_idxs < 0.5])) < 1e-8
+        assert np.max(np.abs(r0[-2:])) < 1e-6
+
+    def test_boozer_penalty_constraints_gradient(self):
+        """
+        Taylor test to verify the gradient of the scalarized constrained optimization problem's
+        objective.
+        """
+        for surfacetype in surfacetypes_list:
+            for stellsym in stellsym_list:
+                for optimize_G in [True, False]:
+                    with self.subTest(surfacetype=surfacetype, stellsym=stellsym, optimize_G=optimize_G):
+                        self.subtest_boozer_penalty_constraints_gradient(surfacetype, stellsym, optimize_G)
+
+    def test_boozer_penalty_constraints_hessian(self):
+        """
+        Taylor test to verify the Hessian of the scalarized constrained optimization problem's
+        objective.
+        """
+        for surfacetype in surfacetypes_list:
+            for stellsym in stellsym_list:
+                for optimize_G in [True, False]:
+                    with self.subTest(surfacetype=surfacetype, stellsym=stellsym, optimize_G=optimize_G):
+                        self.subtest_boozer_penalty_constraints_hessian(surfacetype, stellsym, optimize_G)
+
+    def subtest_boozer_penalty_constraints_gradient(self, surfacetype, stellsym, optimize_G=False):
+        np.random.seed(1)
+        coils, currents, ma = get_ncsx_data()
[…]
