### 2021-07-21 b4f1a0c
Preliminary changes for graph surface framework
