+++ src/simsopt/geo/graph_surface.py
@@ -1,6 +1,19 @@
-import numpy as np
+import abc
 
-from .._core.optimizable import Optimizable
+import numpy as np
+from monty.dev import requires
+
+try:
+    from mayavi import mlab
+except ImportError:
+    mlab = None
+
+try:
+    from pyevtk.hl import gridToVTK
+except ImportError:
+    gridToVTK = None
+
+from .._core.graph_optimizable import Optimizable
 
 
 class Surface(Optimizable):
@@ -8,16 +21,17 @@ class Surface(Optimizable):
     ``Surface`` is a base class for various representations of toroidal
     surfaces in simsopt.
 
-    A ``Surface`` is modelled as a function :math:`\Gamma:[0, 1] \times [0, 1] \to R^3` and is evaluated at quadrature points :math:`\{\phi_1, \ldots, \phi_{n_\phi}\}\times\{\theta_1, \ldots, \theta_{n_\theta}\}`.
-
+    A ``Surface`` is modelled as a function
+    :math:`\Gamma:[0, 1] \times [0, 1] \to R^3` and is evaluated at
+    quadrature points :math:`\{\phi_1, \ldots, \phi_{n_\phi}\}\times\{\theta_1, \ldots, \theta_{n_\theta}\}`.
     """
 
-    def __init__(self):
-        Optimizable.__init__(self)
-        self.dependencies = []
-        self.fixed = np.full(len(self.get_dofs()), False)
+    def __init__(self, **kwargs):
+        Optimizable.__init__(self, x0=self.get_dofs(), **kwargs)
 
-    def plot(self, ax=None, show=True, plot_normal=False, plot_derivative=False, scalars=None, wireframe=True):
+    @requires(mlab is not None, "Install mayavi to generate surface plot")
+    def plot(self, ax=None, show=True, plot_normal=False, plot_derivative=False,
+             scalars=None, wireframe=True):
         """
         Plot the surface using mayavi. 
         Note: the `ax` and `show` parameter can be used to plot more than one surface:
@@ -27,12 +41,9 @@ class Surface(Optimizable):
             ax = surface1.plot(show=False)
             ax = surface2.plot(ax=ax, show=False)
             surface3.plot(ax=ax, show=True)
-
-
         """
         gamma = self.gamma()
 
-        from mayavi import mlab
         mlab.mesh(gamma[:, :, 0], gamma[:, :, 1], gamma[:, :, 2], scalars=scalars)
         if wireframe:
             mlab.mesh(gamma[:, :, 0], gamma[:, :, 1], gamma[:, :, 2], representation='wireframe', color=(0, 0, 0), opacity=0.5)
@@ -48,8 +59,8 @@ class Surface(Optimizable):
         if show:
             mlab.show()
 
+    @requires(gridToVTK is not None, "to_vtk method requires pyevtk module")
     def to_vtk(self, filename):
-        from pyevtk.hl import gridToVTK
         g = self.gamma()
         ntor = g.shape[0]
         npol = g.shape[1]
@@ -66,22 +77,24 @@ class Surface(Optimizable):
             "dtheta": (contig(dtheta[..., 0]), contig(dtheta[..., 1]), contig(dtheta[..., 2])),
         })
 
-    def __repr__(self):
-        return "Surface " + str(hex(id(self)))
+    # Representation is handled by Optimizable object itself
+    # def __repr__(self):
+    #     return "Surface " + str(hex(id(self)))
 
+    @abc.abstractmethod
     def to_RZFourier(self):
         """
-        Return a :obj:`simsopt.geo.surfacerzfourier.SurfaceRZFourier` instance corresponding to the shape of this
-        surface.  All subclasses should implement this abstract
-        method.
[…]
