+++ src/simsopt/geo/graph_surfacerzfourier.py
@@ -1,6 +1,6 @@
 import numpy as np
 import simsoptpp as sopp
-from .surface import Surface
+from .graph_surface import Surface
 
 
 class SurfaceRZFourier(sopp.SurfaceRZFourier, Surface):
@@ -24,16 +24,19 @@ class SurfaceRZFourier(sopp.SurfaceRZFourier, Surface):
     the cos terms for :math:`z`.
     """
 
-    def __init__(self, nfp=1, stellsym=True, mpol=1, ntor=0, quadpoints_phi=63, quadpoints_theta=62):
+    def __init__(self, nfp=1, stellsym=True, mpol=1, ntor=0, quadpoints_phi=63,
+                 quadpoints_theta=62):
         if isinstance(quadpoints_phi, np.ndarray):
             quadpoints_phi = list(quadpoints_phi)
             quadpoints_theta = list(quadpoints_theta)
-        sopp.SurfaceRZFourier.__init__(self, mpol, ntor, nfp, stellsym, quadpoints_phi, quadpoints_theta)
+        sopp.SurfaceRZFourier.__init__(self, mpol, ntor, nfp, stellsym,
+                                       quadpoints_phi, quadpoints_theta)
         self.rc[0, ntor] = 1.0
         self.rc[1, ntor] = 0.1
         self.zs[1, ntor] = 0.1
-        Surface.__init__(self)
-        self.make_names()
+        Surface.__init__(self, external_dof_setter=SurfaceRZFourier.set_dofs,
+                         names=self.make_names())
+        # self.make_names()
 
     def get_dofs(self):
         """
@@ -45,9 +48,10 @@ class SurfaceRZFourier(sopp.SurfaceRZFourier, Surface):
         """
         Form a list of names of the `rc`, `zs`, `rs`, or `zc` array elements.
         """
-        self.names = self.make_names_helper('rc', True) + self.make_names_helper('zs', False)
+        names = self.make_names_helper('rc', True) + self.make_names_helper('zs', False)
         if not self.stellsym:
-            self.names += self.make_names_helper('rs', False) + self.make_names_helper('zc', True)
+            names += self.make_names_helper('rs', False) + self.make_names_helper('zc', True)
+        return names
 
     def make_names_helper(self, prefix, include0):
         if include0:
@@ -65,9 +69,8 @@ class SurfaceRZFourier(sopp.SurfaceRZFourier, Surface):
         """
         Read in a surface from a FOCUS-format file.
         """
-        f = open(filename, 'r')
-        lines = f.readlines()
-        f.close()
+        with open(filename, 'r') as f:
+            lines = f.readlines()
 
         # Read the line containing Nfou and nfp:
         splitline = lines[1].split()
@@ -96,7 +99,8 @@ class SurfaceRZFourier(sopp.SurfaceRZFourier, Surface):
         mpol = int(np.max(m))
         ntor = int(np.max(np.abs(n)))
 
-        surf = cls(mpol=mpol, ntor=ntor, nfp=nfp, stellsym=stellsym, quadpoints_phi=nphi, quadpoints_theta=ntheta)
+        surf = cls(mpol=mpol, ntor=ntor, nfp=nfp, stellsym=stellsym,
+                   quadpoints_phi=nphi, quadpoints_theta=ntheta)
         for j in range(Nfou):
             surf.rc[m[j], n[j] + ntor] = rc[j]
             surf.zs[m[j], n[j] + ntor] = zs[j]
@@ -106,6 +110,7 @@ class SurfaceRZFourier(sopp.SurfaceRZFourier, Surface):
 
         return surf
 
+    # TODO: Needs reimplementation
     def change_resolution(self, mpol, ntor):
         """
[…]
