### 2021-07-21 312f668
Preliminary changes for graph surface framework
