### 2021-07-22 d2e062f
Prevent recompute_bell called twice for x setter
