### 2021-07-22 c2965f0
Documentation on recompute_bell updated
