### 2021-07-22 41fac54
Bug fixes and enhancements in DOF setters
