### 2021-07-22 3aa079b
Remove get_dofs as default argument to Optimizable from Surface class
