+++ src/simsopt/geo/graph_surfacegarabedian.py
@@ -36,37 +36,42 @@ class SurfaceGarabedian(Surface):
         self.nmax = nmax
         self.nfp = nfp
         self.stellsym = True
-        self.allocate()
-        self.recalculate = True
+
+        self.mdim = self.mmax - self.mmin + 1
+        self.ndim = self.nmax - self.nmin + 1
+        self.shape = (self.mdim, self.ndim)
+
         self.recalculate_derivs = True
+
+        Delta = np.zeros(self.shape)
+        Surface.__init__(self, Delta.flatten(),
+                         names=self._make_dof_names())
 
         # Initialize to an axisymmetric torus with major radius 1m and
         # minor radius 0.1m
         self.set_Delta(1, 0, 1.0)
         self.set_Delta(0, 0, 0.1)
-        Surface.__init__(self, x0=self.get_dofs(),
-                         external_dof_setter=SurfaceGarabedian.set_dofs)
 
-    # TODO: Reimplement
+    def _make_dof_names(self):
+        names = []
+        for m in range(self.mmin, self.mmax + 1):
+            for n in range(self.nmin, self.nmax + 1):
+                names.append(f'Delta({m},{n})')
+        return names
+
     def __repr__(self):
-        return "SurfaceGarabedian " + str(hex(id(self))) + " (nfp=" + \
-            str(self.nfp) + ", mmin=" + str(self.mmin) + ", mmax=" + str(self.mmax) \
-            + ", nmin=" + str(self.nmin) + ", nmax=" + str(self.nmax) \
-            + ")"
+        return self.name + f" (nfp={self.nfp}, " + \
+               f"mmin={self.mmin}, mmax={self.mmax}" + \
+               f", nmin={self.nmin}, nmax={self.nmax})"
 
-    def allocate(self):
-        """
-        Create the array for the :math:`\Delta_{m,n}` coefficients.
-        """
-        logger.info("Allocating SurfaceGarabedian")
-        self.mdim = self.mmax - self.mmin + 1
-        self.ndim = self.nmax - self.nmin + 1
-        myshape = (self.mdim, self.ndim)
-        self.Delta = np.zeros(myshape)
-        self.names = []
-        for n in range(self.nmin, self.nmax + 1):
-            for m in range(self.mmin, self.mmax + 1):
-                self.names.append('Delta(' + str(m) + ',' + str(n) + ')')
+    @property
+    def Delta(self):
+        return self.local_full_x.reshape(self.shape)
+
+    @Delta.setter
+    def Delta(self, Delta):
+        assert(self.shape == Delta.shape)
+        self.local_full_x = Delta.flatten()
 
     def get_Delta(self, m, n):
         """
@@ -78,41 +83,37 @@ class SurfaceGarabedian(Surface):
         """
         Set a particular :math:`\Delta_{m,n}` coefficient.
         """
-        self.Delta[m - self.mmin, n - self.nmin] = val
-        self.recalculate = True
+        i = self.ndim * (m - self.mmin) + n - self.nmin
+        self.set(i, val)
+        # self.Delta[m - self.mmin, n - self.nmin] = val
+        # self.recalculate = True
         self.recalculate_derivs = True
 
     def get_dofs(self):
         """
         Return a 1D numpy array with all the degrees of freedom.
         """
-        num_dofs = (self.mmax - self.mmin + 1) * (self.nmax - self.nmin + 1)
-        return np.reshape(self.Delta, (num_dofs,), order='F')
+        self.local_full_x
 
     def set_dofs(self, v):
         """
         Set the shape coefficients from a 1D list/array
         """
-
-        n = self.dof_size
-        if len(v) != n:
-            raise ValueError('Input vector should have ' + str(n) + \
-                             ' elements but instead has ' + str(len(v)))
-
[…]
