### 2021-07-22 cace5ef
Preliminary modifications to surfacexyzfourier for graph framework
