### 2021-07-22 50f8824
Bug fix in argument to base class init
