### 2021-07-22 a4c40ba
Prior tests slightly modified for graph surfaces
