+++ tests/geo/test_graphsurface_xyzfourier.py
@@ -1,14 +1,9 @@
 import unittest
-from pathlib import Path
 
 import numpy as np
 
-from simsopt.geo.graph_surfacerzfourier import SurfaceRZFourier
-from simsopt.geo.graph_surfacegarabedian import SurfaceGarabedian
 from simsopt.geo.graph_surfacexyzfourier import SurfaceXYZFourier
-from .surface_test_helpers import get_ncsx_data, get_surface, get_exact_surface
-
-TEST_DIR = Path(__file__).parent / ".." / "test_files"
+from .surface_test_helpers import get_surface, get_exact_surface
 
 stellsym_list = [True, False]
 
@@ -289,275 +284,6 @@ class SurfaceXYZFourierTests(unittest.TestCase):
         s = SurfaceXYZFourier(mpol=mpol, ntor=ntor, nfp=nfp, stellsym=stellsym, quadpoints_phi=phis, quadpoints_theta=thetas)
 
         s.to_vtk('/tmp/surface')
-
-
-class SurfaceRZFourierTests(unittest.TestCase):
-    def test_aspect_ratio(self):
-        """
-        Test that the aspect ratio of a torus with random minor and major radius 0.1 <= minor_R <= major_R
-        is properly computed to be major_R/minor_R.
-        """
-
-        s = SurfaceRZFourier(nfp=2, mpol=3, ntor=2)
-        s.rc = s.rc * 0
-        s.rs = s.rs * 0
-        s.zc = s.zc * 0
-        s.zs = s.zs * 0
-        r1 = np.random.random_sample() + 0.1
-        r2 = np.random.random_sample() + 0.1
-        major_R = np.max([r1, r2])
-        minor_R = np.min([r1, r2])
-        s.rc[0, 2] = major_R
-        s.rc[1, 2] = minor_R
-        s.zs[1, 2] = minor_R
-        print("AR approx: ", s.aspect_ratio(), "Exact: ", major_R/minor_R)
-        self.assertAlmostEqual(s.aspect_ratio(), major_R/minor_R)
-
-    def test_init(self):
-        s = SurfaceRZFourier(nfp=2, mpol=3, ntor=2)
-        self.assertEqual(s.rc.shape, (4, 5))
-        self.assertEqual(s.zs.shape, (4, 5))
-
-        s = SurfaceRZFourier(nfp=10, mpol=1, ntor=3, stellsym=False)
-        self.assertEqual(s.rc.shape, (2, 7))
-        self.assertEqual(s.zs.shape, (2, 7))
-        self.assertEqual(s.rs.shape, (2, 7))
-        self.assertEqual(s.zc.shape, (2, 7))
-
-    def test_area_volume(self):
-        """
-        Test the calculation of area and volume for an axisymmetric surface
-        """
-        s = SurfaceRZFourier()
-        s.rc[0, 0] = 1.3
-        s.rc[1, 0] = 0.4
-        s.zs[1, 0] = 0.2
-
-        true_area = 15.827322032265993
-        true_volume = 2.0528777154265874
-        self.assertAlmostEqual(s.area(), true_area, places=4)
-        self.assertAlmostEqual(s.volume(), true_volume, places=3)
-
-    def test_get_dofs(self):
-        """
-        Test that we can convert the degrees of freedom into a 1D vector
-        """
-
-        # First try an axisymmetric surface for simplicity:
-        s = SurfaceRZFourier()
-        s.rc[0, 0] = 1.3
-        s.rc[1, 0] = 0.4
-        s.zs[0, 0] = 0.3
-        s.zs[1, 0] = 0.2
-        dofs = s.get_dofs()
-        self.assertEqual(dofs.shape, (3,))
-        self.assertAlmostEqual(dofs[0], 1.3)
[…]
