### 2021-07-22 1de59cd
Cleanup of surface rzfourier tests after splitting from surface tests
