+++ tests/geo/test_graphsurface_rzfourier.py
@@ -4,9 +4,6 @@ from pathlib import Path
 import numpy as np
 
 from simsopt.geo.graph_surfacerzfourier import SurfaceRZFourier
-from simsopt.geo.graph_surfacegarabedian import SurfaceGarabedian
-from simsopt.geo.graph_surfacexyzfourier import SurfaceXYZFourier
-from .surface_test_helpers import get_ncsx_data, get_surface, get_exact_surface
 
 TEST_DIR = Path(__file__).parent / ".." / "test_files"
 
@@ -17,278 +14,6 @@ try:
     pyevtk_found = True
 except ImportError:
     pyevtk_found = False
-
-
-class SurfaceXYZFourierTests(unittest.TestCase):
-    def test_toRZFourier_perfect_torus(self):
-        """
-        This test checks that a perfect torus can be converted from
-        SurfaceXYZFourier to SurfaceRZFourier completely losslessly.
-        """
-        for stellsym in stellsym_list:
-            with self.subTest(stellsym=stellsym):
-                self.subtest_toRZFourier_perfect_torus("SurfaceXYZFourier",
-                                                       stellsym)
-
-    def subtest_toRZFourier_perfect_torus(self, surfacetype, stellsym):
-        """
-        The test obtains a perfect torus as a SurfaceXYZFourier, then converts
-        it to a SurfaceRZFourier.  Next, it computes the cross section of both
-        surfaces at a random angle and compares the pointwise values.
-        """
-        s = get_surface(surfacetype, stellsym)
-        sRZ = s.to_RZFourier()
-
-        np.random.seed(0)
-        angle = np.random.random()*1000
-        scs = s.cross_section(angle, thetas=100)
-        sRZcs = sRZ.cross_section(angle, thetas=100)
-
-        max_pointwise_err = np.max(np.abs(scs - sRZcs))
-        print(max_pointwise_err)
-
-        # compute the cylindrical angle error of the cross section
-        an = np.arctan2(scs[:, 1], scs[:, 0])
-        phi = angle
-        phi = phi - np.sign(phi) * np.floor(np.abs(phi) / (2*np.pi)) * (2. * np.pi)
-        if phi > np.pi:
-            phi = phi - 2. * np.pi
-        if phi < -np.pi:
-            phi = phi + 2. * np.pi
-        max_angle_err = np.max(np.abs(an - phi))
-
-        # check that the angle error is what we expect
-        assert max_angle_err < 1e-12
-        # check that the pointwise error is what we expect
-        assert max_pointwise_err < 1e-12
-
-    def test_toRZFourier_lossless_at_quadraturepoints(self):
-        """
-        This test obtains a more complex surface (not a perfect torus) as a SurfaceXYZFourier, then
-        converts that surface to the SurfaceRZFourier representation.  Then, the test checks that both
-        surface representations coincide at the points where the least squares fit was completed,
-        i.e., the conversion is lossless at the quadrature points.
-
-        Additionally, this test checks that the cross sectional angle is correct.
-        """
-        s = get_exact_surface()
-        sRZ = s.to_RZFourier()
-
-        max_angle_error = -1
-        max_pointwise_error = -1
-        for angle in sRZ.quadpoints_phi:
-            scs = s.cross_section(angle * 2 * np.pi)
-            sRZcs = sRZ.cross_section(angle * 2 * np.pi)
-
[…]
