### 2021-07-22 #3525 @dfm
@ZedThree, @StanczakDominik, @landreman – This is the review thread for the paper. All of our communications will happen here from now on. Thanks again for agreeing to participate!
