### 2021-07-22 #3525 @dfm
☝️ Since the paper lives on the `joss_paper` branch in the repo, I think we'll need to use this command to compile the paper.
