### 2021-07-23 710cd1a
maximize numpy views in assigning dofs
