### 2021-07-23 39f2b32
Formatting of doc strings
