### 2021-07-23 f107343
Make all surface objectives subclasses of graph Optimizable
