### 2021-07-23 f8c7311
Preliminary commit for graph framework conversion
