+++ src/simsopt/geo/graph_surfacexyztensorfourier.py
@@ -1,7 +1,7 @@
 import numpy as np
 import simsoptpp as sopp
-from .surface import Surface
-from .surfacerzfourier import SurfaceRZFourier
+from .graph_surface import Surface
+from .graph_surfacerzfourier import SurfaceRZFourier
 
 
 class SurfaceXYZTensorFourier(sopp.SurfaceXYZTensorFourier, Surface):
@@ -36,17 +36,20 @@ class SurfaceXYZTensorFourier(sopp.SurfaceXYZTensorFourier, Surface):
 
     """
 
-    def __init__(self, nfp=1, stellsym=True, mpol=1, ntor=1, clamped_dims=[False, False, False],
+    def __init__(self, nfp=1, stellsym=True, mpol=1, ntor=1,
+                 clamped_dims=[False, False, False],
                  quadpoints_phi=32, quadpoints_theta=32):
         if isinstance(quadpoints_phi, np.ndarray):
             quadpoints_phi = list(quadpoints_phi)
             quadpoints_theta = list(quadpoints_theta)
-        sopp.SurfaceXYZTensorFourier.__init__(self, mpol, ntor, nfp, stellsym, clamped_dims,
-                                              quadpoints_phi, quadpoints_theta)
+        sopp.SurfaceXYZTensorFourier.__init__(self, mpol, ntor, nfp, stellsym,
+                                              clamped_dims, quadpoints_phi,
+                                              quadpoints_theta)
         self.x[0, 0] = 1.0
         self.x[1, 0] = 0.1
         self.z[mpol+1, 0] = 0.1
-        Surface.__init__(self)
+        Surface.__init__(self, x0=self.get_dofs(),
+                         external_dof_setter=SurfaceXYZTensorFourier.set_dofs)
 
     def get_dofs(self):
         """
@@ -54,25 +57,28 @@ class SurfaceXYZTensorFourier(sopp.SurfaceXYZTensorFourier, Surface):
         """
         return np.asarray(sopp.SurfaceXYZTensorFourier.get_dofs(self))
 
-    def to_RZFourier(self):
-        """
-        Return a SurfaceRZFourier instance corresponding to the shape of this
-        surface.
-        """
-        surf = SurfaceRZFourier(self.mpol, self.ntor, self.nfp, self.stellsym, self.quadpoints_phi, self.quadpoints_theta)
-        gamma = np.zeros((surf.quadpoints_phi.size, surf.quadpoints_theta.size, 3))
-        for idx in range(gamma.shape[0]):
-            gamma[idx, :, :] = self.cross_section(surf.quadpoints_phi[idx]*2*np.pi)
-        surf.least_squares_fit(self.gamma())
-        return surf
-
     def set_dofs(self, dofs):
         """
         Set the dofs associated to this surface.
         """
         sopp.SurfaceXYZTensorFourier.set_dofs(self, dofs)
-        for d in self.dependencies:
-            d.invalidate_cache()
+
+    def recompute_bell(self, parent=None):
+        self.invalidate_cache()
+
+    def to_RZFourier(self):
+        """
+        Return a SurfaceRZFourier instance corresponding to the shape of this
+        surface.
+        """
+        surf = SurfaceRZFourier(self.mpol, self.ntor, self.nfp, self.stellsym,
+                                self.quadpoints_phi, self.quadpoints_theta)
+        gamma = np.zeros((surf.quadpoints_phi.size, surf.quadpoints_theta.size, 3))
+        for idx in range(gamma.shape[0]):
+            gamma[idx, :, :] = self.cross_section(
+                surf.quadpoints_phi[idx]*2*np.pi)
+        surf.least_squares_fit(self.gamma())
+        return surf
 
     def get_stellsym_mask(self):
         """
