### 2021-07-23 81284e6
Minor Cleanup and conversion of surface tests for graph framework
