+++ tests/geo/test_graphsurface_xyzfourier.py
@@ -15,6 +15,155 @@ except ImportError:
 
 
 class SurfaceXYZFourierTests(unittest.TestCase):
+
+    def test_cross_section_torus(self):
+        """
+        Test that the cross sectional area at a certain number of cross sections of a torus
+        is what it should be.  The cross sectional angles are chosen to test any degenerate
+        cases of the bisection algorithm.
+
+        Additionally, this test checks that the cross sectional angle is correct.
+        """
+        mpol = 4
+        ntor = 3
+        nfp = 2
+        phis = np.linspace(0, 1, 31, endpoint=False)
+        thetas = np.linspace(0, 1, 31, endpoint=False)
+
+        np.random.seed(0)
+
+        stellsym = False
+        s = SurfaceXYZFourier(mpol=mpol, ntor=ntor, nfp=nfp, stellsym=stellsym, quadpoints_phi=phis, quadpoints_theta=thetas)
+        s.xc = s.xc * 0
+        s.xs = s.xs * 0
+        s.ys = s.ys * 0
+        s.yc = s.yc * 0
+        s.zs = s.zs * 0
+        s.zc = s.zc * 0
+        r1 = np.random.random_sample() + 0.1
+        r2 = np.random.random_sample() + 0.1
+        major_R = np.max([r1, r2])
+        minor_R = np.min([r1, r2])
+        s.xc[0, ntor] = major_R
+        s.xc[1, ntor] = minor_R
+        s.zs[1, ntor] = minor_R
+
+        num_cs = 9
+        angle = np.zeros((num_cs,))
+        angle[0] = 0.
+        angle[1] = np.pi/2.
+        angle[2] = np.pi
+        angle[3] = 3 * np.pi / 2.
+        angle[4] = 2. * np.pi
+        angle[5] = -np.pi/2.
+        angle[6] = -np.pi
+        angle[7] = -3. * np.pi / 2.
+        angle[8] = -2. * np.pi
+        cs = np.zeros((num_cs, 100, 3))
+        for idx in range(angle.size):
+            cs[idx, :, :] = s.cross_section(angle[idx], thetas=100)
+
+        cs_area = np.zeros((num_cs,))
+        max_angle_error = -1
+
+        from scipy import fftpack
+        for i in range(num_cs):
+
+            phi = angle[i]
+            phi = phi - np.sign(phi) * np.floor(np.abs(phi) / (2*np.pi)) * (2. * np.pi)
+            if phi > np.pi:
+                phi = phi - 2. * np.pi
+            if phi < -np.pi:
+                phi = phi + 2. * np.pi
+
+            # check that the angle of the cross section is what we expect
+            an = np.arctan2(cs[i, :, 1], cs[i, :, 0])
+            curr_angle_err = np.max(np.abs(an - phi))
+
+            if max_angle_error < curr_angle_err:
+                max_angle_error = curr_angle_err
+
+            R = np.sqrt(cs[i, :, 0]**2 + cs[i, :, 1]**2)
+            Z = cs[i, :, 2]
+            Rp = fftpack.diff(R, period=1.)
+            Zp = fftpack.diff(Z, period=1.)
+            cs_area[i] = np.abs(np.mean(Z*Rp))
+        exact_area = np.pi * minor_R**2.
+
+        # check that the cross sectional area is what we expect
+        assert np.max(np.abs(cs_area - exact_area)) < 1e-14
+        # check that the angle error is what we expect
+        assert max_angle_error < 1e-12
+
+    def test_aspect_ratio_random_torus(self):
+        """
+        This is a simple aspect ratio validation on a torus with minor radius = r1
[…]
