+++ tests/geo/test_graphsurface_taylor.py
@@ -1,6 +1,10 @@
-import numpy as np
 import unittest
+import numpy as np
 from simsopt.geo import parameters
+from simsopt.geo.graph_surfacerzfourier import SurfaceRZFourier
+from simsopt.geo.graph_surfacexyzfourier import SurfaceXYZFourier
+from simsopt.geo.graph_surfacexyztensorfourier import SurfaceXYZTensorFourier
+
 parameters['jit'] = False
 
 
@@ -12,7 +16,7 @@ def taylor_test(f, df, x, epsilons=None, direction=None, order=2):
     dfx = df(x)@direction
     if epsilons is None:
         epsilons = np.power(2., -np.asarray(range(8, 20)))
-    print("################################################################################")
+    print("###################################################################")
     err_old = 1e9
     counter = 0
     for eps in epsilons:
@@ -27,7 +31,8 @@ def taylor_test(f, df, x, epsilons=None, direction=None, order=2):
             fplus2eps = f(x + 2*eps * direction)
             fminus2eps = f(x - 2*eps * direction)
             fak = 0.13
-            dfest = ((1/12) * fminus2eps - (2/3) * fminuseps + (2/3)*fpluseps - (1/12)*fplus2eps)/eps
+            dfest = ((1/12) * fminus2eps - (2/3) * fminuseps + (2/3)*fpluseps
+                     - (1/12)*fplus2eps)/eps
         else:
             raise NotImplementedError
         err = np.linalg.norm(dfest - dfx)
@@ -39,7 +44,7 @@ def taylor_test(f, df, x, epsilons=None, direction=None, order=2):
         err_old = err
     if err > 1e-10:
         assert counter > 2
-    print("################################################################################")
+    print("###################################################################")
 
 
 def get_surface(surfacetype, stellsym, phis=None, thetas=None):
@@ -50,15 +55,15 @@ def get_surface(surfacetype, stellsym, phis=None, thetas=None):
     phis = phis if phis is not None else np.linspace(0, 1, 31, endpoint=False)
     thetas = thetas if thetas is not None else np.linspace(0, 1, 31, endpoint=False)
     if surfacetype == "SurfaceRZFourier":
-        from simsopt.geo.surfacerzfourier import SurfaceRZFourier
-        s = SurfaceRZFourier(nfp=nfp, stellsym=stellsym, mpol=mpol, ntor=ntor, quadpoints_phi=phis, quadpoints_theta=thetas)
+        s = SurfaceRZFourier(nfp=nfp, stellsym=stellsym, mpol=mpol, ntor=ntor,
+                             quadpoints_phi=phis, quadpoints_theta=thetas)
         s.set_dofs(s.get_dofs()*0.)
         s.rc[0, ntor + 0] = 1
         s.rc[1, ntor + 0] = 0.3
         s.zs[1, ntor + 0] = 0.3
     elif surfacetype == "SurfaceXYZFourier":
-        from simsopt.geo.surfacexyzfourier import SurfaceXYZFourier
-        s = SurfaceXYZFourier(nfp=nfp, stellsym=stellsym, mpol=mpol, ntor=ntor, quadpoints_phi=phis, quadpoints_theta=thetas)
+        s = SurfaceXYZFourier(nfp=nfp, stellsym=stellsym, mpol=mpol, ntor=ntor,
+                              quadpoints_phi=phis, quadpoints_theta=thetas)
         s.set_dofs(s.get_dofs()*0.)
         s.xc[0, ntor + 1] = 1.
         s.xc[1, ntor + 1] = 0.1
@@ -66,7 +71,6 @@ def get_surface(surfacetype, stellsym, phis=None, thetas=None):
[…]

[…]
