### 2021-07-23 d61d4df
Unittest for iota_target_metric adjoint method.
