### 2021-07-23 749903d
Unittest for iota_target_metric adjoint method.
