+++ src/simsopt/mhd/vmec.py
@@ -33,13 +33,12 @@ except ImportError as e:
 from .._core.optimizable import Optimizable
 from .._core.util import Struct, ObjectiveFailure
 from ..geo.surfacerzfourier import SurfaceRZFourier
+from ..geo.surfaceobjectives import parameter_derivatives
 
 if MPI is not None:
     from ..util.mpi import MpiPartition
 else:
     MpiPartition = None
-
-print('in vmec')
 
 # Flags used by runvmec():
 restart_flag = 1
@@ -167,7 +166,9 @@ class Vmec(Optimizable):
     def __init__(self,
                  filename: Union[str, None] = None,
                  mpi: Union[MpiPartition, None] = None,
-                 keep_all_files: bool = False):
+                 keep_all_files: bool = False,
+                 ntheta=50,
+                 nphi=50):
         if MPI is None:
             raise RuntimeError("mpi4py needs to be installed for running VMEC")
         if vmec is None:
@@ -229,10 +230,14 @@ class Vmec(Optimizable):
         # object, but the mpol/ntor values of either the vmec object
         # or the boundary surface object can be changed independently
         # by the user.
+        quadpoints_theta = np.linspace(0,1.,ntheta,endpoint=False)
+        quadpoints_phi = np.linspace(0,1.,nphi,endpoint=False)
         self.boundary = SurfaceRZFourier(nfp=vi.nfp,
                                          stellsym=not vi.lasym,
                                          mpol=vi.mpol,
-                                         ntor=vi.ntor)
+                                         ntor=vi.ntor,
+                                         quadpoints_theta=quadpoints_theta,
+                                         quadpoints_phi=quadpoints_phi)
         self.free_boundary = bool(vi.lfreeb)
 
         # Transfer boundary shape data from fortran to the ParameterArray:
@@ -607,7 +612,7 @@ class Vmec(Optimizable):
         return np.sum(weight_function(self.s_half_grid) * self.wout.iotas[1:]) \
             / np.sum(weight_function(self.s_half_grid))
 
-    def iota_target(self,iota_target):
+    def iota_target_metric(self,iota_target):
         """
         Computes a metric quantifying the deviation of iota from a prescribed
         target value.
@@ -620,89 +625,131 @@ class Vmec(Optimizable):
                 transform.
         """
         self.run()
-
         return 0.5 * np.sum((self.wout.iotas[1::]
             - iota_target(self.s_half_grid))**2) * self.ds
 
-    def d_iota_target(self,iota_target,delta):
+    def d_iota_target_metric(self,iota_target,delta=1.):
+        """
+        Computes derivatives of iota_target_metric wrt surface parameters using
+        an adjoint method. The parameter delta sets the amplitude of the toroidal
+        current perturbation required for the adjoint solve.
+        """
         if self.indata.ncurr != 1:
-            raise RuntimeError('''d_iota_target cannot be computed without
+            raise RuntimeError('''d_iota_target_metric cannot be computed without
                 running vmec with ncurr = 1''')
 
-        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0,1::] / (4*np.pi*1e-7)
-        ac_aux_f_prev = self.indata.ac_aux_f
[…]
