### 2021-07-25 182b786
Added IotaTargetMetric class. Update to example and tests.
