+++ tests/mhd/test_vmec.py
@@ -200,16 +200,16 @@ class VmecTests(unittest.TestCase):
         yields approximately same result.
         """
         filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename,ntheta=50,nphi=50)
+        vmec = Vmec(filename, ntheta=50, nphi=50)
 
-        target_function = lambda s : np.cos(s)
+        target_function = lambda s: np.cos(s)
         metric1 = vmec.iota_target_metric(target_function)
         stencil = np.ones_like(vmec.s_full_grid)
         stencil[0] = 0.5
         stencil[-1] = 0.5
         metric2 = 0.5 * np.sum((vmec.wout.iotaf - target_function(vmec.s_full_grid))**2 * stencil) * vmec.ds
 
-        self.assertAlmostEqual(metric1,metric2,places=3)
+        self.assertAlmostEqual(metric1, metric2, places=3)
 
     def test_iota_weighted(self):
         """
@@ -217,13 +217,13 @@ class VmecTests(unittest.TestCase):
         approximately the value of iota at center of weight function.
         """
         filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename,ntheta=50,nphi=50)
+        vmec = Vmec(filename, ntheta=50, nphi=50)
 
         vmec.run()
         iota_center = vmec.wout.iotas[50]
         s_center = vmec.s_half_grid[50]
-        target_function = lambda s : np.exp(-(s-s_center)**2/0.01**2)
-        self.assertAlmostEqual(vmec.iota_weighted(target_function),iota_center,places=2)
+        target_function = lambda s: np.exp(-(s-s_center)**2/0.01**2)
+        self.assertAlmostEqual(vmec.iota_weighted(target_function), iota_center, places=2)
 
     def test_well_weighted(self):
         """
@@ -231,7 +231,7 @@ class VmecTests(unittest.TestCase):
         and edge matches well metric computed with V'(0) and V'(1).
         """
         filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename,ntheta=50,nphi=50)
+        vmec = Vmec(filename, ntheta=50, nphi=50)
 
         vmec.run()
         vp_l = 1.5*vmec.wout.vp[1]-0.5*vmec.wout.vp[2]
@@ -239,8 +239,8 @@ class VmecTests(unittest.TestCase):
         well1 = (vp_l-vp_r)/(vp_l+vp_r)
         weight1 = lambda s: np.exp(-s**2/0.01**2)
         weight2 = lambda s: np.exp(-(1-s)**2/0.01**2)
-        well2 = vmec.well_weighted(weight1,weight2)
-        self.assertAlmostEqual(well1,well2,places=2)
+        well2 = vmec.well_weighted(weight1, weight2)
+        self.assertAlmostEqual(well1, well2, places=2)
 
     def test_B_on_arclength_grid(self):
         """
@@ -248,7 +248,7 @@ class VmecTests(unittest.TestCase):
         Check that B is unchanged when evaluated on arclength grid from first run.
         """
         filename = os.path.join(TEST_DIR, 'input.rotating_ellipse')
-        vmec = Vmec(filename,ntheta=50,nphi=50)
+        vmec = Vmec(filename, ntheta=50, nphi=50)
 
         Bx, By, Bz, arclength0 = vmec.B_on_arclength_grid()
         B2 = Bx*Bx + By*By + Bz*Bz
@@ -259,18 +259,18 @@ class VmecTests(unittest.TestCase):
[…]
