+++ src/simsopt/mhd/vmec.py
@@ -230,8 +230,8 @@ class Vmec(Optimizable):
         # object, but the mpol/ntor values of either the vmec object
         # or the boundary surface object can be changed independently
         # by the user.
-        quadpoints_theta = np.linspace(0,1.,ntheta,endpoint=False)
-        quadpoints_phi = np.linspace(0,1.,nphi,endpoint=False)
+        quadpoints_theta = np.linspace(0, 1., ntheta, endpoint=False)
+        quadpoints_phi = np.linspace(0, 1., nphi, endpoint=False)
         self.boundary = SurfaceRZFourier(nfp=vi.nfp,
                                          stellsym=not vi.lasym,
                                          mpol=vi.mpol,
@@ -596,7 +596,7 @@ class Vmec(Optimizable):
         well = (dVds_s0 - dVds_s1) / dVds_s0
         return well
 
-    def iota_weighted(self,weight_function):
+    def iota_weighted(self, weight_function):
         """
         Computed a weighted average of the rotational transform defined by
         the prescribed weight_function, f.
@@ -612,7 +612,7 @@ class Vmec(Optimizable):
         return np.sum(weight_function(self.s_half_grid) * self.wout.iotas[1:]) \
             / np.sum(weight_function(self.s_half_grid))
 
-    def iota_target_metric(self,iota_target):
+    def iota_target_metric(self, iota_target):
         """
         Computes a metric quantifying the deviation of iota from a prescribed
         target value.
@@ -626,9 +626,9 @@ class Vmec(Optimizable):
         """
         self.run()
         return 0.5 * np.sum((self.wout.iotas[1::]
-            - iota_target(self.s_half_grid))**2) * self.ds
+                             - iota_target(self.s_half_grid))**2) * self.ds
 
-    def d_iota_target_metric(self,iota_target,delta=1.):
+    def d_iota_target_metric(self, iota_target, delta=1.):
         """
         Computes derivatives of iota_target_metric wrt surface parameters using
         an adjoint method. The parameter delta sets the amplitude of the toroidal
@@ -638,11 +638,11 @@ class Vmec(Optimizable):
             raise RuntimeError('''d_iota_target_metric cannot be computed without
                 running vmec with ncurr = 1''')
 
-        shape_gradient = self.iota_target_metric_shape_gradient(iota_target,delta)
+        shape_gradient = self.iota_target_metric_shape_gradient(iota_target, delta)
 
-        return parameter_derivatives(self.boundary,shape_gradient)
+        return parameter_derivatives(self.boundary, shape_gradient)
 
-    def iota_target_metric_shape_gradient(self,iota_target,delta):
+    def iota_target_metric_shape_gradient(self, iota_target, delta):
         """
         Computes shape gradient of iota_target_metric, defined as S where,
 
@@ -656,14 +656,14 @@ class Vmec(Optimizable):
         Bx0, By0, Bz0, theta_arclength0 = self.B_on_arclength_grid()
 
         mu0 = 4*np.pi*1e-7
-        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0,1::] / mu0
+        It_half = self.wout.signgs * 2*np.pi * self.wout.bsubumnc[0, 1::] / mu0
[…]
