### 2021-07-26 ba1e590
more tweaks to tests.
