+++ examples/1_Simple/create_individual_maps.py
@@ -1,0 +1,173 @@
+import numpy as np
+from simsopt.field.biotsavart import BiotSavart
+from simsopt.geo.coilcollection import CoilCollection
+from simsopt.util.zoo import get_ncsx_data
+from simsopt.util.constants import PROTON_MASS, ONE_EV
+
+MASS = PROTON_MASS
+EKIN = 5000 * ONE_EV
+v2 = 2 * EKIN/MASS
+
+nphi = 3
+
+# phi_hits = []
+# for i in range(nparticles):
+#     phi_hits.append(np.load(f'/tmp/phihits/24_gc_phi_hits_{i}.npy'))
+
+# phi_hits = np.load('24_bs.npy',allow_pickle=True)
+# phi_hits = np.load('phi_24_n_1000_bsh.npy',allow_pickle=True)
+phi_hits = np.load(f'phi_{nphi}_n_4000_many_u_bsh.npy',allow_pickle=True)
+
+nfp = 3
+phis = [(i/nphi)*(2*np.pi) for i in range(nphi)]
+functions_next = [ [] for k in range(nphi) ]
+functions_same = [ [] for k in range(nphi) ]
+functions_prev = [ [] for k in range(nphi) ]
+functions_exit = [ [] for k in range(nphi) ]
+for i in range(phi_hits.shape[0]):
+    for j in range(phi_hits[i].shape[0]-1):
+        this_idx = int(phi_hits[i][j, 1])
+        next_idx = int(phi_hits[i][j+1, 1])
+        dt = phi_hits[i][j+1, 0] - phi_hits[i][j, 0]
+        rin = np.linalg.norm(phi_hits[i][j, 2:4])
+        rout = np.linalg.norm(phi_hits[i][j+1, 2:4])
+        zin = phi_hits[i][j, 4]
+        zout = phi_hits[i][j+1, 4]
+        vin = phi_hits[i][j, 5]/np.sqrt(v2)
+        vout = phi_hits[i][j+1, 5]/np.sqrt(v2)
+        data = np.asarray([rin, zin, vin, rout, zout, vout, dt])
+        if next_idx == this_idx:
+            functions_same[this_idx].append(data)
+        elif next_idx == (this_idx+1) % nphi:
+            functions_next[this_idx].append(data)
+        elif next_idx == (this_idx-1) % nphi:
+            functions_prev[this_idx].append(data)
+        elif next_idx < 0:
+            functions_exit[this_idx].append(data)
+
+for phiidx in range(nphi):
+    functions_same[phiidx] = np.asarray(functions_same[phiidx])
+    functions_prev[phiidx] = np.asarray(functions_prev[phiidx])
+    functions_next[phiidx] = np.asarray(functions_next[phiidx])
+    functions_exit[phiidx] = np.asarray(functions_exit[phiidx])
+
+
+coils, currents, ma = get_ncsx_data(Nt_coils=25, Nt_ma=20)
+
+# stellarator = CoilCollection(coils, currents, 3, True)
+# coils = stellarator.coils
+# currents = stellarator.currents
+# bs = BiotSavart(coils, currents)
+# data = functions_next[0]
+# AbsB = bs.set_points(np.ascontiguousarray(data[:, 0:3])).AbsB()
+# mus = (v2-data[:, -1])[:, None]/(2*AbsB)
+# print("mus", np.min(mus), np.max(mus), mus)
+# data = functions_same[0]
+# AbsB = bs.set_points(np.ascontiguousarray(data[:, 0:3])).AbsB()
+# mus = (v2-data[:, -1])[:, None]/(2*AbsB)
+# print("mus", np.min(mus), np.max(mus))
+
+
[…]
