+++ examples/1_Simple/tracing_particle.py
@@ -8,6 +8,9 @@ from simsopt.field.tracing import trace_particles_starting_on_curve, SurfaceClas
 from simsopt.geo.curve import curves_to_vtk
 from simsopt.util.zoo import get_ncsx_data
 from simsopt.util.constants import PROTON_MASS, ELEMENTARY_CHARGE, ONE_EV
+from simsopt.geo.surfacexyztensorfourier import SurfaceXYZTensorFourier
+from simsopt.geo.boozersurface import BoozerSurface
+from simsopt.geo.surfaceobjectives import boozer_surface_residual, ToroidalFlux, Area
 import simsoptpp as sopp
 try:
     from mpi4py import MPI
@@ -30,8 +33,8 @@ logger.setLevel(1)
 
 # check whether we're in CI, in that case we make the run a bit cheaper
 ci = "CI" in os.environ and os.environ['CI'].lower() in ['1', 'true']
-nparticles = 3 if ci else 100
-degree = 2 if ci else 3
+nparticles = 3 if ci else 20000
+degree = 2 if ci else 4
 
 
 """
@@ -40,29 +43,60 @@ trajectories of particles
 """
 
 
-coils, currents, ma = get_ncsx_data(Nt_coils=25, Nt_ma=10)
+coils, currents, ma = get_ncsx_data(Nt_coils=25, Nt_ma=20)
 
 stellarator = CoilCollection(coils, currents, 3, True)
 coils = stellarator.coils
 currents = stellarator.currents
 bs = BiotSavart(coils, currents)
-print("Mean(|B|) on axis =", np.mean(np.linalg.norm(bs.set_points(ma.gamma()).B(), axis=1)))
-print("Mean(Axis radius) =", np.mean(np.linalg.norm(ma.gamma(), axis=1)))
-curves_to_vtk(coils + [ma], '/tmp/coils')
+print("Mean(|B|) on axis =", np.mean(np.linalg.norm(bs.set_points(ma.gamma()).B(), axis=1)), flush=True)
+print("Mean(Axis radius) =", np.mean(np.linalg.norm(ma.gamma(), axis=1)), flush=True)
 
-mpol = 5
-ntor = 5
+mpol = 7
+ntor = 7
 stellsym = True
 nfp = 3
-phis = np.linspace(0, 1, nfp*2*ntor+1, endpoint=False)
+
+phis = np.linspace(0, 1/nfp, 2*ntor+1, endpoint=False)
 thetas = np.linspace(0, 1, 2*mpol+1, endpoint=False)
-s = SurfaceRZFourier(
+s = SurfaceXYZTensorFourier(
     mpol=mpol, ntor=ntor, stellsym=stellsym, nfp=nfp, quadpoints_phi=phis, quadpoints_theta=thetas)
+s.fit_to_curve(ma, 0.10, flip_theta=True)
+iota = -0.4
+
+ar = Area(s)
+ar_target = ar.J()
+
+boozer_surface = BoozerSurface(bs, s, ar, ar_target)
+G0 = 2. * np.pi * np.sum(np.abs(bs.coil_currents)) * (4 * np.pi * 10**(-7) / (2 * np.pi))
+
+res = boozer_surface.minimize_boozer_penalty_constraints_LBFGS(tol=1e-10, maxiter=300, constraint_weight=100., iota=iota, G=G0)
+print(f"After LBFGS:   iota={res['iota']:.3f}, area={s.area():.3f}, ||residual||={np.linalg.norm(boozer_surface_residual(s, res['iota'], res['G'], bs, derivatives=0)):.3e}", flush=True)
[…]
