### 2021-07-27 8acbd0a
Conversion of mhd modules to graph framework
