+++ src/simsopt/mhd/spec.py
@@ -21,39 +21,32 @@ except ImportError as e:
     MPI = None 
     logger.warning(str(e))
 
-# spec_found = True
 try:
     import spec
 except ImportError as e:
     spec = None
     logger.warning(str(e))
-    # spec_found = False
 
-# py_spec_found = True
 try:
     import py_spec
 except ImportError as e:
     py_spec = None
     logger.warning(str(e))
-    # py_spec_found = False
 
-# pyoculus_found = True
 try:
     import pyoculus
 except ImportError as e:
     pyoculus = None
     logger.warning(str(e))
-    # pyoculus_found = False
 
-from .._core.optimizable import Optimizable
+from .._core.graph_optimizable import Optimizable
 from .._core.util import ObjectiveFailure
-from ..geo.surfacerzfourier import SurfaceRZFourier
+from ..geo.graph_surfacerzfourier import SurfaceRZFourier
 from ..util.dev import SimsoptRequires
 if MPI is not None:
     from ..util.mpi import MpiPartition
 else:
     MpiPartition = None
-#from ..util.mpi import MpiPartition
 
 
 @SimsoptRequires(MPI is not None, "mpi4py needs to be installed for running SPEC")
@@ -94,12 +87,9 @@ class Spec(Optimizable):
                  verbose: bool = True,
                  keep_all_files: bool = False):
 
-        #if not spec_found:
         if spec is None:
             raise RuntimeError(
                 "Using Spec requires spec python wrapper to be installed.")
-
-        #if not py_spec_found:
         if py_spec is None:
             raise RuntimeError(
                 "Using Spec requires py_spec to be installed.")
@@ -135,12 +125,11 @@ class Spec(Optimizable):
             # Read default input file, which should be in the same
             # directory as this file:
             filename = os.path.join(os.path.dirname(__file__), 'defaults.sp')
-            logger.info("Initializing a SPEC object from defaults in " \
-                        + filename)
+            logger.info(f"Initializing a SPEC object from defaults in {filename}")
         else:
             if not filename.endswith('.sp'):
-                filename = filename + '.sp'
-            logger.info("Initializing a SPEC object from file: " + filename)
+                filename = f"{filename}.sp"
+            logger.info(f"Initializing a SPEC object from file: {filename}")
 
         self.init(filename)
         self.extension = filename[:-3]
@@ -150,8 +139,8 @@ class Spec(Optimizable):
         # Create a surface object for the boundary:
         si = spec.inputlist  # Shorthand
         stellsym = bool(si.istellsym)
-        print("In __init__, si.istellsym=", si.istellsym, " stellsym=", stellsym)
-        self.boundary = SurfaceRZFourier(nfp=si.nfp,
+        print(f"In __init__, si.istellsym={si.istellsym} stellsym={stellsym}")
+        self._boundary = SurfaceRZFourier(nfp=si.nfp,
                                          stellsym=stellsym,
                                          mpol=si.mpol,
                                          ntor=si.ntor)
@@ -166,14 +155,32 @@ class Spec(Optimizable):
                     self.boundary.rs[m, n + si.ntor] = si.rbs[n + si.mntor, m + si.mmpol]
                     self.boundary.zc[m, n + si.ntor] = si.zbc[n + si.mntor, m + si.mmpol]
 
-        self.depends_on = ["boundary"]
+        # self.depends_on = ["boundary"]
         self.need_to_run_code = True
         self.counter = -1
 
         # By default, all dofs owned by SPEC directly, as opposed to
         # dofs owned by the boundary surface object, are fixed.
-        self.fixed = np.full(len(self.get_dofs()), True)
-        self.names = ['phiedge', 'curtor']
[…]
