+++ src/simsopt/mhd/vmec.py
@@ -30,9 +30,9 @@ except ImportError as e:
     vmec = None
     logger.warning(str(e))
 
-from .._core.optimizable import Optimizable
+from .._core.graph_optimizable import Optimizable
 from .._core.util import Struct, ObjectiveFailure
-from ..geo.surfacerzfourier import SurfaceRZFourier
+from ..geo.graph_surfacerzfourier import SurfaceRZFourier
 
 if MPI is not None:
     from ..util.mpi import MpiPartition
@@ -179,10 +179,9 @@ class Vmec(Optimizable):
             # Read default input file, which should be in the same
             # directory as this file:
             filename = os.path.join(os.path.dirname(__file__), 'input.default')
-            logger.info("Initializing a VMEC object from defaults in "
-                        + filename)
+            logger.info(f"Initializing a VMEC object from defaults in {filename}")
         else:
-            logger.info("Initializing a VMEC object from file: " + filename)
+            logger.info(f"Initializing a VMEC object from file: {filename}")
         self.input_file = filename
 
         # Get MPI communicator:
@@ -228,7 +227,7 @@ class Vmec(Optimizable):
         # object, but the mpol/ntor values of either the vmec object
         # or the boundary surface object can be changed independently
         # by the user.
-        self.boundary = SurfaceRZFourier(nfp=vi.nfp,
+        self._boundary = SurfaceRZFourier(nfp=vi.nfp,
                                          stellsym=not vi.lasym,
                                          mpol=vi.mpol,
                                          ntor=vi.ntor)
@@ -240,11 +239,25 @@ class Vmec(Optimizable):
                 self.boundary.rc[m, n + vi.ntor] = vi.rbc[101 + n, m]
                 self.boundary.zs[m, n + vi.ntor] = vi.zbs[101 + n, m]
         # Handle a few variables that are not Parameters:
-        self.depends_on = ["boundary"]
         self.need_to_run_code = True
 
-        self.fixed = np.full(len(self.get_dofs()), True)
-        self.names = ['delt', 'tcon0', 'phiedge', 'curtor', 'gamma']
+        x0 = self.get_dofs(),
+        fixed = np.full(len(x0), True)
+        names = ['delt', 'tcon0', 'phiedge', 'curtor', 'gamma']
+        super().__init__(x0=x0, fixed=fixed, names=names,
+                         opts_in=[self.boundary],
+                         external_dof_setter=Vmec.set_dofs)
+
+    @property
+    def boundary(self):
+        return self._boundary
+
+    @boundary.setter
+    def boundary(self, boundary):
+        if not boundary is self._boundary:
+            self.remove_parent(self._boundary)
+            self._boundary = boundary
+            self.append_parent(boundary)
 
     def get_dofs(self):
         return np.array([self.indata.delt, self.indata.tcon0,
@@ -258,6 +271,9 @@ class Vmec(Optimizable):
         self.indata.phiedge = x[2]
         self.indata.curtor = x[3]
         self.indata.gamma = x[4]
+
+    def recompute_bell(self, parent=None):
+        self.need_to_run_code = True
 
     def run(self):
         """
@@ -392,70 +408,22 @@ class Vmec(Optimizable):
         data in a ``wout`` attribute of this Vmec object.
         """
         ierr = 0
-        logger.info("Attempting to read file " + self.output_file)
-        # vmec.read_wout_mod.read_wout_file(self.output_file, ierr)
-        #if ierr == 0:
[…]
