### 2021-07-27 f850e5a
Minor cleanup in CMakeLists.txt
