### 2021-08-03 #3525 @StanczakDominik
Hey, I just wanted to report back and say that I'm sorry for the delay. I've had an intense few weeks. I'm now heading for my first vacation in a couple years (gee thanks, COVID!). Once I'm back next Tuesday, I'm making going through the rest of the package my first priority. The parts I've seen so far are looking great, by the way!
