### 2021-08-06 #3525 @ZedThree
Mostly done, I need to dig a bit more into the functionality and API docs. But generally, everything's been of a high standard so far.
