### 2021-08-23 hiddenSymmetries/simsopt#134 @StanczakDominik
Mayavi vs Pyvista, from the docker standpoint
I'm doing my JOSS review out of the provided Docker image and I'm just lamenting the fact that Mayavi keeps crashing my Jupyter server.

From what I can tell, it seems that Mayavi is used only in these two locations:

https://github.com/hiddenSymmetries/simsopt/blob/77c32e08b33c660bfd300ba2c2f5b2fbc9bc3ab3/src/simsopt/geo/surface.py#L21-L50

https://github.com/hiddenSymmetries/simsopt/blob/af234540745b7e37c5d6cc737e74cf65d7d09299/src/simsopt/geo/curve.py#L140-L149

I suspect that if this was (simply, it looks like) rewritten using PyVista, we could then take advantage of https://docs.pyvista.org/extras/docker.html and get this to work from within the notebook! Is there anything I'm missing here? In particular, is there any tighter coupling/investment between Simsopt and Mayavi that I'm missing?

---

Of course, the whole issue of running this out of a Docker container is a side effect of SPEC being private, and the Docker image being the only way I can get my hands on it for review purposes - but that's a rant that I don't want to get into 😅

---

Let me add that this should in no way be considered a blocker for the JOSS review. Just putting the idea out there.
