### 2021-08-24 #3525 @StanczakDominik
Mostly done with my review! Everything looks, well, stellar. 🌞

---

One concern I have is the inclusion/dependency on SPEC. Based on some of my digging around (and https://github.com/hiddenSymmetries/simsopt/issues/35#issuecomment-854786329 ), my understanding is that:

* SIMSOPT wraps SPEC.
* SPEC itself is closed source.
* The provided Docker image contains compiled SPEC binaries, which SIMSOPT can use.
* The creation of those binaries is carried out during the Docker image build, using keys to the private repo as Github secrets.
* This means that I have been unable to run the SPEC-dependent examples locally, only through Docker.
* This, in turn, means I've been unable to actually visualize the 3D results wherever the plotting is written using Mayavi, due to X11/docker weirdness (see https://github.com/hiddenSymmetries/simsopt/issues/134 for a future-work proposal to sidestep the issue).

I have a few ideas about this, in order of increasing feasibility:

1. Open source SPEC! This is of course both the ideal world scenario and out of your hands, and was apparently in-progress-but-will-take-a-while in April: https://github.com/hiddenSymmetries/simsopt/issues/35#issuecomment-815885445 ;
2. Contribute the SPEC binary building process of the SIMSOPT docker build upstream, so SPEC can remain closed source but provide the binaries (probably a waste of time depending on how close SPEC is to becoming open source, though);
3. Note more visibly (in the paper? in the SIMSOPT readme? maybe both?) that, as of `{date}`, SPEC is on the way to open source, but not there yet, and for now people can use the Docker image that provides its binaries. This kind of thing should be stated openly, just so people don't have to dig into this issue themselves.
	* Ideally, add some notes on the 3d viz out of Docker issue, which I'm not sure has been considered (I imagine SIMSOPT devs have SPEC access!).

---

Besides that, though, really a cool piece of software and I definitely hope I'll get to use it some day :)
