### 2021-08-25 hiddenSymmetries/simsopt#134 @mbkumar
@StanczakDominik,
Sorry to hear that Stan. How about increasing the memory of the container to 8 GB and try again in case you are running the container on mac?

@florianwechsung,
could you please look into Stan's suggestion when you have time?
