### 2021-08-25 hiddenSymmetries/simsopt#134 @florianwechsung
Hi @StanczakDominik

There isn't any tight coupling between SIMSOPT and mayavi, and I'm all onboard with getting rid of mayavi, as it has given me trouble a few times now.

Most objects at this point can be exported to vtk for plotting purposes (using https://github.com/pyscience-projects/pyevtk), but having some form of visualisation inside jupyter notebooks would be nice. I will have a look at pyvista when I'm back from holiday next week.
