### 2021-08-31 #3525 @ZedThree
Hi @landreman, sorry for the silence, I've been off on leave for a good chunk of August.

Overall, SIMSOPT is a really great package, and I would definitely recommend it to colleagues. It was easy to get started with it, lots of easy-to-follow examples, and a decent API documentation.

I had a similar comment to @StanczakDominik about SPEC, so I won't repeat that.

My other comment was that there didn't seem to be much in the way of documentation on the C++ part. I can see how most of it relates to the equivalent Python parts, but it would be good to have a few words in the documentation about it. I guess this is more important for developers than users.
