### 2021-09-01 #3525 @landreman
@ZedThree Thanks for taking the time to look at the code and for your comments. I agree that the C++ components could use more comments and documentation - this is on my to-do list.
