### 2021-09-08 hiddenSymmetries/simsopt#134 @landreman
Resolved by #137
