### 2021-09-09 #3525 @mbkumar
@dfm
version number incremented to v0.4.7
DOI is 10.5281/zenodo.5498111
